+++ examples/wxapp/lib/sentry-miniapp.js
@@ -1,19 +1,19 @@
-(function(g,E){typeof exports=="object"&&typeof module!="undefined"?E(exports):typeof define=="function"&&define.amd?define(["exports"],E):(g=typeof globalThis!="undefined"?globalThis:g||self,E(g.SentryMiniapp={}))})(this,(function(g){"use strict";var gl=Object.defineProperty,yl=Object.defineProperties;var Sl=Object.getOwnPropertyDescriptors;var Xt=Object.getOwnPropertySymbols;var gi=Object.prototype.hasOwnProperty,yi=Object.prototype.propertyIsEnumerable;var or=(g,E)=>(E=Symbol[g])?E:Symbol.for("Symbol."+g),bl=g=>{throw TypeError(g)};var mi=(g,E,R)=>E in g?gl(g,E,{enumerable:!0,configurable:!0,writable:!0,value:R}):g[E]=R,l=(g,E)=>{for(var R in E||(E={}))gi.call(E,R)&&mi(g,R,E[R]);if(Xt)for(var R of Xt(E))yi.call(E,R)&&mi(g,R,E[R]);return g},I=(g,E)=>yl(g,Sl(E));var Si=(g,E)=>{var R={};for(var D in g)gi.call(g,D)&&E.indexOf(D)<0&&(R[D]=g[D]);if(g!=null&&Xt)for(var D of Xt(g))E.indexOf(D)<0&&yi.call(g,D)&&(R[D]=g[D]);return R};var z=(g,E,R)=>new Promise((D,U)=>{var y=O=>{try{q(R.next(O))}catch(X){U(X)}},C=O=>{try{q(R.throw(O))}catch(X){U(X)}},q=O=>O.done?D(O.value):Promise.resolve(O.value).then(y,C);q((R=R.apply(g,E)).next())}),El=function(g,E){this[0]=g,this[1]=E};var bi=g=>{var E=g[or("asyncIterator")],R=!1,D,U={};return E==null?(E=g[or("iterator")](),D=y=>U[y]=C=>E[y](C)):(E=E.call(g),D=y=>U[y]=C=>{if(R){if(R=!1,y==="throw")throw C;return C}return R=!0,{done:!1,value:new El(new Promise(q=>{var O=E[y](C);O instanceof Object||bl("Object expected"),q(O)}),1)}}),U[or("iterator")]=()=>U,D("next"),"throw"in E?D("throw"):U.throw=y=>{throw y},"return"in E&&D("return"),U};class E{constructor(e){if(this._entries=[],typeof e=="string")this._parseString(e);else if(Array.isArray(e))for(const n of e)Array.isArray(n)&&n.length>=2&&this._entries.push([n[0]||"",n[1]||""]);else if(e&&typeof e=="object")if(e instanceof E)this._entries=e._entries.map(([n,r])=>[n,r]);else for(const[n,r]of Object.entries(e))this._entries.push([n,r])}get size(){return this._entries.length}_parseString(e){if(e.startsWith("?")&&(e=e.slice(1)),!e)return;const n=e.split("&");for(const r of n){const s=r.indexOf("=");if(s===-1)r&&this._entries.push([decodeURIComponent(r),""]);else{const i=r.slice(0,s),o=r.slice(s+1);i&&this._entries.push([decodeURIComponent(i),decodeURIComponent(o)])}}}append(e,n){this._entries.push([e,String(n)])}delete(e){this._entries=this._entries.filter(([n])=>n!==e)}get(e){const n=this._entries.find(([r])=>r===e);return n?n[1]:null}getAll(e){return this._entries.filter(([n])=>n===e).map(([,n])=>n)}has(e){return this._entries.some(([n])=>n===e)}set(e,n){const r=String(n);let s=!1;if(this._entries=this._entries.filter(([i])=>i===e?s?!1:(s=!0,!0):!0),s){const i=this._entries.findIndex(([o])=>o===e);i!==-1&&(this._entries[i]=[e,r])}else this._entries.push([e,r])}sort(){this._entries.sort(([e],[n])=>e<n?-1:e>n?1:0)}toString(){return this._entries.map(([e,n])=>`${encodeURIComponent(e)}=${encodeURIComponent(n)}`).join("&")}forEach(e){for(const[n,r]of this._entries)e(r,n,this)}*keys(){for(const[e]of this._entries)yield e}*values(){for(const[,e]of this._entries)yield e}*entries(){for(const e of this._entries)yield e}*[Symbol.iterator](){yield*bi(this.entries())}}function R(){if(typeof globalThis!="undefined")return globalThis;try{return Function("return this")()}catch(t){}if(typeof window!="undefined")return window;if(typeof global!="undefined")return global;if(typeof self!="undefined")return self}function D(){try{const t=R();if(!t){console.warn("[sentry-miniapp] Unable to detect global object, polyfills may not work correctly");return}typeof t.URLSearchParams=="undefined"&&(t.URLSearchParams=E);const e=Function("return this")();e&&typeof e.URLSearchParams=="undefined"&&(e.URLSearchParams=E),typeof globalThis!="undefined"&&typeof globalThis.URLSearchParams=="undefined"&&(globalThis.URLSearchParams=E)}catch(t){console.warn("[sentry-miniapp] Failed to install polyfills:",t)}}function U(){D()}U();const y=typeof __SENTRY_DEBUG__=="undefined"||__SENTRY_DEBUG__,C=globalThis,q="10.68.0";function O(){return X(C),C}function X(t){const e=t.__SENTRY__=t.__SENTRY__||{};return e.version=e.version||q,e[q]=e[q]||{}}function le(t,e,n=C){const r=n.__SENTRY__=n.__SENTRY__||{},s=r[q]=r[q]||{};return s[t]||(s[t]=e())}const Ei="Sentry Logger ",ar={};function fe(t){if(!("console"in C))return t();const e=C.console,n={},r=Object.keys(ar);r.forEach(s=>{const i=ar[s];n[s]=e[s],e[s]=i});try{return t()}finally{r.forEach(s=>{e[s]=n[s]})}}function Ti(){Zt().enabled=!0}function vi(){Zt().enabled=!1}function cr(){return Zt().enabled}function Ii(...t){Qt("log",...t)}function wi(...t){Qt("warn",...t)}function ki(...t){Qt("error",...t)}function Qt(t,...e){y&&cr()&&fe(()=>{C.console[t](`${Ei}[${t}]:`,...e)})}function Zt(){return y?le("loggerSettings",()=>({enabled:!1})):{enabled:!1}}const b={enable:Ti,disable:vi,isEnabled:cr,log:Ii,warn:wi,error:ki},ur=50,Ne="?",lr=/\(error: (.*)\)/,fr=/captureMessage|captureException/;function Ai(...t){const e=t.sort((n,r)=>n[0]-r[0]).map(n=>n[1]);return(n,r=0,s=0)=>{const i=[],o=n.split(`
-`);for(let c=r;c<o.length;c++){let a=o[c];a.length>1024&&(a=a.slice(0,1024));const u=lr.test(a)?a.replace(lr,"$1"):a;if(!u.includes("Error: ")){for(const p of e){const f=p(u);if(f){i.push(f);break}}if(i.length>=ur+s)break}}return Ri(i.slice(s))}}function Ri(t){if(!t.length)return[];const e=Array.from(t);return/sentryWrapped/.test(Xe(e).function||"")&&e.pop(),e.reverse(),fr.test(Xe(e).function||"")&&(e.pop(),fr.test(Xe(e).function||"")&&e.pop()),e.slice(0,ur).map(n=>I(l({},n),{filename:n.filename||Xe(e).filename,function:n.function||Ne}))}function Xe(t){return t[t.length-1]||{}}const en="<anonymous>";function Ci(t){try{return!t||typeof t!="function"?en:t.name||en}catch(e){return en}}const dr=Object.prototype.toString;function pr(t){switch(dr.call(t)){case"[object Error]":case"[object Exception]":case"[object DOMException]":case"[object WebAssembly.Exception]":return!0;default:return _r(t,Error)}}function tn(t,e){return dr.call(t)===`[object ${e}]`}function nn(t){return tn(t,"String")}function rn(t){return typeof t=="object"&&t!==null&&"__sentry_template_string__"in t&&"__sentry_template_values__"in t}function hr(t){return t===null||rn(t)||typeof t!="object"&&typeof t!="function"}function sn(t){return tn(t,"Object")}function on(t){return typeof t=="object"&&t!==null}function Pi(t){return typeof Event!="undefined"&&_r(t,Event)}function Ni(t){return tn(t,"RegExp")}function de(t){return!!(t!=null&&t.then&&typeof t.then=="function")}function _r(t,e){try{return t instanceof e}catch(n){return!1}}function Q(t,e,n){try{Object.defineProperty(t,e,{value:n,writable:!0,configurable:!0})}catch(r){y&&b.log(`Failed to add non-enumerable property "${String(e)}" to object`,t)}}function Di(t){if(pr(t))return l({message:t.message,name:t.name,stack:t.stack},mr(t));if(Pi(t)){const{type:e,target:n,currentTarget:r,detail:s}=t;return l(l({type:e,target:n,currentTarget:r},s?{detail:s}:{}),mr(t))}return t}function mr(t){return on(t)?Object.fromEntries(Object.entries(t)):{}}let pe;function Qe(t){if(pe!==void 0)return pe?pe(t):t();const e=Symbol.for("__SENTRY_SAFE_RANDOM_ID_WRAPPER__"),n=C;return e in n&&typeof n[e]=="function"?(pe=n[e],pe(t)):(pe=null,t())}function Ze(){return Qe(()=>Math.random())}function ee(){return Qe(()=>Date.now())}const Oi=Symbol.for("sentry.skipNormalization"),Mi=Symbol.for("sentry.overrideNormalizationDepth");function xi(t){return!!t[Oi]}function Fi(t){const e=t[Mi];return typeof e=="number"?e:void 0}let Li;function te(t,e=100,n=1/0){try{return an("",t,e,n)}catch(r){return{ERROR:`**non-serializable** (${r})`}}}function an(t,e,n=1/0,r=1/0,s=Bi()){const[i,o]=s;if(e==null||["boolean","string"].includes(typeof e)||typeof e=="number"&&Number.isFinite(e))return e;const c=$i(t,e);if(!c.startsWith("[object "))return c;if(xi(e))return e;const a=Fi(e),u=a!==void 0?a:n;if(u===0)return c.replace("object ","");if(i(e))return"[Circular ~]";const p=e;if(p&&typeof p.toJSON=="function")try{const h=p.toJSON();return an("",h,u-1,r,s)}catch(h){}const f=Array.isArray(e)?[]:{};let _=0;const d=Di(e);for(const h in d){if(!Object.prototype.hasOwnProperty.call(d,h))continue;if(_>=r){f[h]="[MaxProperties ~]";break}const m=d[h];f[h]=an(h,m,u-1,r,s),_++}return o(e),f}function $i(t,e){try{return typeof global!="undefined"&&e===global?"[Global]":typeof e=="number"&&!Number.isFinite(e)?`[${e}]`:typeof e=="function"?`[Function: ${Ci(e)}]`:typeof e=="symbol"?`[${String(e)}]`:typeof e=="bigint"?`[BigInt: ${String(e)}]`:`[object ${ji(e)}]`}catch(n){return`**non-serializable** (${n})`}}function ji(t){const e=Object.getPrototypeOf(t);return e!=null&&e.constructor?e.constructor.name:"null prototype"}function Bi(){const t=new WeakSet;function e(r){return t.has(r)?!0:(t.add(r),!1)}function n(r){t.delete(r)}return[e,n]}function cn(t,e=0){return typeof t!="string"||e===0||t.length<=e?t:`${t.slice(0,e)}...`}function De(t,e,n=!1){return nn(t)?Ni(e)?e.test(t):nn(e)?n?t===e:t.includes(e):typeof e=="function"?e(t):!1:!1}function et(t,e=[],n=!1){for(const r of e)if(De(t,r,n))return!0;return!1}function Ui(){const t=C;return t.crypto||t.msCrypto}let un;function Hi(){return Ze()*16}function W(t=Ui()){try{if(t!=null&&t.randomUUID)return Qe(()=>t.randomUUID()).replace(/-/g,"")}catch(e){}return un||(un="10000000100040008000"+1e11),un.replace(/[018]/g,e=>(e^(Hi()&15)>>e/4).toString(16))}function gr(t){var e,n;return(n=(e=t.exception)==null?void 0:e.values)==null?void 0:n[0]}function he(t){const{message:e,event_id:n}=t;if(e)return e;const r=gr(t);return r?r.type&&r.value?`${r.type}: ${r.value}`:r.type||r.value||n||"<unknown>":n||"<unknown>"}function zi(t,e){const n=gr(t);if(!n)return;const r={type:"generic",handled:!0},s=n.mechanism;if(n.mechanism=l(l(l({},r),s),e),e&&"data"in e){const i=l(l({},s==null?void 0:s.data),e.data);n.mechanism.data=i}}function yr(t){if(qi(t))return!0;try{Q(t,"__sentry_captured__",!0)}catch(e){}return!1}function qi(t){try{return t.__sentry_captured__}catch(e){}}const Sr=1e3;function Oe(){return ee()/Sr}function Wi(){const{performance:t}=C;if(!(t!=null&&t.now)||!t.timeOrigin)return Oe;const e=t.timeOrigin;return()=>(e+Qe(()=>t.now()))/Sr}let tt;function _e(){return(tt!=null?tt:tt=Wi())()}function br(t){const e=_e(),n={sid:W(),init:!0,timestamp:e,started:e,duration:0,status:"ok",errors:0,ignoreDuration:!1,toJSON:()=>Gi(n)};return t&&se(n,t),n}function se(t,e={}){if(e.user&&(!t.ipAddress&&e.user.ip_address&&(t.ipAddress=e.user.ip_address),!t.did&&!e.did&&(t.did=e.user.id||e.user.email||e.user.username)),t.timestamp=e.timestamp||_e(),e.abnormal_mechanism&&(t.abnormal_mechanism=e.abnormal_mechanism),e.ignoreDuration&&(t.ignoreDuration=e.ignoreDuration),e.sid&&(t.sid=e.sid.length===32?e.sid:W()),e.init!==void 0&&(t.init=e.init),!t.did&&e.did&&(t.did=`${e.did}`),typeof e.started=="number"&&(t.started=e.started),t.ignoreDuration)t.duration=void 0;else if(typeof e.duration=="number")t.duration=e.duration;else{const n=t.timestamp-t.started;t.duration=n>=0?n:0}e.release&&(t.release=e.release),e.environment&&(t.environment=e.environment),!t.ipAddress&&e.ipAddress&&(t.ipAddress=e.ipAddress),!t.userAgent&&e.userAgent&&(t.userAgent=e.userAgent),typeof e.errors=="number"&&(t.errors=e.errors),e.status&&(t.status=e.status)}function Er(t,e){let n={};e?n={status:e}:t.status==="ok"&&(n={status:"exited"}),se(t,n)}function Gi(t){return{sid:`${t.sid}`,init:t.init,started:new Date(t.started*1e3).toISOString(),timestamp:new Date(t.timestamp*1e3).toISOString(),status:t.status,errors:t.errors,did:typeof t.did=="number"||typeof t.did=="string"?`${t.did}`:void 0,duration:t.duration,abnormal_mechanism:t.abnormal_mechanism,attrs:{release:t.release,environment:t.environment,ip_address:t.ipAddress,user_agent:t.userAgent}}}function Me(t,e,n=2){if(!e||typeof e!="object"||n<=0)return e;if(t&&Object.keys(e).length===0)return t;const r=l({},t);for(const s in e)Object.prototype.hasOwnProperty.call(e,s)&&(r[s]=Me(r[s],e[s],n-1));return r}function me(){return W()}function ge(){return W().substring(16)}function Tr(t){try{const e=C.WeakRef;if(typeof e=="function")return new e(t)}catch(e){}return t}function vr(t){if(t){if(typeof t=="object"&&"deref"in t&&typeof t.deref=="function")try{return t.deref()}catch(e){return}return t}}const ln="_sentrySpan";function nt(t,e){e?Q(t,ln,Tr(e)):delete t[ln]}function xe(t){return vr(t[ln])}const Ji=100;class Y{constructor(){this._notifyingListeners=!1,this._scopeListeners=[],this._eventProcessors=[],this._breadcrumbs=[],this._attachments=[],this._user={},this._tags={},this._attributes={},this._extra={},this._contexts={},this._sdkProcessingMetadata={},this._propagationContext={traceId:me(),sampleRand:Ze()}}clone(){const e=new Y;return e._breadcrumbs=[...this._breadcrumbs],e._tags=l({},this._tags),e._attributes=l({},this._attributes),e._extra=l({},this._extra),e._contexts=l({},this._contexts),this._contexts.flags&&(e._contexts.flags={values:[...this._contexts.flags.values]}),e._user=this._user,e._level=this._level,e._session=this._session,e._transactionName=this._transactionName,e._fingerprint=this._fingerprint,e._eventProcessors=[...this._eventProcessors],e._attachments=[...this._attachments],e._sdkProcessingMetadata=l({},this._sdkProcessingMetadata),e._propagationContext=l({},this._propagationContext),e._client=this._client,e._lastEventId=this._lastEventId,e._conversationId=this._conversationId,nt(e,xe(this)),e}setClient(e){this._client=e}setLastEventId(e){this._lastEventId=e}getClient(){return this._client}lastEventId(){return this._lastEventId}addScopeListener(e){this._scopeListeners.push(e)}addEventProcessor(e){return this._eventProcessors.push(e),this}setUser(e){return this._user=e||{email:void 0,id:void 0,ip_address:void 0,username:void 0},this._session&&se(this._session,{user:e}),this._notifyScopeListeners(),this}getUser(){return this._user}setConversationId(e){return this._conversationId=e||void 0,this._notifyScopeListeners(),this}setTags(e){return this._tags=l(l({},this._tags),e),this._notifyScopeListeners(),this}setTag(e,n){return this.setTags({[e]:n})}setAttributes(e){return this._attributes=l(l({},this._attributes),e),this._notifyScopeListeners(),this}setAttribute(e,n){return this.setAttributes({[e]:n})}removeAttribute(e){return e in this._attributes&&(delete this._attributes[e],this._notifyScopeListeners()),this}setExtras(e){return this._extra=l(l({},this._extra),e),this._notifyScopeListeners(),this}setExtra(e,n){return this._extra=I(l({},this._extra),{[e]:n}),this._notifyScopeListeners(),this}setFingerprint(e){return this._fingerprint=e,this._notifyScopeListeners(),this}setLevel(e){return this._level=e,this._notifyScopeListeners(),this}setTransactionName(e){return this._transactionName=e,this._notifyScopeListeners(),this}setContext(e,n){return n===null?delete this._contexts[e]:this._contexts[e]=n,this._notifyScopeListeners(),this}setSession(e){return e?this._session=e:delete this._session,this._notifyScopeListeners(),this}getSession(){return this._session}update(e){if(!e)return this;const n=typeof e=="function"?e(this):e,r=n instanceof Y?n.getScopeData():sn(n)?e:void 0,{tags:s,attributes:i,extra:o,user:c,contexts:a,level:u,fingerprint:p=[],propagationContext:f,conversationId:_}=r||{};return this._tags=l(l({},this._tags),s),this._attributes=l(l({},this._attributes),i),this._extra=l(l({},this._extra),o),this._contexts=l(l({},this._contexts),a),c&&Object.keys(c).length&&(this._user=c),u&&(this._level=u),p.length&&(this._fingerprint=p),f&&(this._propagationContext=f),_&&(this._conversationId=_),this}clear(){return this._breadcrumbs=[],this._tags={},this._attributes={},this._extra={},this._user={},this._contexts={},this._level=void 0,this._transactionName=void 0,this._fingerprint=void 0,this._session=void 0,this._conversationId=void 0,nt(this,void 0),this._attachments=[],this.setPropagationContext({traceId:me(),sampleRand:Ze()}),this._notifyScopeListeners(),this}addBreadcrumb(e,n){var i;const r=typeof n=="number"?n:Ji;if(r<=0)return this;const s=I(l({timestamp:Oe()},e),{message:e.message?cn(e.message,2048):e.message});return this._breadcrumbs.push(s),this._breadcrumbs.length>r&&(this._breadcrumbs=this._breadcrumbs.slice(-r),(i=this._client)==null||i.recordDroppedEvent("buffer_overflow","log_item")),this._notifyScopeListeners(),this}getLastBreadcrumb(){return this._breadcrumbs[this._breadcrumbs.length-1]}clearBreadcrumbs(){return this._breadcrumbs=[],this._notifyScopeListeners(),this}addAttachment(e){return this._attachments.push(e),this}clearAttachments(){return this._attachments=[],this}getScopeData(){return{breadcrumbs:this._breadcrumbs,attachments:this._attachments,contexts:this._contexts,tags:this._tags,attributes:this._attributes,extra:this._extra,user:this._user,level:this._level,fingerprint:this._fingerprint||[],eventProcessors:this._eventProcessors,propagationContext:this._propagationContext,sdkProcessingMetadata:this._sdkProcessingMetadata,transactionName:this._transactionName,span:xe(this),conversationId:this._conversationId}}setSDKProcessingMetadata(e){return this._sdkProcessingMetadata=Me(this._sdkProcessingMetadata,e,2),this}setPropagationContext(e){return this._propagationContext=e,this}getPropagationContext(){return this._propagationContext}captureException(e,n){const r=(n==null?void 0:n.event_id)||W();if(!this._client)return y&&b.warn("No client configured on scope - will not capture exception!"),r;const s=new Error("Sentry syntheticException");return this._client.captureException(e,I(l({originalException:e,syntheticException:s},n),{event_id:r}),this),r}captureMessage(e,n,r){var o;const s=(r==null?void 0:r.event_id)||W();if(!this._client)return y&&b.warn("No client configured on scope - will not capture message!"),s;const i=(o=r==null?void 0:r.syntheticException)!=null?o:new Error(e);return this._client.captureMessage(e,n,I(l({originalException:e,syntheticException:i},r),{event_id:s}),this),s}captureEvent(e,n){const r=e.event_id||(n==null?void 0:n.event_id)||W();return this._client?(this._client.captureEvent(e,I(l({},n),{event_id:r}),this),r):(y&&b.warn("No client configured on scope - will not capture event!"),r)}_notifyScopeListeners(){this._notifyingListeners||(this._notifyingListeners=!0,this._scopeListeners.forEach(e=>{e(this)}),this._notifyingListeners=!1)}}function Yi(){return le("defaultCurrentScope",()=>new Y)}function Ki(){return le("defaultIsolationScope",()=>new Y)}const Ir=t=>t instanceof Promise&&!t[wr],wr=Symbol("chained PromiseLike"),kr=(t,e,n)=>{const r=t.then(s=>(e(s),s),s=>{throw n(s),s});return Ir(r)&&Ir(t)?r:Vi(t,r)},Vi=(t,e)=>{if(!e)return t;let n=!1;for(const r in t){if(r in e)continue;n=!0;const s=t[r];typeof s=="function"?Object.defineProperty(e,r,{value:(...i)=>s.apply(t,i),enumerable:!0,configurable:!0,writable:!0}):e[r]=s}return n&&Object.assign(e,{[wr]:!0}),e};class Xi{constructor(e,n){let r;e?r=e:r=new Y;let s;n?s=n:s=new Y,this._stack=[{scope:r}],this._isolationScope=s}withScope(e){const n=this._pushScope();let r;try{r=e(n)}catch(s){throw this._popScope(),s}return de(r)?kr(r,()=>this._popScope(),()=>this._popScope()):(this._popScope(),r)}getClient(){return this.getStackTop().client}getScope(){return this.getStackTop().scope}getIsolationScope(){return this._isolationScope}getStackTop(){return this._stack[this._stack.length-1]}_pushScope(){const e=this.getScope().clone();return this._stack.push({client:this.getClient(),scope:e}),e}_popScope(){return this._stack.length<=1?!1:!!this._stack.pop()}}function ye(){const t=O(),e=X(t);return e.stack=e.stack||new Xi(Yi(),Ki())}function Qi(t){return ye().withScope(t)}function Zi(t,e){const n=ye();return n.withScope(()=>(n.getStackTop().scope=t,e(t)))}function Ar(t){return ye().withScope(()=>t(ye().getIsolationScope()))}function eo(){return{withIsolationScope:Ar,withScope:Qi,withSetScope:Zi,withSetIsolationScope:(t,e)=>Ar(e),getCurrentScope:()=>ye().getScope(),getIsolationScope:()=>ye().getIsolationScope()}}function Se(t){const e=X(t);return e.acs?e.acs:eo()}function to(t){return typeof t=="object"&&t!=null&&!Array.isArray(t)&&Object.keys(t).includes("value")}function no(t,e){var c;const{value:n,unit:r}=to(t)?t:{value:t,unit:void 0},s=ro(n),i=r&&typeof r=="string"?{unit:r}:{};if(s)return l(l({},s),i);if(!e||e==="skip-undefined"&&n===void 0)return;let o="";try{o=(c=JSON.stringify(n))!=null?c:""}catch(a){}return l({value:o,type:"string"},i)}function rt(t,e=!1){const n={};for(const[r,s]of Object.entries(t!=null?t:{})){const i=no(s,e);i&&(n[r]=i)}return n}function ro(t){if(Array.isArray(t))return{value:t,type:"array"};const e=typeof t=="string"?"string":typeof t=="boolean"?"boolean":typeof t=="number"&&!Number.isNaN(t)?Number.isInteger(t)?"integer":"double":null;if(e)return{value:t,type:e}}function T(){const t=O();return Se(t).getCurrentScope()}function F(){const t=O();return Se(t).getIsolationScope()}function so(){return le("globalScope",()=>new Y)}function ie(...t){const e=O(),n=Se(e);if(t.length===2){const[r,s]=t;return r?n.withSetScope(r,s):n.withScope(s)}return n.withScope(t[0])}function P(){return T().getClient()}function Rr(t){const e=t.getPropagationContext(),{traceId:n,parentSpanId:r,propagationSpanId:s}=e,i={trace_id:n,span_id:s||ge()};return r&&(i.parent_span_id=r),i}const Fe="sentry.source",Cr="sentry.sample_rate",io="sentry.previous_trace_sample_rate",st="sentry.op",fn="sentry.origin",oo="sentry.status.message",Pr="sentry.measurement_unit",Nr="sentry.measurement_value",Dr="sentry.custom_span_name",dn="sentry.profile_id",pn="sentry.exclusive_time",Or=0,it=1,j=2;function ao(t){if(t<400&&t>=100)return{code:it};if(t>=400&&t<500)switch(t){case 401:return{code:j,message:"unauthenticated"};case 403:return{code:j,message:"permission_denied"};case 404:return{code:j,message:"not_found"};case 409:return{code:j,message:"already_exists"};case 413:return{code:j,message:"failed_precondition"};case 429:return{code:j,message:"resource_exhausted"};case 499:return{code:j,message:"cancelled"};default:return{code:j,message:"invalid_argument"}}if(t>=500&&t<600)switch(t){case 501:return{code:j,message:"unimplemented"};case 503:return{code:j,message:"unavailable"};case 504:return{code:j,message:"deadline_exceeded"};default:return{code:j,message:"internal_error"}}return{code:j,message:"internal_error"}}function co(t,e){t.setAttribute("http.response.status_code",e);const n=ao(e);n.message!=="unknown_error"&&t.setStatus(n)}const Mr="_sentryScope",xr="_sentryIsolationScope",uo=Symbol.for("sentry.otelSourceInference"),lo=Symbol.for("sentry.otelSourceExplicitlySet"),fo=Symbol.for("sentry.tracerProviderSpan");function Le(t,e,n){t&&(Q(t,xr,Tr(n)),Q(t,Mr,e))}function be(t){const e=t;return{scope:e[Mr],isolationScope:vr(e[xr])}}function Fr(t){return t[uo]===!0}function po(t){Q(t,lo,!0)}function Lr(t){return t[fo]===!0}const hn="sentry-",ho=8192;function _o(t){const e=go(t);if(!e)return;const n=Object.entries(e).reduce((r,[s,i])=>{if(s.startsWith(hn)){const o=s.slice(hn.length);r[o]=i}return r},{});if(Object.keys(n).length>0)return n}function mo(t){if(!t)return;const e=Object.entries(t).reduce((n,[r,s])=>(s&&(n[`${hn}${r}`]=s),n),{});return yo(e)}function go(t){if(!(!t||!nn(t)&&!Array.isArray(t)))return Array.isArray(t)?t.reduce((e,n)=>{const r=$r(n);return Object.entries(r).forEach(([s,i])=>{e[s]=i}),e},{}):$r(t)}function $r(t){return t.split(",").map(e=>{const n=e.indexOf("=");if(n===-1)return[];const r=e.slice(0,n),s=e.slice(n+1);return[r,s].map(i=>{try{return decodeURIComponent(i.trim())}catch(o){return}})}).reduce((e,[n,r])=>(n&&r&&(e[n]=r),e),{})}function yo(t){if(Object.keys(t).length!==0)return Object.entries(t).reduce((e,[n,r],s)=>{const i=`${encodeURIComponent(n)}=${encodeURIComponent(r)}`,o=s===0?i:`${e},${i}`;return o.length>ho?(y&&b.warn(`Not adding key: ${n} with val: ${r} to baggage header due to exceeding baggage size limits.`),e):o},"")}const So=/^o(\d+)\./,bo=/^(?:(\w+):)\/\/(?:(\w+)(?::(\w+)?)?@)((?:\[[:.%\w]+\]|[\w.-]+))(?::(\d+))?\/(.+)/;function Eo(t){return t==="http"||t==="https"}function Ee(t,e=!1){const{host:n,path:r,pass:s,port:i,projectId:o,protocol:c,publicKey:a}=t;return`${c}://${a}${e&&s?`:${s}`:""}@${n}${i?`:${i}`:""}/${r&&`${r}/`}${o}`}function To(t){const e=bo.exec(t);if(!e){fe(()=>{console.error(`Invalid Sentry Dsn: ${t}`)});return}const[n,r,s="",i="",o="",c=""]=e.slice(1);let a="",u=c;const p=u.split("/");if(p.length>1&&(a=p.slice(0,-1).join("/"),u=p.pop()),u){const f=u.match(/^\d+/);f&&(u=f[0])}return jr({host:i,pass:s,path:a,projectId:u,port:o,protocol:n,publicKey:r})}function jr(t){return{protocol:t.protocol,publicKey:t.publicKey||"",pass:t.pass||"",host:t.host,port:t.port||"",path:t.path||"",projectId:t.projectId}}function vo(t){if(!y)return!0;const{port:e,projectId:n,protocol:r}=t;return["protocol","publicKey","host","projectId"].find(o=>t[o]?!1:(b.error(`Invalid Sentry Dsn: ${o} missing`),!0))?!1:n.match(/^\d+$/)?Eo(r)?e&&isNaN(parseInt(e,10))?(b.error(`Invalid Sentry Dsn: Invalid port ${e}`),!1):!0:(b.error(`Invalid Sentry Dsn: Invalid protocol ${r}`),!1):(b.error(`Invalid Sentry Dsn: Invalid projectId ${n}`),!1)}function Io(t){const e=t.match(So);return e==null?void 0:e[1]}function wo(t){const e=t.getOptions(),{host:n}=t.getDsn()||{};let r;return e.orgId?r=String(e.orgId):n&&(r=Io(n)),r}function ko(t){const e=typeof t=="string"?To(t):jr(t);if(!(!e||!vo(e)))return e}function _n(t){if(typeof t=="boolean")return Number(t);const e=typeof t=="string"?parseFloat(t):t;if(!(typeof e!="number"||isNaN(e)||e<0||e>1))return e}const Ao=new RegExp("^[ \\t]*([0-9a-f]{32})?-?([0-9a-f]{16})?-?([01])?[ \\t]*$");function Br(t=me(),e=ge(),n){let r="";return n!==void 0&&(r=n?"-1":"-0"),`${t}-${e}${r}`}function Ur(t=me(),e=ge(),n){return`00-${t}-${e}-${n?"01":"00"}`}const Hr=0,ot=1;let zr=!1;function Ro(t){const{spanId:e,traceId:n}=t.spanContext(),{data:r,op:s,parent_span_id:i,status:o,origin:c,links:a}=B(t);return{parent_span_id:i,span_id:e,trace_id:n,data:r,op:s,status:o,origin:c,links:a}}function qr(t){const{spanId:e,traceId:n,isRemote:r}=t.spanContext(),s=r?e:B(t).parent_span_id,i=be(t).scope,o=r?(i==null?void 0:i.getPropagationContext().propagationSpanId)||ge():e;return{parent_span_id:s,span_id:o,trace_id:n}}function Co(t){const{traceId:e,spanId:n}=t.spanContext(),r=oe(t);return Br(e,n,r)}function Po(t){const{traceId:e,spanId:n}=t.spanContext(),r=oe(t);return Ur(e,n,r)}function Wr(t){if(t&&t.length>0)return t.map(o=>{var c=o,{context:a}=c,u=a,{spanId:e,traceId:n,traceFlags:r}=u,s=Si(u,["spanId","traceId","traceFlags"]),{attributes:i}=c;return l({span_id:e,trace_id:n,sampled:r===ot,attributes:i},s)})}function No(t){if(t!=null&&t.length)return t.map(({context:{spanId:e,traceId:n,traceFlags:r},attributes:s})=>({span_id:e,trace_id:n,sampled:r===ot,attributes:s}))}function Te(t){return typeof t=="number"?Gr(t):Array.isArray(t)?t[0]+t[1]/1e9:t instanceof Date?Gr(t.getTime()):_e()}function Gr(t){return t>9999999999?t/1e3:t}function B(t){if(xo(t))return t.getSpanJSON();const{spanId:e,traceId:n}=t.spanContext();if(Mo(t)){const{attributes:r,startTime:s,name:i,endTime:o,status:c,links:a}=t;return{span_id:e,trace_id:n,data:r,description:i,parent_span_id:Do(t),start_timestamp:Te(s),timestamp:Te(o)||void 0,status:Jr(c),op:r[st],origin:r[fn],links:Wr(a)}}return{span_id:e,trace_id:n,start_timestamp:0,data:{}}}function Do(t){var e;return"parentSpanId"in t?t.parentSpanId:"parentSpanContext"in t?(e=t.parentSpanContext)==null?void 0:e.spanId:void 0}function Oo(t){var e;return I(l({},t),{attributes:rt(t.attributes),links:(e=t.links)==null?void 0:e.map(n=>I(l({},n),{attributes:rt(n.attributes)}))})}function Mo(t){const e=t;return!!e.attributes&&!!e.startTime&&!!e.name&&!!e.endTime&&!!e.status}function xo(t){return typeof t.getSpanJSON=="function"}function oe(t){const{traceFlags:e}=t.spanContext();return e===ot}function Jr(t){if(!(!t||t.code===Or))return t.code===it?"ok":t.message||"internal_error"}function Yr(t){return!t||t.code===it||t.code===Or||t.message==="cancelled"?"ok":"error"}function Fo(t,e){const n=Yr(e)==="error"?e==null?void 0:e.message:void 0;return l(l({},n&&{[oo]:n}),t)}const $e="_sentryChildSpans",mn="_sentryRootSpan";function at(t,e){const n=t[mn]||t;Q(e,mn,n),t[$e]?t[$e].add(e):Q(t,$e,new Set([e]))}function Lo(t){const e=new Set;function n(r){if(!e.has(r)&&oe(r)){e.add(r);const s=r[$e]?Array.from(r[$e]):[];for(const i of s)n(i)}}return n(t),Array.from(e)}const K=$o;function $o(t){return t[mn]||t}function gn(){const t=O(),e=Se(t);return e.getActiveSpan?e.getActiveSpan():xe(T())}function yn(){zr||(fe(()=>{console.warn("[Sentry] Returning null from `beforeSendSpan` is disallowed. To drop certain spans, configure the respective integrations directly or use `ignoreSpans`.")}),zr=!0)}function je(t){var n;if(typeof __SENTRY_TRACING__=="boolean"&&!__SENTRY_TRACING__)return!1;const e=t||((n=P())==null?void 0:n.getOptions());return!!e&&(e.tracesSampleRate!=null||!!e.tracesSampler)}function Kr(t){b.log(`Ignoring span ${t.op} - ${t.description} because it matches \`ignoreSpans\`.`)}function ct(t,e){if(!(e!=null&&e.length))return!1;for(const n of e){if(Uo(n)){if(t.description&&De(t.description,n))return y&&Kr(t),!0;continue}const r=!!n.attributes&&Object.keys(n.attributes).length>0;if(!n.name&&!n.op&&!r)continue;const s=n.name?t.description&&De(t.description,n.name):!0,i=n.op?t.op&&De(t.op,n.op):!0,o=n.attributes?Object.entries(n.attributes).every(([c,a])=>{var u;return jo((u=t.attributes)==null?void 0:u[c],a)}):!0;if(s&&i&&o)return y&&Kr(t),!0}return!1}function jo(t,e){return typeof t=="string"&&(typeof e=="string"||e instanceof RegExp)?De(t,e):Array.isArray(t)&&Array.isArray(e)?t.length===e.length&&t.every((n,r)=>n===e[r]):t===e}function Bo(t,e){const n=e.parent_span_id,r=e.span_id;if(n)for(const s of t)s.parent_span_id===r&&(s.parent_span_id=n)}function Uo(t){return typeof t=="string"||t instanceof RegExp}const Vr=Symbol.for("sentry.nonRecordingSpan");class ut{constructor(e={}){this._traceId=e.traceId||me(),this._spanId=e.spanId||ge(),this.dropReason=e.dropReason,Q(this,Vr,!0)}spanContext(){return{spanId:this._spanId,traceId:this._traceId,traceFlags:Hr}}end(e){}setAttribute(e,n){return this}setAttributes(e){return this}setStatus(e){return this}updateName(e){return this}isRecording(){return!1}addEvent(e,n,r){return this}addLink(e){return this}addLinks(e){return this}recordException(e,n){}}function Be(t){return!!t&&t[Vr]===!0}const Sn="production",Xr="_frozenDsc";function Qr(t,e){Q(t,Xr,e)}function Zr(t,e){const n=e.getOptions(),{publicKey:r}=e.getDsn()||{},s={environment:n.environment||Sn,release:n.release,public_key:r,trace_id:t,org_id:wo(e)};return e.emit("createDsc",s),s}function lt(t,e){const n=e.getPropagationContext();return n.dsc||Zr(n.traceId,t)}function ve(t){var S,v,N,x,$;const e=P();if(!e)return{};const n=K(t),r=B(n),s=r.data,i=n.spanContext().traceState,o=(v=(S=i==null?void 0:i.get("sentry.sample_rate"))!=null?S:s[Cr])!=null?v:s[io];function c(k){return(typeof o=="number"||typeof o=="string")&&(k.sample_rate=`${o}`),k}const a=n[Xr];if(a)return c(a);const u=Be(n),p=u&&n.dropReason==="ignored";if(u&&(!je(e.getOptions())||p)){const k=be(n).scope;if(k){const J=l({},lt(e,k));return p&&(J.sampled="false"),c(J)}}const f=i==null?void 0:i.get("sentry.dsc"),_=f&&_o(f);if(_)return c(_);const d=Zr(t.spanContext().traceId,e),h=(N=s[Fe])!=null?N:s["sentry.segment.name.source"],m=r.description;return h!=="url"&&m&&(d.transaction=m),je()&&(d.sampled=String(oe(n)),d.sample_rand=($=i==null?void 0:i.get("sentry.sample_rand"))!=null?$:(x=be(n).scope)==null?void 0:x.getPropagationContext().sampleRand.toString()),c(d),e.emit("createDsc",d,n),d}function es(t){return!!t&&typeof t=="function"&&"_streamed"in t&&!!t._streamed}function ae(t,e=[]){return[t,e]}function ts(t,e){const[n,r]=t;return[n,[...r,e]]}function bn(t,e){const n=t[1];for(const r of n){const s=r[0].type;if(e(r,s))return!0}return!1}function En(t,e){return bn(t,(n,r)=>e.includes(r))}function Tn(t){const e=X(C);return e.encodePolyfill?e.encodePolyfill(t):new TextEncoder().encode(t)}function Ho(t){const[e,n]=t;let r=JSON.stringify(e);function s(i){typeof r=="string"?r=typeof i=="string"?r+i:[Tn(r),i]:r.push(typeof i=="string"?Tn(i):i)}for(const i of n){const[o,c]=i;if(s(`
+(function(g,E){typeof exports=="object"&&typeof module!="undefined"?E(exports):typeof define=="function"&&define.amd?define(["exports"],E):(g=typeof globalThis!="undefined"?globalThis:g||self,E(g.SentryMiniapp={}))})(this,(function(g){"use strict";var ml=Object.defineProperty,gl=Object.defineProperties;var yl=Object.getOwnPropertyDescriptors;var Xt=Object.getOwnPropertySymbols;var gi=Object.prototype.hasOwnProperty,yi=Object.prototype.propertyIsEnumerable;var or=(g,E)=>(E=Symbol[g])?E:Symbol.for("Symbol."+g),Sl=g=>{throw TypeError(g)};var mi=(g,E,R)=>E in g?ml(g,E,{enumerable:!0,configurable:!0,writable:!0,value:R}):g[E]=R,l=(g,E)=>{for(var R in E||(E={}))gi.call(E,R)&&mi(g,R,E[R]);if(Xt)for(var R of Xt(E))yi.call(E,R)&&mi(g,R,E[R]);return g},I=(g,E)=>gl(g,yl(E));var Si=(g,E)=>{var R={};for(var O in g)gi.call(g,O)&&E.indexOf(O)<0&&(R[O]=g[O]);if(g!=null&&Xt)for(var O of Xt(g))E.indexOf(O)<0&&yi.call(g,O)&&(R[O]=g[O]);return R};var z=(g,E,R)=>new Promise((O,U)=>{var y=D=>{try{q(R.next(D))}catch(X){U(X)}},C=D=>{try{q(R.throw(D))}catch(X){U(X)}},q=D=>D.done?O(D.value):Promise.resolve(D.value).then(y,C);q((R=R.apply(g,E)).next())}),bl=function(g,E){this[0]=g,this[1]=E};var bi=g=>{var E=g[or("asyncIterator")],R=!1,O,U={};return E==null?(E=g[or("iterator")](),O=y=>U[y]=C=>E[y](C)):(E=E.call(g),O=y=>U[y]=C=>{if(R){if(R=!1,y==="throw")throw C;return C}return R=!0,{done:!1,value:new bl(new Promise(q=>{var D=E[y](C);D instanceof Object||Sl("Object expected"),q(D)}),1)}}),U[or("iterator")]=()=>U,O("next"),"throw"in E?O("throw"):U.throw=y=>{throw y},"return"in E&&O("return"),U};class E{constructor(e){if(this._entries=[],typeof e=="string")this._parseString(e);else if(Array.isArray(e))for(const n of e)Array.isArray(n)&&n.length>=2&&this._entries.push([n[0]||"",n[1]||""]);else if(e&&typeof e=="object")if(e instanceof E)this._entries=e._entries.map(([n,r])=>[n,r]);else for(const[n,r]of Object.entries(e))this._entries.push([n,r])}get size(){return this._entries.length}_parseString(e){if(e.startsWith("?")&&(e=e.slice(1)),!e)return;const n=e.split("&");for(const r of n){const s=r.indexOf("=");if(s===-1)r&&this._entries.push([decodeURIComponent(r),""]);else{const i=r.slice(0,s),o=r.slice(s+1);i&&this._entries.push([decodeURIComponent(i),decodeURIComponent(o)])}}}append(e,n){this._entries.push([e,String(n)])}delete(e){this._entries=this._entries.filter(([n])=>n!==e)}get(e){const n=this._entries.find(([r])=>r===e);return n?n[1]:null}getAll(e){return this._entries.filter(([n])=>n===e).map(([,n])=>n)}has(e){return this._entries.some(([n])=>n===e)}set(e,n){const r=String(n);let s=!1;if(this._entries=this._entries.filter(([i])=>i===e?s?!1:(s=!0,!0):!0),s){const i=this._entries.findIndex(([o])=>o===e);i!==-1&&(this._entries[i]=[e,r])}else this._entries.push([e,r])}sort(){this._entries.sort(([e],[n])=>e<n?-1:e>n?1:0)}toString(){return this._entries.map(([e,n])=>`${encodeURIComponent(e)}=${encodeURIComponent(n)}`).join("&")}forEach(e){for(const[n,r]of this._entries)e(r,n,this)}*keys(){for(const[e]of this._entries)yield e}*values(){for(const[,e]of this._entries)yield e}*entries(){for(const e of this._entries)yield e}*[Symbol.iterator](){yield*bi(this.entries())}}function R(){if(typeof globalThis!="undefined")return globalThis;try{return Function("return this")()}catch(t){}if(typeof window!="undefined")return window;if(typeof global!="undefined")return global;if(typeof self!="undefined")return self}function O(){try{const t=R();if(!t){console.warn("[sentry-miniapp] Unable to detect global object, polyfills may not work correctly");return}typeof t.URLSearchParams=="undefined"&&(t.URLSearchParams=E);const e=Function("return this")();e&&typeof e.URLSearchParams=="undefined"&&(e.URLSearchParams=E),typeof globalThis!="undefined"&&typeof globalThis.URLSearchParams=="undefined"&&(globalThis.URLSearchParams=E)}catch(t){console.warn("[sentry-miniapp] Failed to install polyfills:",t)}}function U(){O()}U();const y=typeof __SENTRY_DEBUG__=="undefined"||__SENTRY_DEBUG__,C=globalThis,q="10.68.0";function D(){return X(C),C}function X(t){const e=t.__SENTRY__=t.__SENTRY__||{};return e.version=e.version||q,e[q]=e[q]||{}}function le(t,e,n=C){const r=n.__SENTRY__=n.__SENTRY__||{},s=r[q]=r[q]||{};return s[t]||(s[t]=e())}const Ei="Sentry Logger ",ar={};function fe(t){if(!("console"in C))return t();const e=C.console,n={},r=Object.keys(ar);r.forEach(s=>{const i=ar[s];n[s]=e[s],e[s]=i});try{return t()}finally{r.forEach(s=>{e[s]=n[s]})}}function Ti(){Zt().enabled=!0}function vi(){Zt().enabled=!1}function cr(){return Zt().enabled}function Ii(...t){Qt("log",...t)}function wi(...t){Qt("warn",...t)}function ki(...t){Qt("error",...t)}function Qt(t,...e){y&&cr()&&fe(()=>{C.console[t](`${Ei}[${t}]:`,...e)})}function Zt(){return y?le("loggerSettings",()=>({enabled:!1})):{enabled:!1}}const b={enable:Ti,disable:vi,isEnabled:cr,log:Ii,warn:wi,error:ki},ur=50,Ne="?",lr=/\(error: (.*)\)/,fr=/captureMessage|captureException/;function Ai(...t){const e=t.sort((n,r)=>n[0]-r[0]).map(n=>n[1]);return(n,r=0,s=0)=>{const i=[],o=n.split(`
+`);for(let c=r;c<o.length;c++){let a=o[c];a.length>1024&&(a=a.slice(0,1024));const u=lr.test(a)?a.replace(lr,"$1"):a;if(!u.includes("Error: ")){for(const p of e){const f=p(u);if(f){i.push(f);break}}if(i.length>=ur+s)break}}return Ri(i.slice(s))}}function Ri(t){if(!t.length)return[];const e=Array.from(t);return/sentryWrapped/.test(Xe(e).function||"")&&e.pop(),e.reverse(),fr.test(Xe(e).function||"")&&(e.pop(),fr.test(Xe(e).function||"")&&e.pop()),e.slice(0,ur).map(n=>I(l({},n),{filename:n.filename||Xe(e).filename,function:n.function||Ne}))}function Xe(t){return t[t.length-1]||{}}const en="<anonymous>";function Ci(t){try{return!t||typeof t!="function"?en:t.name||en}catch(e){return en}}const dr=Object.prototype.toString;function pr(t){switch(dr.call(t)){case"[object Error]":case"[object Exception]":case"[object DOMException]":case"[object WebAssembly.Exception]":return!0;default:return _r(t,Error)}}function tn(t,e){return dr.call(t)===`[object ${e}]`}function nn(t){return tn(t,"String")}function rn(t){return typeof t=="object"&&t!==null&&"__sentry_template_string__"in t&&"__sentry_template_values__"in t}function hr(t){return t===null||rn(t)||typeof t!="object"&&typeof t!="function"}function sn(t){return tn(t,"Object")}function on(t){return typeof t=="object"&&t!==null}function Pi(t){return typeof Event!="undefined"&&_r(t,Event)}function Ni(t){return tn(t,"RegExp")}function de(t){return!!(t!=null&&t.then&&typeof t.then=="function")}function _r(t,e){try{return t instanceof e}catch(n){return!1}}function Q(t,e,n){try{Object.defineProperty(t,e,{value:n,writable:!0,configurable:!0})}catch(r){y&&b.log(`Failed to add non-enumerable property "${String(e)}" to object`,t)}}function Oi(t){if(pr(t))return l({message:t.message,name:t.name,stack:t.stack},mr(t));if(Pi(t)){const{type:e,target:n,currentTarget:r,detail:s}=t;return l(l({type:e,target:n,currentTarget:r},s?{detail:s}:{}),mr(t))}return t}function mr(t){return on(t)?Object.fromEntries(Object.entries(t)):{}}let pe;function Qe(t){if(pe!==void 0)return pe?pe(t):t();const e=Symbol.for("__SENTRY_SAFE_RANDOM_ID_WRAPPER__"),n=C;return e in n&&typeof n[e]=="function"?(pe=n[e],pe(t)):(pe=null,t())}function Ze(){return Qe(()=>Math.random())}function ee(){return Qe(()=>Date.now())}const Di=Symbol.for("sentry.skipNormalization"),Mi=Symbol.for("sentry.overrideNormalizationDepth");function xi(t){return!!t[Di]}function Fi(t){const e=t[Mi];return typeof e=="number"?e:void 0}let Li;function te(t,e=100,n=1/0){try{return an("",t,e,n)}catch(r){return{ERROR:`**non-serializable** (${r})`}}}function an(t,e,n=1/0,r=1/0,s=Bi()){const[i,o]=s;if(e==null||["boolean","string"].includes(typeof e)||typeof e=="number"&&Number.isFinite(e))return e;const c=$i(t,e);if(!c.startsWith("[object "))return c;if(xi(e))return e;const a=Fi(e),u=a!==void 0?a:n;if(u===0)return c.replace("object ","");if(i(e))return"[Circular ~]";const p=e;if(p&&typeof p.toJSON=="function")try{const h=p.toJSON();return an("",h,u-1,r,s)}catch(h){}const f=Array.isArray(e)?[]:{};let _=0;const d=Oi(e);for(const h in d){if(!Object.prototype.hasOwnProperty.call(d,h))continue;if(_>=r){f[h]="[MaxProperties ~]";break}const m=d[h];f[h]=an(h,m,u-1,r,s),_++}return o(e),f}function $i(t,e){try{return typeof global!="undefined"&&e===global?"[Global]":typeof e=="number"&&!Number.isFinite(e)?`[${e}]`:typeof e=="function"?`[Function: ${Ci(e)}]`:typeof e=="symbol"?`[${String(e)}]`:typeof e=="bigint"?`[BigInt: ${String(e)}]`:`[object ${ji(e)}]`}catch(n){return`**non-serializable** (${n})`}}function ji(t){const e=Object.getPrototypeOf(t);return e!=null&&e.constructor?e.constructor.name:"null prototype"}function Bi(){const t=new WeakSet;function e(r){return t.has(r)?!0:(t.add(r),!1)}function n(r){t.delete(r)}return[e,n]}function cn(t,e=0){return typeof t!="string"||e===0||t.length<=e?t:`${t.slice(0,e)}...`}function Oe(t,e,n=!1){return nn(t)?Ni(e)?e.test(t):nn(e)?n?t===e:t.includes(e):typeof e=="function"?e(t):!1:!1}function et(t,e=[],n=!1){for(const r of e)if(Oe(t,r,n))return!0;return!1}function Ui(){const t=C;return t.crypto||t.msCrypto}let un;function Hi(){return Ze()*16}function W(t=Ui()){try{if(t!=null&&t.randomUUID)return Qe(()=>t.randomUUID()).replace(/-/g,"")}catch(e){}return un||(un="10000000100040008000"+1e11),un.replace(/[018]/g,e=>(e^(Hi()&15)>>e/4).toString(16))}function gr(t){var e,n;return(n=(e=t.exception)==null?void 0:e.values)==null?void 0:n[0]}function he(t){const{message:e,event_id:n}=t;if(e)return e;const r=gr(t);return r?r.type&&r.value?`${r.type}: ${r.value}`:r.type||r.value||n||"<unknown>":n||"<unknown>"}function zi(t,e){const n=gr(t);if(!n)return;const r={type:"generic",handled:!0},s=n.mechanism;if(n.mechanism=l(l(l({},r),s),e),e&&"data"in e){const i=l(l({},s==null?void 0:s.data),e.data);n.mechanism.data=i}}function yr(t){if(qi(t))return!0;try{Q(t,"__sentry_captured__",!0)}catch(e){}return!1}function qi(t){try{return t.__sentry_captured__}catch(e){}}const Sr=1e3;function De(){return ee()/Sr}function Wi(){const{performance:t}=C;if(!(t!=null&&t.now)||!t.timeOrigin)return De;const e=t.timeOrigin;return()=>(e+Qe(()=>t.now()))/Sr}let tt;function _e(){return(tt!=null?tt:tt=Wi())()}function br(t){const e=_e(),n={sid:W(),init:!0,timestamp:e,started:e,duration:0,status:"ok",errors:0,ignoreDuration:!1,toJSON:()=>Gi(n)};return t&&se(n,t),n}function se(t,e={}){if(e.user&&(!t.ipAddress&&e.user.ip_address&&(t.ipAddress=e.user.ip_address),!t.did&&!e.did&&(t.did=e.user.id||e.user.email||e.user.username)),t.timestamp=e.timestamp||_e(),e.abnormal_mechanism&&(t.abnormal_mechanism=e.abnormal_mechanism),e.ignoreDuration&&(t.ignoreDuration=e.ignoreDuration),e.sid&&(t.sid=e.sid.length===32?e.sid:W()),e.init!==void 0&&(t.init=e.init),!t.did&&e.did&&(t.did=`${e.did}`),typeof e.started=="number"&&(t.started=e.started),t.ignoreDuration)t.duration=void 0;else if(typeof e.duration=="number")t.duration=e.duration;else{const n=t.timestamp-t.started;t.duration=n>=0?n:0}e.release&&(t.release=e.release),e.environment&&(t.environment=e.environment),!t.ipAddress&&e.ipAddress&&(t.ipAddress=e.ipAddress),!t.userAgent&&e.userAgent&&(t.userAgent=e.userAgent),typeof e.errors=="number"&&(t.errors=e.errors),e.status&&(t.status=e.status)}function Er(t,e){let n={};e?n={status:e}:t.status==="ok"&&(n={status:"exited"}),se(t,n)}function Gi(t){return{sid:`${t.sid}`,init:t.init,started:new Date(t.started*1e3).toISOString(),timestamp:new Date(t.timestamp*1e3).toISOString(),status:t.status,errors:t.errors,did:typeof t.did=="number"||typeof t.did=="string"?`${t.did}`:void 0,duration:t.duration,abnormal_mechanism:t.abnormal_mechanism,attrs:{release:t.release,environment:t.environment,ip_address:t.ipAddress,user_agent:t.userAgent}}}function Me(t,e,n=2){if(!e||typeof e!="object"||n<=0)return e;if(t&&Object.keys(e).length===0)return t;const r=l({},t);for(const s in e)Object.prototype.hasOwnProperty.call(e,s)&&(r[s]=Me(r[s],e[s],n-1));return r}function me(){return W()}function ge(){return W().substring(16)}function Tr(t){try{const e=C.WeakRef;if(typeof e=="function")return new e(t)}catch(e){}return t}function vr(t){if(t){if(typeof t=="object"&&"deref"in t&&typeof t.deref=="function")try{return t.deref()}catch(e){return}return t}}const ln="_sentrySpan";function nt(t,e){e?Q(t,ln,Tr(e)):delete t[ln]}function xe(t){return vr(t[ln])}const Ji=100;class Y{constructor(){this._notifyingListeners=!1,this._scopeListeners=[],this._eventProcessors=[],this._breadcrumbs=[],this._attachments=[],this._user={},this._tags={},this._attributes={},this._extra={},this._contexts={},this._sdkProcessingMetadata={},this._propagationContext={traceId:me(),sampleRand:Ze()}}clone(){const e=new Y;return e._breadcrumbs=[...this._breadcrumbs],e._tags=l({},this._tags),e._attributes=l({},this._attributes),e._extra=l({},this._extra),e._contexts=l({},this._contexts),this._contexts.flags&&(e._contexts.flags={values:[...this._contexts.flags.values]}),e._user=this._user,e._level=this._level,e._session=this._session,e._transactionName=this._transactionName,e._fingerprint=this._fingerprint,e._eventProcessors=[...this._eventProcessors],e._attachments=[...this._attachments],e._sdkProcessingMetadata=l({},this._sdkProcessingMetadata),e._propagationContext=l({},this._propagationContext),e._client=this._client,e._lastEventId=this._lastEventId,e._conversationId=this._conversationId,nt(e,xe(this)),e}setClient(e){this._client=e}setLastEventId(e){this._lastEventId=e}getClient(){return this._client}lastEventId(){return this._lastEventId}addScopeListener(e){this._scopeListeners.push(e)}addEventProcessor(e){return this._eventProcessors.push(e),this}setUser(e){return this._user=e||{email:void 0,id:void 0,ip_address:void 0,username:void 0},this._session&&se(this._session,{user:e}),this._notifyScopeListeners(),this}getUser(){return this._user}setConversationId(e){return this._conversationId=e||void 0,this._notifyScopeListeners(),this}setTags(e){return this._tags=l(l({},this._tags),e),this._notifyScopeListeners(),this}setTag(e,n){return this.setTags({[e]:n})}setAttributes(e){return this._attributes=l(l({},this._attributes),e),this._notifyScopeListeners(),this}setAttribute(e,n){return this.setAttributes({[e]:n})}removeAttribute(e){return e in this._attributes&&(delete this._attributes[e],this._notifyScopeListeners()),this}setExtras(e){return this._extra=l(l({},this._extra),e),this._notifyScopeListeners(),this}setExtra(e,n){return this._extra=I(l({},this._extra),{[e]:n}),this._notifyScopeListeners(),this}setFingerprint(e){return this._fingerprint=e,this._notifyScopeListeners(),this}setLevel(e){return this._level=e,this._notifyScopeListeners(),this}setTransactionName(e){return this._transactionName=e,this._notifyScopeListeners(),this}setContext(e,n){return n===null?delete this._contexts[e]:this._contexts[e]=n,this._notifyScopeListeners(),this}setSession(e){return e?this._session=e:delete this._session,this._notifyScopeListeners(),this}getSession(){return this._session}update(e){if(!e)return this;const n=typeof e=="function"?e(this):e,r=n instanceof Y?n.getScopeData():sn(n)?e:void 0,{tags:s,attributes:i,extra:o,user:c,contexts:a,level:u,fingerprint:p=[],propagationContext:f,conversationId:_}=r||{};return this._tags=l(l({},this._tags),s),this._attributes=l(l({},this._attributes),i),this._extra=l(l({},this._extra),o),this._contexts=l(l({},this._contexts),a),c&&Object.keys(c).length&&(this._user=c),u&&(this._level=u),p.length&&(this._fingerprint=p),f&&(this._propagationContext=f),_&&(this._conversationId=_),this}clear(){return this._breadcrumbs=[],this._tags={},this._attributes={},this._extra={},this._user={},this._contexts={},this._level=void 0,this._transactionName=void 0,this._fingerprint=void 0,this._session=void 0,this._conversationId=void 0,nt(this,void 0),this._attachments=[],this.setPropagationContext({traceId:me(),sampleRand:Ze()}),this._notifyScopeListeners(),this}addBreadcrumb(e,n){var i;const r=typeof n=="number"?n:Ji;if(r<=0)return this;const s=I(l({timestamp:De()},e),{message:e.message?cn(e.message,2048):e.message});return this._breadcrumbs.push(s),this._breadcrumbs.length>r&&(this._breadcrumbs=this._breadcrumbs.slice(-r),(i=this._client)==null||i.recordDroppedEvent("buffer_overflow","log_item")),this._notifyScopeListeners(),this}getLastBreadcrumb(){return this._breadcrumbs[this._breadcrumbs.length-1]}clearBreadcrumbs(){return this._breadcrumbs=[],this._notifyScopeListeners(),this}addAttachment(e){return this._attachments.push(e),this}clearAttachments(){return this._attachments=[],this}getScopeData(){return{breadcrumbs:this._breadcrumbs,attachments:this._attachments,contexts:this._contexts,tags:this._tags,attributes:this._attributes,extra:this._extra,user:this._user,level:this._level,fingerprint:this._fingerprint||[],eventProcessors:this._eventProcessors,propagationContext:this._propagationContext,sdkProcessingMetadata:this._sdkProcessingMetadata,transactionName:this._transactionName,span:xe(this),conversationId:this._conversationId}}setSDKProcessingMetadata(e){return this._sdkProcessingMetadata=Me(this._sdkProcessingMetadata,e,2),this}setPropagationContext(e){return this._propagationContext=e,this}getPropagationContext(){return this._propagationContext}captureException(e,n){const r=(n==null?void 0:n.event_id)||W();if(!this._client)return y&&b.warn("No client configured on scope - will not capture exception!"),r;const s=new Error("Sentry syntheticException");return this._client.captureException(e,I(l({originalException:e,syntheticException:s},n),{event_id:r}),this),r}captureMessage(e,n,r){var o;const s=(r==null?void 0:r.event_id)||W();if(!this._client)return y&&b.warn("No client configured on scope - will not capture message!"),s;const i=(o=r==null?void 0:r.syntheticException)!=null?o:new Error(e);return this._client.captureMessage(e,n,I(l({originalException:e,syntheticException:i},r),{event_id:s}),this),s}captureEvent(e,n){const r=e.event_id||(n==null?void 0:n.event_id)||W();return this._client?(this._client.captureEvent(e,I(l({},n),{event_id:r}),this),r):(y&&b.warn("No client configured on scope - will not capture event!"),r)}_notifyScopeListeners(){this._notifyingListeners||(this._notifyingListeners=!0,this._scopeListeners.forEach(e=>{e(this)}),this._notifyingListeners=!1)}}function Yi(){return le("defaultCurrentScope",()=>new Y)}function Ki(){return le("defaultIsolationScope",()=>new Y)}const Ir=t=>t instanceof Promise&&!t[wr],wr=Symbol("chained PromiseLike"),kr=(t,e,n)=>{const r=t.then(s=>(e(s),s),s=>{throw n(s),s});return Ir(r)&&Ir(t)?r:Vi(t,r)},Vi=(t,e)=>{if(!e)return t;let n=!1;for(const r in t){if(r in e)continue;n=!0;const s=t[r];typeof s=="function"?Object.defineProperty(e,r,{value:(...i)=>s.apply(t,i),enumerable:!0,configurable:!0,writable:!0}):e[r]=s}return n&&Object.assign(e,{[wr]:!0}),e};class Xi{constructor(e,n){let r;e?r=e:r=new Y;let s;n?s=n:s=new Y,this._stack=[{scope:r}],this._isolationScope=s}withScope(e){const n=this._pushScope();let r;try{r=e(n)}catch(s){throw this._popScope(),s}return de(r)?kr(r,()=>this._popScope(),()=>this._popScope()):(this._popScope(),r)}getClient(){return this.getStackTop().client}getScope(){return this.getStackTop().scope}getIsolationScope(){return this._isolationScope}getStackTop(){return this._stack[this._stack.length-1]}_pushScope(){const e=this.getScope().clone();return this._stack.push({client:this.getClient(),scope:e}),e}_popScope(){return this._stack.length<=1?!1:!!this._stack.pop()}}function ye(){const t=D(),e=X(t);return e.stack=e.stack||new Xi(Yi(),Ki())}function Qi(t){return ye().withScope(t)}function Zi(t,e){const n=ye();return n.withScope(()=>(n.getStackTop().scope=t,e(t)))}function Ar(t){return ye().withScope(()=>t(ye().getIsolationScope()))}function eo(){return{withIsolationScope:Ar,withScope:Qi,withSetScope:Zi,withSetIsolationScope:(t,e)=>Ar(e),getCurrentScope:()=>ye().getScope(),getIsolationScope:()=>ye().getIsolationScope()}}function Se(t){const e=X(t);return e.acs?e.acs:eo()}function to(t){return typeof t=="object"&&t!=null&&!Array.isArray(t)&&Object.keys(t).includes("value")}function no(t,e){var c;const{value:n,unit:r}=to(t)?t:{value:t,unit:void 0},s=ro(n),i=r&&typeof r=="string"?{unit:r}:{};if(s)return l(l({},s),i);if(!e||e==="skip-undefined"&&n===void 0)return;let o="";try{o=(c=JSON.stringify(n))!=null?c:""}catch(a){}return l({value:o,type:"string"},i)}function rt(t,e=!1){const n={};for(const[r,s]of Object.entries(t!=null?t:{})){const i=no(s,e);i&&(n[r]=i)}return n}function ro(t){if(Array.isArray(t))return{value:t,type:"array"};const e=typeof t=="string"?"string":typeof t=="boolean"?"boolean":typeof t=="number"&&!Number.isNaN(t)?Number.isInteger(t)?"integer":"double":null;if(e)return{value:t,type:e}}function T(){const t=D();return Se(t).getCurrentScope()}function F(){const t=D();return Se(t).getIsolationScope()}function so(){return le("globalScope",()=>new Y)}function ie(...t){const e=D(),n=Se(e);if(t.length===2){const[r,s]=t;return r?n.withSetScope(r,s):n.withScope(s)}return n.withScope(t[0])}function P(){return T().getClient()}function Rr(t){const e=t.getPropagationContext(),{traceId:n,parentSpanId:r,propagationSpanId:s}=e,i={trace_id:n,span_id:s||ge()};return r&&(i.parent_span_id=r),i}const Fe="sentry.source",Cr="sentry.sample_rate",io="sentry.previous_trace_sample_rate",st="sentry.op",fn="sentry.origin",oo="sentry.status.message",Pr="sentry.measurement_unit",Nr="sentry.measurement_value",Or="sentry.custom_span_name",dn="sentry.profile_id",pn="sentry.exclusive_time",Dr=0,it=1,j=2;function ao(t){if(t<400&&t>=100)return{code:it};if(t>=400&&t<500)switch(t){case 401:return{code:j,message:"unauthenticated"};case 403:return{code:j,message:"permission_denied"};case 404:return{code:j,message:"not_found"};case 409:return{code:j,message:"already_exists"};case 413:return{code:j,message:"failed_precondition"};case 429:return{code:j,message:"resource_exhausted"};case 499:return{code:j,message:"cancelled"};default:return{code:j,message:"invalid_argument"}}if(t>=500&&t<600)switch(t){case 501:return{code:j,message:"unimplemented"};case 503:return{code:j,message:"unavailable"};case 504:return{code:j,message:"deadline_exceeded"};default:return{code:j,message:"internal_error"}}return{code:j,message:"internal_error"}}function co(t,e){t.setAttribute("http.response.status_code",e);const n=ao(e);n.message!=="unknown_error"&&t.setStatus(n)}const Mr="_sentryScope",xr="_sentryIsolationScope",uo=Symbol.for("sentry.otelSourceInference"),lo=Symbol.for("sentry.otelSourceExplicitlySet"),fo=Symbol.for("sentry.tracerProviderSpan");function Le(t,e,n){t&&(Q(t,xr,Tr(n)),Q(t,Mr,e))}function be(t){const e=t;return{scope:e[Mr],isolationScope:vr(e[xr])}}function Fr(t){return t[uo]===!0}function po(t){Q(t,lo,!0)}function Lr(t){return t[fo]===!0}const hn="sentry-",ho=8192;function _o(t){const e=go(t);if(!e)return;const n=Object.entries(e).reduce((r,[s,i])=>{if(s.startsWith(hn)){const o=s.slice(hn.length);r[o]=i}return r},{});if(Object.keys(n).length>0)return n}function mo(t){if(!t)return;const e=Object.entries(t).reduce((n,[r,s])=>(s&&(n[`${hn}${r}`]=s),n),{});return yo(e)}function go(t){if(!(!t||!nn(t)&&!Array.isArray(t)))return Array.isArray(t)?t.reduce((e,n)=>{const r=$r(n);return Object.entries(r).forEach(([s,i])=>{e[s]=i}),e},{}):$r(t)}function $r(t){return t.split(",").map(e=>{const n=e.indexOf("=");if(n===-1)return[];const r=e.slice(0,n),s=e.slice(n+1);return[r,s].map(i=>{try{return decodeURIComponent(i.trim())}catch(o){return}})}).reduce((e,[n,r])=>(n&&r&&(e[n]=r),e),{})}function yo(t){if(Object.keys(t).length!==0)return Object.entries(t).reduce((e,[n,r],s)=>{const i=`${encodeURIComponent(n)}=${encodeURIComponent(r)}`,o=s===0?i:`${e},${i}`;return o.length>ho?(y&&b.warn(`Not adding key: ${n} with val: ${r} to baggage header due to exceeding baggage size limits.`),e):o},"")}const So=/^o(\d+)\./,bo=/^(?:(\w+):)\/\/(?:(\w+)(?::(\w+)?)?@)((?:\[[:.%\w]+\]|[\w.-]+))(?::(\d+))?\/(.+)/;function Eo(t){return t==="http"||t==="https"}function Ee(t,e=!1){const{host:n,path:r,pass:s,port:i,projectId:o,protocol:c,publicKey:a}=t;return`${c}://${a}${e&&s?`:${s}`:""}@${n}${i?`:${i}`:""}/${r&&`${r}/`}${o}`}function To(t){const e=bo.exec(t);if(!e){fe(()=>{console.error(`Invalid Sentry Dsn: ${t}`)});return}const[n,r,s="",i="",o="",c=""]=e.slice(1);let a="",u=c;const p=u.split("/");if(p.length>1&&(a=p.slice(0,-1).join("/"),u=p.pop()),u){const f=u.match(/^\d+/);f&&(u=f[0])}return jr({host:i,pass:s,path:a,projectId:u,port:o,protocol:n,publicKey:r})}function jr(t){return{protocol:t.protocol,publicKey:t.publicKey||"",pass:t.pass||"",host:t.host,port:t.port||"",path:t.path||"",projectId:t.projectId}}function vo(t){if(!y)return!0;const{port:e,projectId:n,protocol:r}=t;return["protocol","publicKey","host","projectId"].find(o=>t[o]?!1:(b.error(`Invalid Sentry Dsn: ${o} missing`),!0))?!1:n.match(/^\d+$/)?Eo(r)?e&&isNaN(parseInt(e,10))?(b.error(`Invalid Sentry Dsn: Invalid port ${e}`),!1):!0:(b.error(`Invalid Sentry Dsn: Invalid protocol ${r}`),!1):(b.error(`Invalid Sentry Dsn: Invalid projectId ${n}`),!1)}function Io(t){const e=t.match(So);return e==null?void 0:e[1]}function wo(t){const e=t.getOptions(),{host:n}=t.getDsn()||{};let r;return e.orgId?r=String(e.orgId):n&&(r=Io(n)),r}function ko(t){const e=typeof t=="string"?To(t):jr(t);if(!(!e||!vo(e)))return e}function _n(t){if(typeof t=="boolean")return Number(t);const e=typeof t=="string"?parseFloat(t):t;if(!(typeof e!="number"||isNaN(e)||e<0||e>1))return e}const Ao=new RegExp("^[ \\t]*([0-9a-f]{32})?-?([0-9a-f]{16})?-?([01])?[ \\t]*$");function Br(t=me(),e=ge(),n){let r="";return n!==void 0&&(r=n?"-1":"-0"),`${t}-${e}${r}`}function Ur(t=me(),e=ge(),n){return`00-${t}-${e}-${n?"01":"00"}`}const Hr=0,ot=1;let zr=!1;function Ro(t){const{spanId:e,traceId:n}=t.spanContext(),{data:r,op:s,parent_span_id:i,status:o,origin:c,links:a}=B(t);return{parent_span_id:i,span_id:e,trace_id:n,data:r,op:s,status:o,origin:c,links:a}}function qr(t){const{spanId:e,traceId:n,isRemote:r}=t.spanContext(),s=r?e:B(t).parent_span_id,i=be(t).scope,o=r?(i==null?void 0:i.getPropagationContext().propagationSpanId)||ge():e;return{parent_span_id:s,span_id:o,trace_id:n}}function Co(t){const{traceId:e,spanId:n}=t.spanContext(),r=oe(t);return Br(e,n,r)}function Po(t){const{traceId:e,spanId:n}=t.spanContext(),r=oe(t);return Ur(e,n,r)}function Wr(t){if(t&&t.length>0)return t.map(o=>{var c=o,{context:a}=c,u=a,{spanId:e,traceId:n,traceFlags:r}=u,s=Si(u,["spanId","traceId","traceFlags"]),{attributes:i}=c;return l({span_id:e,trace_id:n,sampled:r===ot,attributes:i},s)})}function No(t){if(t!=null&&t.length)return t.map(({context:{spanId:e,traceId:n,traceFlags:r},attributes:s})=>({span_id:e,trace_id:n,sampled:r===ot,attributes:s}))}function Te(t){return typeof t=="number"?Gr(t):Array.isArray(t)?t[0]+t[1]/1e9:t instanceof Date?Gr(t.getTime()):_e()}function Gr(t){return t>9999999999?t/1e3:t}function B(t){if(xo(t))return t.getSpanJSON();const{spanId:e,traceId:n}=t.spanContext();if(Mo(t)){const{attributes:r,startTime:s,name:i,endTime:o,status:c,links:a}=t;return{span_id:e,trace_id:n,data:r,description:i,parent_span_id:Oo(t),start_timestamp:Te(s),timestamp:Te(o)||void 0,status:Jr(c),op:r[st],origin:r[fn],links:Wr(a)}}return{span_id:e,trace_id:n,start_timestamp:0,data:{}}}function Oo(t){var e;return"parentSpanId"in t?t.parentSpanId:"parentSpanContext"in t?(e=t.parentSpanContext)==null?void 0:e.spanId:void 0}function Do(t){var e;return I(l({},t),{attributes:rt(t.attributes),links:(e=t.links)==null?void 0:e.map(n=>I(l({},n),{attributes:rt(n.attributes)}))})}function Mo(t){const e=t;return!!e.attributes&&!!e.startTime&&!!e.name&&!!e.endTime&&!!e.status}function xo(t){return typeof t.getSpanJSON=="function"}function oe(t){const{traceFlags:e}=t.spanContext();return e===ot}function Jr(t){if(!(!t||t.code===Dr))return t.code===it?"ok":t.message||"internal_error"}function Yr(t){return!t||t.code===it||t.code===Dr||t.message==="cancelled"?"ok":"error"}function Fo(t,e){const n=Yr(e)==="error"?e==null?void 0:e.message:void 0;return l(l({},n&&{[oo]:n}),t)}const $e="_sentryChildSpans",mn="_sentryRootSpan";function at(t,e){const n=t[mn]||t;Q(e,mn,n),t[$e]?t[$e].add(e):Q(t,$e,new Set([e]))}function Lo(t){const e=new Set;function n(r){if(!e.has(r)&&oe(r)){e.add(r);const s=r[$e]?Array.from(r[$e]):[];for(const i of s)n(i)}}return n(t),Array.from(e)}const K=$o;function $o(t){return t[mn]||t}function gn(){const t=D(),e=Se(t);return e.getActiveSpan?e.getActiveSpan():xe(T())}function yn(){zr||(fe(()=>{console.warn("[Sentry] Returning null from `beforeSendSpan` is disallowed. To drop certain spans, configure the respective integrations directly or use `ignoreSpans`.")}),zr=!0)}function je(t){var n;if(typeof __SENTRY_TRACING__=="boolean"&&!__SENTRY_TRACING__)return!1;const e=t||((n=P())==null?void 0:n.getOptions());return!!e&&(e.tracesSampleRate!=null||!!e.tracesSampler)}function Kr(t){b.log(`Ignoring span ${t.op} - ${t.description} because it matches \`ignoreSpans\`.`)}function ct(t,e){if(!(e!=null&&e.length))return!1;for(const n of e){if(Uo(n)){if(t.description&&Oe(t.description,n))return y&&Kr(t),!0;continue}const r=!!n.attributes&&Object.keys(n.attributes).length>0;if(!n.name&&!n.op&&!r)continue;const s=n.name?t.description&&Oe(t.description,n.name):!0,i=n.op?t.op&&Oe(t.op,n.op):!0,o=n.attributes?Object.entries(n.attributes).every(([c,a])=>{var u;return jo((u=t.attributes)==null?void 0:u[c],a)}):!0;if(s&&i&&o)return y&&Kr(t),!0}return!1}function jo(t,e){return typeof t=="string"&&(typeof e=="string"||e instanceof RegExp)?Oe(t,e):Array.isArray(t)&&Array.isArray(e)?t.length===e.length&&t.every((n,r)=>n===e[r]):t===e}function Bo(t,e){const n=e.parent_span_id,r=e.span_id;if(n)for(const s of t)s.parent_span_id===r&&(s.parent_span_id=n)}function Uo(t){return typeof t=="string"||t instanceof RegExp}const Vr=Symbol.for("sentry.nonRecordingSpan");class ut{constructor(e={}){this._traceId=e.traceId||me(),this._spanId=e.spanId||ge(),this.dropReason=e.dropReason,Q(this,Vr,!0)}spanContext(){return{spanId:this._spanId,traceId:this._traceId,traceFlags:Hr}}end(e){}setAttribute(e,n){return this}setAttributes(e){return this}setStatus(e){return this}updateName(e){return this}isRecording(){return!1}addEvent(e,n,r){return this}addLink(e){return this}addLinks(e){return this}recordException(e,n){}}function Be(t){return!!t&&t[Vr]===!0}const Sn="production",Xr="_frozenDsc";function Qr(t,e){Q(t,Xr,e)}function Zr(t,e){const n=e.getOptions(),{publicKey:r}=e.getDsn()||{},s={environment:n.environment||Sn,release:n.release,public_key:r,trace_id:t,org_id:wo(e)};return e.emit("createDsc",s),s}function lt(t,e){const n=e.getPropagationContext();return n.dsc||Zr(n.traceId,t)}function ve(t){var S,v,N,x,$;const e=P();if(!e)return{};const n=K(t),r=B(n),s=r.data,i=n.spanContext().traceState,o=(v=(S=i==null?void 0:i.get("sentry.sample_rate"))!=null?S:s[Cr])!=null?v:s[io];function c(k){return(typeof o=="number"||typeof o=="string")&&(k.sample_rate=`${o}`),k}const a=n[Xr];if(a)return c(a);const u=Be(n),p=u&&n.dropReason==="ignored";if(u&&(!je(e.getOptions())||p)){const k=be(n).scope;if(k){const J=l({},lt(e,k));return p&&(J.sampled="false"),c(J)}}const f=i==null?void 0:i.get("sentry.dsc"),_=f&&_o(f);if(_)return c(_);const d=Zr(t.spanContext().traceId,e),h=(N=s[Fe])!=null?N:s["sentry.segment.name.source"],m=r.description;return h!=="url"&&m&&(d.transaction=m),je()&&(d.sampled=String(oe(n)),d.sample_rand=($=i==null?void 0:i.get("sentry.sample_rand"))!=null?$:(x=be(n).scope)==null?void 0:x.getPropagationContext().sampleRand.toString()),c(d),e.emit("createDsc",d,n),d}function es(t){return!!t&&typeof t=="function"&&"_streamed"in t&&!!t._streamed}function ae(t,e=[]){return[t,e]}function ts(t,e){const[n,r]=t;return[n,[...r,e]]}function bn(t,e){const n=t[1];for(const r of n){const s=r[0].type;if(e(r,s))return!0}return!1}function En(t,e){return bn(t,(n,r)=>e.includes(r))}function Tn(t){const e=X(C);return e.encodePolyfill?e.encodePolyfill(t):new TextEncoder().encode(t)}function Ho(t){const[e,n]=t;let r=JSON.stringify(e);function s(i){typeof r=="string"?r=typeof i=="string"?r+i:[Tn(r),i]:r.push(typeof i=="string"?Tn(i):i)}for(const i of n){const[o,c]=i;if(s(`
 ${JSON.stringify(o)}
 `),typeof c=="string"||c instanceof Uint8Array)s(c);else{let a;try{a=JSON.stringify(c)}catch(u){a=JSON.stringify(te(c))}s(a)}}return typeof r=="string"?r:zo(r)}function zo(t){const e=t.reduce((s,i)=>s+i.length,0),n=new Uint8Array(e);let r=0;for(const s of t)n.set(s,r),r+=s.length;return n}function qo(t){return[{type:"span"},t]}function Wo(t){const e=typeof t.data=="string"?Tn(t.data):t.data;return[{type:"attachment",length:e.length,filename:t.filename,content_type:t.contentType,attachment_type:t.attachmentType},e]}const ns={sessions:"session",event:"error",client_report:"internal",user_report:"default",profile_chunk:"profile",replay_event:"replay",replay_recording:"replay",check_in:"monitor",raw_security:"security",log:"log_item",trace_metric:"metric"};function Go(t){return t in ns}function rs(t){return Go(t)?ns[t]:t}function ss(t){if(!(t!=null&&t.sdk))return;const{name:e,version:n}=t.sdk;return{name:e,version:n}}function Jo(t,e,n,r){var i;const s=(i=t.sdkProcessingMetadata)==null?void 0:i.dynamicSamplingContext;return l(l(l({event_id:t.event_id,sent_at:new Date(ee()).toISOString()},e&&{sdk:e}),!!n&&r&&{dsn:Ee(r)}),s&&{trace:s})}function Yo(t,e){var r,s,i,o;if(!e)return t;const n=t.sdk||{};return t.sdk=I(l({},n),{name:n.name||e.name,version:n.version||e.version,integrations:[...((r=t.sdk)==null?void 0:r.integrations)||[],...e.integrations||[]],packages:[...((s=t.sdk)==null?void 0:s.packages)||[],...e.packages||[]],settings:(i=t.sdk)!=null&&i.settings||e.settings?l(l({},(o=t.sdk)==null?void 0:o.settings),e.settings):void 0}),t}function Ko(t,e,n,r){const s=ss(n),i=l(l({sent_at:new Date(ee()).toISOString()},s&&{sdk:s}),!!r&&e&&{dsn:Ee(e)}),o="aggregates"in t?[{type:"sessions"},t]:[{type:"session"},t.toJSON()];return ae(i,[o])}function Vo(t,e,n,r){const s=ss(n),i=t.type&&t.type!=="replay_event"?t.type:"event";Yo(t,n==null?void 0:n.sdk);const o=Jo(t,s,r,e);return delete t.sdkProcessingMetadata,ae(o,[[{type:i},t]])}function Xo(t,e){function n(d){return!!d.trace_id&&!!d.public_key}const r=ve(t[0]),s=e==null?void 0:e.getDsn(),i=e==null?void 0:e.getOptions().tunnel,o=l(l({sent_at:new Date(ee()).toISOString()},n(r)&&{trace:r}),!!i&&s&&{dsn:Ee(s)}),{beforeSendSpan:c,ignoreSpans:a}=(e==null?void 0:e.getOptions())||{},u=a!=null&&a.length?t.filter(d=>{const h=B(d);return!ct({description:h.description,op:h.op,attributes:h.data},a)}):t,p=t.length-u.length;p&&(e==null||e.recordDroppedEvent("before_send","span",p));const f=c?d=>{const h=B(d),m=es(c)?h:c(h);return m||(yn(),h)}:B,_=[];for(const d of u){const h=f(d);h&&_.push(qo(h))}return ae(o,_)}function Qo(t){if(!y)return;const{description:e="< unknown name >",op:n="< unknown op >",parent_span_id:r}=B(t),{spanId:s}=t.spanContext(),i=oe(t),o=K(t),c=o===t,a=`[Tracing] Starting ${i?"sampled":"unsampled"} ${c?"root ":""}span`,u=[`op: ${n}`,`name: ${e}`,`ID: ${s}`];if(r&&u.push(`parent ID: ${r}`),!c){const{op:p,description:f}=B(o);u.push(`root ID: ${o.spanContext().spanId}`),p&&u.push(`root op: ${p}`),f&&u.push(`root description: ${f}`)}b.log(`${a}
   ${u.join(`
-  `)}`)}function Zo(t){if(!y)return;const{description:e="< unknown name >",op:n="< unknown op >"}=B(t),{spanId:r}=t.spanContext(),i=K(t)===t,o=`[Tracing] Finishing "${n}" ${i?"root ":""}span "${e}" with ID ${r}`;b.log(o)}function Ie(t,e,n,r=gn()){const s=r&&K(r);s&&(y&&b.log(`[Measurement] Setting measurement on root span: ${t} = ${e} ${n}`),s.addEvent(t,{[Nr]:e,[Pr]:n}))}function is(t){if(!t||t.length===0)return;const e={};return t.forEach(n=>{const r=n.attributes||{},s=r[Pr],i=r[Nr];typeof s=="string"&&typeof i=="number"&&(e[n.name]={value:i,unit:s})}),e}function os(){return X(O()).segmentSpanCaptureStrategy}function Ue(t){return t.getOptions().traceLifecycle==="stream"}const as=1e3;class vn{constructor(e={}){this._traceId=e.traceId||me(),this._spanId=e.spanId||ge(),this._startTime=e.startTimestamp||_e(),this._links=e.links,this._attributes={},this.setAttributes(l({[fn]:"manual",[st]:e.op},e.attributes)),this._name=e.name,e.parentSpanId&&(this._parentSpanId=e.parentSpanId),"sampled"in e&&(this._sampled=e.sampled),e.endTimestamp&&(this._endTime=e.endTimestamp),this._events=[],this._isStandaloneSpan=e.isStandalone,this._endTime&&this._onSpanEnded()}addLink(e){return this._frozen?this:(this._links?this._links.push(e):this._links=[e],this)}addLinks(e){return this._frozen?this:(this._links?this._links.push(...e):this._links=e,this)}recordException(e,n){}spanContext(){const{_spanId:e,_traceId:n,_sampled:r}=this;return{spanId:e,traceId:n,traceFlags:r?ot:Hr}}setAttribute(e,n){return this._frozen?this:(n===void 0?delete this._attributes[e]:this._attributes[e]=n,e===Fe&&n!==void 0&&Fr(this)&&po(this),this)}setAttributes(e){return Object.keys(e).forEach(n=>this.setAttribute(n,e[n])),this}updateStartTime(e){this._frozen||(this._startTime=Te(e))}setStatus(e){return this._frozen?this:(this._status=e,this)}updateName(e){return this._frozen?this:(this._name=e,Fr(this)||this.setAttribute(Fe,"custom"),this)}end(e){if(this._endTime){this._frozen=Lr(this);return}this._endTime=Te(e),Zo(this),this._onSpanEnded(),this._frozen=Lr(this)}getSpanJSON(){return{data:this._attributes,description:this._name,op:this._attributes[st],parent_span_id:this._parentSpanId,span_id:this._spanId,start_timestamp:this._startTime,status:Jr(this._status),timestamp:this._endTime,trace_id:this._traceId,origin:this._attributes[fn],profile_id:this._attributes[dn],exclusive_time:this._attributes[pn],measurements:is(this._events),is_segment:this._isStandaloneSpan&&K(this)===this||void 0,segment_id:this._isStandaloneSpan?K(this).spanContext().spanId:void 0,links:Wr(this._links)}}getStreamedSpanJSON(){var e,n;return{name:(e=this._name)!=null?e:"",span_id:this._spanId,trace_id:this._traceId,parent_span_id:this._parentSpanId,start_timestamp:this._startTime,end_timestamp:(n=this._endTime)!=null?n:this._startTime,is_segment:this._isStandaloneSpan||this===K(this),status:Yr(this._status),attributes:Fo(this._attributes,this._status),links:No(this._links)}}isRecording(){return!this._endTime&&!!this._sampled}addEvent(e,n,r){if(this._frozen)return this;y&&b.log("[Tracing] Adding an event to span:",e);const s=cs(n)?n:r||_e(),i=cs(n)?{}:n||{},o={name:e,time:Te(s),attributes:i};return this._events.push(o),this}isStandaloneSpan(){return!!this._isStandaloneSpan}_onSpanEnded(){const e=P();e&&(e.emit("spanEnd",this),this._isStandaloneSpan||e.emit("afterSpanEnd",this));const n=K(this),r=this._isStandaloneSpan||this===n;if(this._isStandaloneSpan){this._sampled?ta(Xo([this],e)):(y&&b.log("[Tracing] Discarding standalone span because its trace was not chosen to be sampled."),e&&e.recordDroppedEvent("sample_rate","span"));return}if(!r){const o=os();if(o){const c=be(this).scope||T();o.onChildSpanEnded(this,n,a=>this._convertSpanToTransaction(a),c)}return}if(e&&Ue(e)){e.emit("afterSegmentSpanEnd",this);return}const s=be(this).scope||T(),i=os();if(i)i.onSegmentSpanEnded(o=>this._convertSpanToTransaction(o),s);else{const o=this._convertSpanToTransaction();o&&s.captureEvent(o)}}_convertSpanToTransaction(e={}){var f,_,d,h;if(!us(B(this)))return;this._name||(y&&b.warn("Transaction has no name, falling back to `<unlabeled transaction>`."),this._name="<unlabeled transaction>");const{scope:n,isolationScope:r}=be(this),s=(f=n==null?void 0:n.getScopeData().sdkProcessingMetadata)==null?void 0:f.normalizedRequest;if(this._sampled!==!0)return;(_=e.onSpanCaptured)==null||_.call(e,this);const i=[];for(const m of Lo(this)){if(m===this||ea(m)||(d=e.isSpanAlreadyCaptured)!=null&&d.call(e,m))continue;const S=B(m);us(S)&&((h=e.onSpanCaptured)==null||h.call(e,m),i.push(S))}const o=this._attributes[Fe];delete this._attributes[Dr];let c=!1;i.forEach(m=>{var S;delete m.data[Dr],(S=m.op)!=null&&S.startsWith("gen_ai.")&&(c=!0)});const a=l({contexts:{trace:Ro(this)},spans:i.length>as?i.sort((m,S)=>m.start_timestamp-S.start_timestamp).slice(0,as):i,start_timestamp:this._startTime,timestamp:this._endTime,transaction:this._name,type:"transaction",sdkProcessingMetadata:{capturedSpanScope:n,capturedSpanIsolationScope:r,dynamicSamplingContext:ve(this),hasGenAiSpans:c},request:s},o&&{transaction_info:{source:o}}),u=is(this._events);return u&&Object.keys(u).length&&(y&&b.log("[Measurements] Adding measurements to transaction event",JSON.stringify(u,void 0,2)),a.measurements=u),a}}function cs(t){return t&&typeof t=="number"||t instanceof Date||Array.isArray(t)}function us(t){return!!t.start_timestamp&&!!t.timestamp&&!!t.span_id&&!!t.trace_id}function ea(t){return t instanceof vn&&t.isStandaloneSpan()}function ta(t){const e=P();if(!e)return;const n=t[1];if(!n||n.length===0){e.recordDroppedEvent("before_send","span");return}e.sendEnvelope(t)}function na(t,e,n=()=>{},r=()=>{}){let s;try{s=t()}catch(i){throw e(i),n(),i}return ra(s,e,n,r)}function ra(t,e,n,r){return de(t)?kr(t,s=>{n(),r(s)},s=>{e(s),n()}):(n(),r(t),t)}function sa(t,e,n){if(!je(t))return[!1];let r,s;typeof t.tracesSampler=="function"?(s=t.tracesSampler(I(l({},e),{inheritOrSampleWith:c=>typeof e.parentSampleRate=="number"?e.parentSampleRate:typeof e.parentSampled=="boolean"?Number(e.parentSampled):c})),r=!0):e.parentSampled!==void 0?s=e.parentSampled:typeof t.tracesSampleRate!="undefined"&&(s=t.tracesSampleRate,r=!0);const i=_n(s);if(i===void 0)return y&&b.warn(`[Tracing] Discarding root span because of invalid sample rate. Sample rate must be a boolean or a number between 0 and 1. Got ${JSON.stringify(s)} of type ${JSON.stringify(typeof s)}.`),[!1];if(!i)return y&&b.log(`[Tracing] Discarding transaction because ${typeof t.tracesSampler=="function"?"tracesSampler returned 0 or false":"a negative sampling decision was inherited or tracesSampleRate is set to 0"}`),[!1,i,r];const o=n<i;return o||y&&b.log(`[Tracing] Discarding transaction because it's not included in the random sample (sampling rate = ${Number(s)})`),[o,i,r]}const ia="__SENTRY_SUPPRESS_TRACING__";function He(t,e){const n=dt();if(n.startSpan)return n.startSpan(t,e);const r=ps(t),{forceTransaction:s,parentSpan:i,scope:o}=t,c=o==null?void 0:o.clone();return ie(c,()=>ca(i)(()=>{const u=T(),p=_s(u,i),f=P(),d=t.onlyIfParent&&!p?fs(u,f):ds({parentSpan:p,spanArguments:r,forceTransaction:s,scope:u});return(!la(d)||!p)&&nt(u,d),na(()=>e(d),()=>{const{status:h}=B(d);d.isRecording()&&(!h||h==="ok")&&d.setStatus({code:j,message:"internal_error"})},()=>{d.end()})}))}function ft(t){const e=dt();return e.startInactiveSpan?e.startInactiveSpan(t):oa(t)}function oa(t){const e=ps(t),{forceTransaction:n,parentSpan:r}=t;return(t.scope?i=>ie(t.scope,i):r!==void 0?i=>ls(r,i):i=>i())(()=>{const i=T(),o=_s(i,r),c=P();return t.onlyIfParent&&!o?fs(i,c):ds({parentSpan:o,spanArguments:e,forceTransaction:n,scope:i})})}function ls(t,e){const n=dt();return n.withActiveSpan?n.withActiveSpan(t,e):ie(r=>(nt(r,t||void 0),e(r)))}function In(t=T()){const e=dt();return e.isTracingSuppressed?e.isTracingSuppressed(t):t.getScopeData().sdkProcessingMetadata[ia]===!0}function fs(t,e){e==null||e.recordDroppedEvent("no_parent_span","span");const n=new ut({traceId:t.getPropagationContext().traceId});return Le(n,t,F()),n}function ds({parentSpan:t,spanArguments:e,forceTransaction:n,scope:r}){var c;const s=F();if(!je()){const a=l(l({},s.getPropagationContext()),r.getPropagationContext()),u=t?t.spanContext().traceId:a.traceId,p=new ut({traceId:u});return t&&!n&&at(t,p),Le(p,r,s),p}const i=P();if(ua(i,e)){In(r)||i==null||i.recordDroppedEvent("ignored","span");const a=new ut({dropReason:"ignored",traceId:(c=t==null?void 0:t.spanContext().traceId)!=null?c:r.getPropagationContext().traceId});return t&&!n&&at(t,a),Le(a,r,s),a}let o;if(t&&!n)o=aa(t,r,e,s),at(t,o);else if(t){const a=ve(t),{traceId:u,spanId:p}=t.spanContext(),f=oe(t);o=hs(l({traceId:u,parentSpanId:p},e),r,s,f),Qr(o,a)}else{const{traceId:a,dsc:u,parentSpanId:p,sampled:f}=l(l({},s.getPropagationContext()),r.getPropagationContext());o=hs(l({traceId:a,parentSpanId:p},e),r,s,f),u&&Qr(o,u)}return Qo(o),o}function ps(t){const e=t.experimental||{},n=l({isStandalone:e.standalone},t);if(t.startTime){const r=l({},n);return r.startTimestamp=Te(t.startTime),delete r.startTime,r}return n}function dt(){const t=O();return Se(t)}function hs(t,e,n,r){var S,v;const s=P(),i=(s==null?void 0:s.getOptions())||{},{name:o=""}=t,c={spanAttributes:l({},t.attributes),spanName:o,parentSampled:r};s==null||s.emit("beforeSampling",c,{decision:!1});const a=(S=c.parentSampled)!=null?S:r,u=c.spanAttributes,p=e.getPropagationContext(),f=In(e),[_,d,h]=f?[!1]:sa(i,{name:o,parentSampled:a,attributes:u,normalizedRequest:n.getScopeData().sdkProcessingMetadata.normalizedRequest,parentSampleRate:_n((v=p.dsc)==null?void 0:v.sample_rate)},p.sampleRand),m=new vn(I(l({},t),{attributes:l({[Fe]:"custom",[Cr]:d!==void 0&&h?d:void 0},u),sampled:_}));return!_&&s&&!f&&(y&&b.log("[Tracing] Discarding root span because its trace was not chosen to be sampled."),s.recordDroppedEvent("sample_rate",Ue(s)?"span":"transaction")),Le(m,e,n),s&&s.emit("spanStart",m),m}function aa(t,e,n,r){const{spanId:s,traceId:i}=t.spanContext(),o=In(e),c=o?!1:oe(t),a=c?new vn(I(l({},n),{parentSpanId:s,traceId:i,sampled:c})):new ut({traceId:i});at(t,a),Le(a,e,r);const u=P();return u&&(Ue(u)&&Be(a)&&(Be(t)&&t.dropReason?(a.dropReason=t.dropReason,u.recordDroppedEvent(t.dropReason,"span")):o||(a.dropReason="sample_rate",u.recordDroppedEvent("sample_rate","span"))),u.emit("spanStart",a),n.endTimestamp&&(u.emit("spanEnd",a),u.emit("afterSpanEnd",a))),a}function _s(t,e){if(e)return e;if(e===null)return;const n=xe(t);if(!n)return;const r=P();return(r?r.getOptions():{}).parentSpanIsAlwaysRootSpan?K(n):n}function ca(t){return t!==void 0?e=>ls(t,e):e=>e()}function ua(t,e){var r;const n=t==null?void 0:t.getOptions().ignoreSpans;return!t||!Ue(t)||!(n!=null&&n.length)?!1:ct({description:e.name||"",op:((r=e.attributes)==null?void 0:r[st])||e.op,attributes:e.attributes},n)}function la(t){return Be(t)&&t.dropReason==="ignored"}function fa(t,e){const{fingerprint:n,span:r,breadcrumbs:s,sdkProcessingMetadata:i}=e;da(t,e),r&&_a(t,r),ma(t,n),pa(t,s),ha(t,i)}function ms(t,e){const{extra:n,tags:r,attributes:s,user:i,contexts:o,level:c,sdkProcessingMetadata:a,breadcrumbs:u,fingerprint:p,eventProcessors:f,attachments:_,propagationContext:d,transactionName:h,span:m}=e;ze(t,"extra",n),ze(t,"tags",r),ze(t,"attributes",s),ze(t,"user",i),ze(t,"contexts",o),t.sdkProcessingMetadata=Me(t.sdkProcessingMetadata,a,2),c&&(t.level=c),h&&(t.transactionName=h),m&&(t.span=m),u.length&&(t.breadcrumbs=[...t.breadcrumbs,...u]),p.length&&(t.fingerprint=[...t.fingerprint,...p]),f.length&&(t.eventProcessors=[...t.eventProcessors,...f]),_.length&&(t.attachments=[...t.attachments,..._]),t.propagationContext=l(l({},t.propagationContext),d)}function ze(t,e,n){t[e]=Me(t[e],n,1)}function wn(t,e){const n=so().getScopeData();return t&&ms(n,t.getScopeData()),e&&ms(n,e.getScopeData()),n}function da(t,e){const{extra:n,tags:r,user:s,contexts:i,level:o,transactionName:c}=e;Object.keys(n).length&&(t.extra=l(l({},n),t.extra)),Object.keys(r).length&&(t.tags=l(l({},r),t.tags)),Object.keys(s).length&&(t.user=l(l({},s),t.user)),Object.keys(i).length&&(t.contexts=l(l({},i),t.contexts)),o&&(t.level=o),c&&t.type!=="transaction"&&(t.transaction=c)}function pa(t,e){const n=[...t.breadcrumbs||[],...e];t.breadcrumbs=n.length?n:void 0}function ha(t,e){t.sdkProcessingMetadata=l(l({},t.sdkProcessingMetadata),e)}function _a(t,e){t.contexts=l({trace:qr(e)},t.contexts),t.sdkProcessingMetadata=l({dynamicSamplingContext:ve(e)},t.sdkProcessingMetadata);const n=K(e),r=B(n).description;r&&!t.transaction&&t.type==="transaction"&&(t.transaction=r)}function ma(t,e){t.fingerprint=t.fingerprint?Array.isArray(t.fingerprint)?t.fingerprint:[t.fingerprint]:[],e&&(t.fingerprint=t.fingerprint.concat(e)),t.fingerprint.length||delete t.fingerprint}const kn=0,gs=1,ys=2;function An(t){return new qe(e=>{e(t)})}function Rn(t){return new qe((e,n)=>{n(t)})}class qe{constructor(e){this._state=kn,this._handlers=[],this._runExecutor(e)}then(e,n){return new qe((r,s)=>{this._handlers.push([!1,i=>{if(!e)r(i);else try{r(e(i))}catch(o){s(o)}},i=>{if(!n)s(i);else try{r(n(i))}catch(o){s(o)}}]),this._executeHandlers()})}catch(e){return this.then(n=>n,e)}finally(e){return new qe((n,r)=>{let s,i;return this.then(o=>{i=!1,s=o,e&&e()},o=>{i=!0,s=o,e&&e()}).then(()=>{if(i){r(s);return}n(s)})})}_executeHandlers(){if(this._state===kn)return;const e=this._handlers.slice();this._handlers=[],e.forEach(n=>{n[0]||(this._state===gs&&n[1](this._value),this._state===ys&&n[2](this._value),n[0]=!0)})}_runExecutor(e){const n=(i,o)=>{if(this._state===kn){if(de(o)){o.then(r,s);return}this._state=i,this._value=o,this._executeHandlers()}},r=i=>{n(gs,i)},s=i=>{n(ys,i)};try{e(r,s)}catch(i){s(i)}}}function ga(t,e,n,r=0){try{const s=Cn(e,n,t,r);return de(s)?s:An(s)}catch(s){return Rn(s)}}function Cn(t,e,n,r){const s=n[r];if(!t||!s)return t;const i=s(l({},t),e);return y&&i===null&&b.log(`Event processor "${s.id||"?"}" dropped event`),de(i)?i.then(o=>Cn(o,e,n,r+1)):Cn(i,e,n,r+1)}let Z,Ss,bs,ne;function ya(t){const e=C._sentryDebugIds,n=C._debugIds;if(!e&&!n)return{};const r=e?Object.keys(e):[],s=n?Object.keys(n):[];if(ne&&r.length===Ss&&s.length===bs)return ne;Ss=r.length,bs=s.length,ne={},Z||(Z={});const i=(o,c)=>{for(const a of o){const u=c[a],p=Z==null?void 0:Z[a];if(p&&ne&&u)ne[p[0]]=u,Z&&(Z[a]=[p[0],u]);else if(u){const f=t(a);for(let _=f.length-1;_>=0;_--){const d=f[_],h=d==null?void 0:d.filename;if(h&&ne&&Z){ne[h]=u,Z[a]=[h,u];break}}}}};return e&&i(r,e),n&&i(s,n),ne}function Sa(t,e,n,r,s,i){const{normalizeDepth:o=3,normalizeMaxBreadth:c=1e3}=t,a=I(l({},e),{event_id:e.event_id||n.event_id||W(),timestamp:e.timestamp||Oe()}),u=n.integrations||t.integrations.map(v=>v.name);ba(a,t),va(a,u),s&&s.emit("applyFrameMetadata",e),e.type===void 0&&Ea(a,t.stackParser);const p=wa(r,n.captureContext);n.mechanism&&zi(a,n.mechanism);const f=s?s.getEventProcessors():[],_=wn(i,p),d=[...n.attachments||[],..._.attachments];d.length&&(n.attachments=d),fa(a,_);const h=[...f,..._.eventProcessors];return(n.data&&n.data.__sentry__===!0?An(a):ga(h,a,n)).then(v=>(v&&Ta(v),typeof o=="number"&&o>0?Ia(v,o,c):v))}function ba(t,e){var c,a;const{environment:n,release:r,dist:s,maxValueLength:i}=e;t.environment=t.environment||n||Sn,!t.release&&r&&(t.release=r),!t.dist&&s&&(t.dist=s);const o=t.request;o!=null&&o.url&&i&&(o.url=cn(o.url,i)),i&&((a=(c=t.exception)==null?void 0:c.values)==null||a.forEach(u=>{u.value&&(u.value=cn(u.value,i))}))}function Ea(t,e){var r,s;const n=ya(e);(s=(r=t.exception)==null?void 0:r.values)==null||s.forEach(i=>{var o,c;(c=(o=i.stacktrace)==null?void 0:o.frames)==null||c.forEach(a=>{a.filename&&(a.debug_id=n[a.filename])})})}function Ta(t){var r,s;const e={};if((s=(r=t.exception)==null?void 0:r.values)==null||s.forEach(i=>{var o,c;(c=(o=i.stacktrace)==null?void 0:o.frames)==null||c.forEach(a=>{a.debug_id&&(a.abs_path?e[a.abs_path]=a.debug_id:a.filename&&(e[a.filename]=a.debug_id),delete a.debug_id)})}),Object.keys(e).length===0)return;t.debug_meta=t.debug_meta||{},t.debug_meta.images=t.debug_meta.images||[];const n=t.debug_meta.images;Object.entries(e).forEach(([i,o])=>{n.push({type:"sourcemap",code_file:i,debug_id:o})})}function va(t,e){e.length>0&&(t.sdk=t.sdk||{},t.sdk.integrations=[...t.sdk.integrations||[],...e])}function Ia(t,e,n){var s,i;if(!t)return null;const r=l(l(l(l(l({},t),t.breadcrumbs&&{breadcrumbs:t.breadcrumbs.map(o=>l(l({},o),o.data&&{data:te(o.data,e,n)}))}),t.user&&{user:te(t.user,e,n)}),t.contexts&&{contexts:te(t.contexts,e,n)}),t.extra&&{extra:te(t.extra,e,n)});return(s=t.contexts)!=null&&s.trace&&r.contexts&&(r.contexts.trace=t.contexts.trace,t.contexts.trace.data&&(r.contexts.trace.data=te(t.contexts.trace.data,e,n))),t.spans&&(r.spans=t.spans.map(o=>l(l({},o),o.data&&{data:te(o.data,e,n)}))),(i=t.contexts)!=null&&i.flags&&r.contexts&&(r.contexts.flags=te(t.contexts.flags,3,n)),r}function wa(t,e){if(!e)return t;const n=t?t.clone():new Y;return n.update(e),n}function ka(t){if(t)return Aa(t)?{captureContext:t}:Ca(t)?{captureContext:t}:t}function Aa(t){return t instanceof Y||typeof t=="function"}const Ra=["user","level","extra","contexts","tags","fingerprint","propagationContext"];function Ca(t){return Object.keys(t).some(e=>Ra.includes(e))}function we(t,e){return T().captureException(t,ka(e))}function Pa(t,e){const n=typeof e=="string"?e:void 0,r=typeof e!="string"?{captureContext:e}:void 0;return T().captureMessage(t,n,r)}function Na(t,e){return T().captureEvent(t,e)}function re(t,e){F().setContext(t,e)}function Da(t){F().setExtras(t)}function Oa(t,e){F().setExtra(t,e)}function Ma(t){F().setTags(t)}function xa(t,e){F().setTag(t,e)}function Fa(t){F().setUser(t)}function La(){return F().lastEventId()}function Es(t){return z(this,null,function*(){const e=P();return e?e.flush(t):(y&&b.warn("Cannot flush events. No client defined."),Promise.resolve(!1))})}function $a(t){return z(this,null,function*(){const e=P();return e?e.close(t):(y&&b.warn("Cannot flush events and disable SDK. No client defined."),Promise.resolve(!1))})}function Ts(){const t=P();return(t==null?void 0:t.getOptions().enabled)!==!1&&!!(t!=null&&t.getTransport())}function ja(t){F().addEventProcessor(t)}function vs(t){const e=F(),{user:n}=wn(e,T()),{userAgent:r}=C.navigator||{},s=br(l(l({user:n},r&&{userAgent:r}),t)),i=e.getSession();return(i==null?void 0:i.status)==="ok"&&se(i,{status:"exited"}),pt(),e.setSession(s),s}function pt(){const t=F(),n=T().getSession()||t.getSession();n&&Er(n),Is(),t.setSession()}function Is(){const t=F(),e=P(),n=t.getSession();n&&e&&e.captureSession(n)}function Pn(t=!1){if(t){pt();return}Is()}function Nn(t){return typeof t=="object"&&typeof t.unref=="function"&&t.unref(),t}const Ba="7";function Ua(t){const e=t.protocol?`${t.protocol}:`:"",n=t.port?`:${t.port}`:"";return`${e}//${t.host}${n}${t.path?`/${t.path}`:""}/api/`}function Ha(t){return`${Ua(t)}${t.projectId}/envelope/`}function za(t,e){const n={sentry_version:Ba};return t.publicKey&&(n.sentry_key=t.publicKey),e&&(n.sentry_client=`${e.name}/${e.version}`),new URLSearchParams(n).toString()}function qa(t,e,n){return e||`${Ha(t)}?${za(t,n)}`}const We=[];function Wa(t,e){const n={};return e.forEach(r=>{r!=null&&r.beforeSetup&&r.beforeSetup(t)}),e.forEach(r=>{r&&ks(t,r,n)}),n}function ws(t,e){for(const n of e)n!=null&&n.afterAllSetup&&n.afterAllSetup(t)}function ks(t,e,n){if(n[e.name]){y&&b.log(`Integration skipped because it was already installed: ${e.name}`);return}if(n[e.name]=e,!We.includes(e.name)&&typeof e.setupOnce=="function"&&(e.setupOnce(),We.push(e.name)),e.setup&&typeof e.setup=="function"&&e.setup(t),typeof e.preprocessEvent=="function"){const r=e.preprocessEvent.bind(e);t.on("preprocessEvent",(s,i)=>r(s,i,t))}if(typeof e.processEvent=="function"){const r=e.processEvent.bind(e),s=Object.assign((i,o)=>r(i,o,t),{id:e.name});t.addEventProcessor(s)}["processSpan","processSegmentSpan"].forEach(r=>{const s=e[r];typeof s=="function"&&t.on(r,i=>s.call(e,i,t))}),y&&b.log(`Integration installed: ${e.name}`)}function Ga(t){const e=P();if(!e){y&&b.warn(`Cannot add integration "${t.name}" because no SDK Client is available.`);return}e.addIntegration(t)}function Tl(t){return t}const Ja="sentry.timestamp.sequence";let Dn=0,On;function Ya(t){const e=Math.floor(t*1e3);On!==void 0&&e!==On&&(Dn=0);const n=Dn;return Dn++,On=e,{key:Ja,value:{value:n,type:"integer"}}}function Ka(t,e){return e?ie(e,()=>{const n=gn(),r=n?qr(n):Rr(e);return[n?ve(n):lt(t,e),r]}):[void 0,void 0]}const Va={trace:1,debug:5,info:9,warn:13,error:17,fatal:21};function Xa(){return typeof __SENTRY_BROWSER_BUNDLE__!="undefined"&&!!__SENTRY_BROWSER_BUNDLE__}function Qa(){return!Xa()&&Object.prototype.toString.call(typeof process!="undefined"?process:0)==="[object process]"}function Mn(){return typeof window!="undefined"&&(!Qa()||Za())}function Za(){const t=C.process;return(t==null?void 0:t.type)==="renderer"}function ec(t,e){const n=e?"auto":"never";return[{type:"log",item_count:t.length,content_type:"application/vnd.sentry.items.log+json"},I(l({version:2},Mn()&&{ingest_settings:{infer_ip:n,infer_user_agent:n}}),{items:t})]}function tc(t,e,n,r,s){const i={};return e!=null&&e.sdk&&(i.sdk={name:e.sdk.name,version:e.sdk.version}),n&&r&&(i.dsn=Ee(r)),ae(i,[ec(t,s)])}const nc=100;function V(t,e,n,r=!0){n&&(!t[e]||r)&&(t[e]=n)}function rc(t,e){const n=xn(),r=Rs(t);r===void 0?n.set(t,[e]):r.length>=nc?(As(t,r),n.set(t,[e])):n.set(t,[...r,e])}function sc(t,e=T(),n=rc){var Kt,Pe,_i;const r=(Kt=e==null?void 0:e.getClient())!=null?Kt:P();if(!r){y&&b.warn("No client available to capture log.");return}const{release:s,environment:i,enableLogs:o=!1,beforeSendLog:c}=r.getOptions();if(!o){y&&b.warn("logging option not enabled, log will not be captured.");return}const[,a]=Ka(r,e),u=l({},t.attributes),{user:{id:p,email:f,username:_},attributes:d={}}=wn(F(),e);V(u,"user.id",p,!1),V(u,"user.email",f,!1),V(u,"user.name",_,!1),V(u,"sentry.release",s),V(u,"sentry.environment",i);const{name:h,version:m}=(_i=(Pe=r.getSdkMetadata())==null?void 0:Pe.sdk)!=null?_i:{};V(u,"sentry.sdk.name",h),V(u,"sentry.sdk.version",m);const S=r.getIntegrationByName("Replay"),v=S==null?void 0:S.getReplayId(!0);V(u,"sentry.replay_id",v),v&&(S==null?void 0:S.getRecordingMode())==="buffer"&&V(u,"sentry._internal.replay_is_buffering",!0);const N=t.message;if(rn(N)){const{__sentry_template_string__:hl,__sentry_template_values__:Vt=[]}=N;Vt!=null&&Vt.length&&(u["sentry.message.template"]=hl),Vt.forEach((_l,ml)=>{u[`sentry.message.parameter.${ml}`]=_l})}const x=xe(e);V(u,"sentry.trace.parent_span_id",x==null?void 0:x.spanContext().spanId);const $=I(l({},t),{attributes:u});r.emit("beforeCaptureLog",$);const k=c?fe(()=>c($)):$;if(!k){r.recordDroppedEvent("before_send","log_item",1),y&&b.warn("beforeSendLog returned null, log will not be captured.");return}const{level:J,message:ce,attributes:Yt={},severityNumber:A}=k,L=_e(),H=Ya(L),ue={timestamp:L,level:J,body:Fn(String(ce)),trace_id:a==null?void 0:a.trace_id,severity_number:A!=null?A:Va[J],attributes:ic(I(l(l({},rt(d)),rt(Yt,!0)),{[H.key]:H.value}))};n(r,ue),r.emit("afterCaptureLog",k)}function As(t,e){var i;const n=(i=e!=null?e:Rs(t))!=null?i:[];if(n.length===0)return;const r=t.getOptions(),s=tc(n,r._metadata,r.tunnel,t.getDsn(),t.getDataCollectionOptions().userInfo);xn().set(t,[]),t.emit("flushLogs"),t.sendEnvelope(s)}function Rs(t){return xn().get(t)}function xn(){return le("clientToLogBufferMap",()=>new WeakMap)}function ic(t){const e={};for(const[n,r]of Object.entries(t)){const s=Fn(n);r.type==="string"?e[s]=I(l({},r),{value:Fn(r.value)}):e[s]=r}return e}function Fn(t){const e=Object(t),n=e.isWellFormed,r=e.toWellFormed;return typeof n=="function"&&typeof r=="function"?n.call(t)?t:r.call(t):t}function oc(t,e){const n=e?"auto":"never";return[{type:"trace_metric",item_count:t.length,content_type:"application/vnd.sentry.items.trace-metric+json"},I(l({version:2},Mn()&&{ingest_settings:{infer_ip:n,infer_user_agent:n}}),{items:t})]}function ac(t,e,n,r,s){const i={};return e!=null&&e.sdk&&(i.sdk={name:e.sdk.name,version:e.sdk.version}),n&&r&&(i.dsn=Ee(r)),ae(i,[oc(t,s)])}function cc(t,e){var i;const n=(i=e!=null?e:uc(t))!=null?i:[];if(n.length===0)return;const r=t.getOptions(),s=ac(n,r._metadata,r.tunnel,t.getDsn(),t.getDataCollectionOptions().userInfo);Cs().set(t,[]),t.emit("flushMetrics"),t.sendEnvelope(s)}function uc(t){return Cs().get(t)}function Cs(){return le("clientToMetricBufferMap",()=>new WeakMap)}function lc(t){const e={trace_id:t.trace_id,span_id:t.span_id,parent_span_id:t.parent_span_id,name:t.description||"",start_timestamp:t.start_timestamp,end_timestamp:t.timestamp||t.start_timestamp,status:!t.status||t.status==="ok"||t.status==="cancelled"?"ok":"error",is_segment:!1,attributes:l({},t.data),links:t.links};return Oo(e)}function fc(t,e){var i,o,c;if(t.type!=="transaction"||!((i=t.spans)!=null&&i.length)||!((o=t.sdkProcessingMetadata)!=null&&o.hasGenAiSpans)||e.getOptions().streamGenAiSpans===!1||Ue(e))return;const n=[],r=[];for(const a of t.spans)(c=a.op)!=null&&c.startsWith("gen_ai.")?n.push(lc(a)):r.push(a);if(n.length===0)return;t.spans=r;const s=e.getDataCollectionOptions().userInfo?"auto":"never";return[{type:"span",item_count:n.length,content_type:"application/vnd.sentry.items.span.v2+json"},I(l({version:2},Mn()&&{ingest_settings:{infer_ip:s,infer_user_agent:s}}),{items:n})]}const Ln=Symbol.for("SentryBufferFullError");function Ps(t=100){const e=new Set;function n(){return e.size<t}function r(o){e.delete(o)}function s(o){if(!n())return Rn(Ln);const c=o();return e.add(c),c.then(()=>r(c),()=>r(c)),c}function i(o){if(!e.size)return An(!0);const c=Promise.allSettled(Array.from(e)).then(()=>!0);if(!o)return c;const a=[c,new Promise(u=>Nn(setTimeout(()=>u(!1),o)))];return Promise.race(a)}return{get $(){return Array.from(e)},add:s,drain:i}}const dc=60*1e3;function Ns(t,e=ee()){const n=parseInt(`${t}`,10);if(!isNaN(n))return n*1e3;const r=Date.parse(`${t}`);return isNaN(r)?dc:r-e}function pc(t,e){return t[e]||t.all||0}function hc(t,e,n=ee()){return pc(t,e)>n}function _c(t,{statusCode:e,headers:n},r=ee()){const s=l({},t),i=n==null?void 0:n["x-sentry-rate-limits"],o=n==null?void 0:n["retry-after"];if(i)for(const c of i.trim().split(",")){const[a,u,,,p]=c.split(":",5),f=parseInt(a,10),_=(isNaN(f)?60:f)*1e3;if(!u)s.all=r+_;else for(const d of u.split(";"))d==="metric_bucket"?(!p||p.split(";").includes("custom"))&&(s[d]=r+_):s[d]=r+_}else o?s.all=r+Ns(o,r):e===429&&(s.all=r+60*1e3);return s}const Ds=64;function mc(t,e,n=Ps(t.bufferSize||Ds)){let r={};const s=o=>n.drain(o);function i(o){const c=[];if(bn(o,(f,_)=>{const d=rs(_);hc(r,d)?t.recordDroppedEvent("ratelimit_backoff",d):c.push(f)}),c.length===0)return Promise.resolve({});const a=ae(o[0],c),u=f=>{if(En(a,["client_report"])){y&&b.warn(`Dropping client report. Will not send outcomes (reason: ${f}).`);return}bn(a,(_,d)=>{t.recordDroppedEvent(f,rs(d))})},p=()=>e({body:Ho(a)}).then(f=>f.statusCode===413?(y&&b.error("Sentry responded with status code 413. Envelope was discarded due to exceeding size limits."),u("send_error"),f):(y&&f.statusCode!==void 0&&(f.statusCode<200||f.statusCode>=300)&&b.warn(`Sentry responded with status code ${f.statusCode} to sent event.`),r=_c(r,f),f),f=>{throw u("network_error"),y&&b.error("Encountered error running transport request:",f),f});return n.add(p).then(f=>f,f=>{if(f===Ln)return y&&b.error("Skipped sending event because buffer is full."),u("queue_overflow"),Promise.resolve({});throw f})}return{send:i,flush:s}}function gc(t,e,n){const r=[{type:"client_report"},{timestamp:Oe(),discarded_events:t}];return ae(e?{dsn:e}:{},[r])}function Os(t){const e=[];t.message&&e.push(t.message);try{const n=t.exception.values[t.exception.values.length-1];n!=null&&n.value&&(e.push(n.value),n.type&&e.push(`${n.type}: ${n.value}`))}catch(n){}return e}function yc(t){var a,u,p;const{trace_id:e,parent_span_id:n,span_id:r,status:s,origin:i,data:o,op:c}=(u=(a=t.contexts)==null?void 0:a.trace)!=null?u:{};return{data:o!=null?o:{},description:t.transaction,op:c,parent_span_id:n,span_id:r!=null?r:"",start_timestamp:(p=t.start_timestamp)!=null?p:0,status:s,timestamp:t.timestamp,trace_id:e!=null?e:"",origin:i,profile_id:o==null?void 0:o[dn],exclusive_time:o==null?void 0:o[pn],measurements:t.measurements,is_segment:!0}}function Sc(t){return{type:"transaction",timestamp:t.timestamp,start_timestamp:t.start_timestamp,transaction:t.description,contexts:{trace:{trace_id:t.trace_id,span_id:t.span_id,parent_span_id:t.parent_span_id,op:t.op,status:t.status,origin:t.origin,data:l(l(l({},t.data),t.profile_id&&{[dn]:t.profile_id}),t.exclusive_time&&{[pn]:t.exclusive_time})}},measurements:t.measurements}}const ht=["forwarded","-ip","remote-","via","-user"];function bc(t){return t===!0?{userInfo:!0,cookies:!0,httpHeaders:{request:!0,response:!0},httpBodies:["incomingRequest","outgoingRequest","incomingResponse","outgoingResponse"],urlQueryParams:!0,graphQL:{document:!0,variables:!0},genAI:{inputs:!0,outputs:!0},databaseQueryData:!0,stackFrameVariables:!0,frameContextLines:7}:{userInfo:!1,cookies:{deny:ht},httpHeaders:{request:{deny:ht},response:{deny:ht}},httpBodies:[],urlQueryParams:{deny:ht},graphQL:{document:!0,variables:!0},genAI:{inputs:!1,outputs:!1},databaseQueryData:!1,stackFrameVariables:!0,frameContextLines:7}}const Ec={userInfo:!0,cookies:!0,httpHeaders:{request:!0,response:!0},httpBodies:["incomingRequest","outgoingRequest","incomingResponse","outgoingResponse"],urlQueryParams:!0,graphQL:{document:!0,variables:!0},genAI:{inputs:!0,outputs:!0},databaseQueryData:!0,stackFrameVariables:!0,frameContextLines:5};function Tc(t){var r,s,i,o,c,a,u,p,f,_,d,h,m,S,v,N,x,$,k,J,ce;const e=t.dataCollection!=null?Ec:bc(t.sendDefaultPii),n=(r=t.dataCollection)!=null?r:{};return{userInfo:(s=n.userInfo)!=null?s:e.userInfo,cookies:(i=n.cookies)!=null?i:e.cookies,httpHeaders:{request:(c=(o=n.httpHeaders)==null?void 0:o.request)!=null?c:e.httpHeaders.request,response:(u=(a=n.httpHeaders)==null?void 0:a.response)!=null?u:e.httpHeaders.response},httpBodies:(p=n.httpBodies)!=null?p:e.httpBodies,urlQueryParams:(_=(f=n.urlQueryParams)!=null?f:n.queryParams)!=null?_:e.urlQueryParams,graphQL:{document:(h=(d=n.graphQL)==null?void 0:d.document)!=null?h:e.graphQL.document,variables:(S=(m=n.graphQL)==null?void 0:m.variables)!=null?S:e.graphQL.variables},genAI:{inputs:(N=(v=n.genAI)==null?void 0:v.inputs)!=null?N:e.genAI.inputs,outputs:($=(x=n.genAI)==null?void 0:x.outputs)!=null?$:e.genAI.outputs},databaseQueryData:(k=n.databaseQueryData)!=null?k:e.databaseQueryData,stackFrameVariables:(J=n.stackFrameVariables)!=null?J:e.stackFrameVariables,frameContextLines:(ce=n.frameContextLines)!=null?ce:e.frameContextLines}}const Ms="Not capturing exception because it's already been captured.",xs="Discarded session because of missing or non-string release",Fs=Symbol.for("SentryInternalError"),Ls=Symbol.for("SentryDoNotSendEventError"),vc=5e3;function _t(t){return{message:t,[Fs]:!0}}function $n(t){return{message:t,[Ls]:!0}}function $s(t){return on(t)&&Fs in t}function js(t){return on(t)&&Ls in t}function Bs(t,e,n,r,s){let i=0,o,c=!1;t.on(n,()=>{i=0,clearTimeout(o),c=!1}),t.on(e,a=>{var u;if(i+=r(a),i>=8e5)s(t);else if(!c){const p=(u=t.getOptions()._flushInterval)!=null?u:vc;p>0&&(c=!0,o=Nn(setTimeout(()=>{s(t)},p)))}}),t.on("flush",()=>{s(t)})}class Ic{constructor(e){var r,s,i,o,c,a,u;if(this._options=e,this._integrations={},this._numProcessing=0,this._outcomes={},this._hooks={},this._eventProcessors=[],this._promiseBuffer=Ps((s=(r=e.transportOptions)==null?void 0:r.bufferSize)!=null?s:Ds),this._dataCollection=Tc(e),e.dsn?this._dsn=ko(e.dsn):y&&b.warn("No DSN provided, client will not send events."),this._dsn){const p=qa(this._dsn,e.tunnel,e._metadata?e._metadata.sdk:void 0);this._transport=e.transport(I(l({tunnel:this._options.tunnel,recordDroppedEvent:this.recordDroppedEvent.bind(this)},e.transportOptions),{url:p}))}this._options.enableLogs=(o=this._options.enableLogs)!=null?o:(i=this._options._experiments)==null?void 0:i.enableLogs,this._options.enableLogs&&Bs(this,"afterCaptureLog","flushLogs",Rc,As),((u=(a=this._options.enableMetrics)!=null?a:(c=this._options._experiments)==null?void 0:c.enableMetrics)!=null?u:!0)&&Bs(this,"afterCaptureMetric","flushMetrics",Ac,cc)}captureException(e,n,r){const s=W();if(yr(e))return y&&b.log(Ms),s;const i=l({event_id:s},n);return this._process(()=>this.eventFromException(e,i).then(o=>this._captureEvent(o,i,r)).then(o=>o),"error"),i.event_id}captureMessage(e,n,r,s){const i=l({event_id:W()},r),o=rn(e)?e:String(e),c=hr(e),a=c?this.eventFromMessage(o,n,i):this.eventFromException(e,i);return this._process(()=>a.then(u=>this._captureEvent(u,i,s)),c?"unknown":"error"),i.event_id}captureEvent(e,n,r){const s=W();if(n!=null&&n.originalException&&yr(n.originalException))return y&&b.log(Ms),s;const i=l({event_id:s},n),o=e.sdkProcessingMetadata||{},c=o.capturedSpanScope,a=o.capturedSpanIsolationScope,u=Us(e.type);return this._process(()=>this._captureEvent(e,i,c||r,a),u),i.event_id}captureSession(e){this.sendSession(e),se(e,{init:!1})}getDsn(){return this._dsn}getOptions(){return this._options}getDataCollectionOptions(){return this._dataCollection}getSdkMetadata(){return this._options._metadata}getTransport(){return this._transport}flush(e){return z(this,null,function*(){const n=this._transport;if(this.emit("flush"),!n)return!0;const r=yield this._isClientDoneProcessing(e),s=yield n.flush(e);return r&&s})}close(e){return z(this,null,function*(){const n=yield this.flush(e);return this.getOptions().enabled=!1,this.emit("close"),n})}getEventProcessors(){return this._eventProcessors}addEventProcessor(e){this._eventProcessors.push(e)}init(){(this._isEnabled()||this._options.integrations.some(({name:e})=>e.startsWith("Spotlight")))&&this._setupIntegrations()}getIntegrationByName(e){return this._integrations[e]}getIntegrationNames(){return Object.keys(this._integrations)}addIntegration(e){const n=this._integrations[e.name];!n&&e.beforeSetup&&e.beforeSetup(this),ks(this,e,this._integrations),n||ws(this,[e])}sendEvent(e,n={}){this.emit("beforeSendEvent",e,n);const r=fc(e,this);let s=Vo(e,this._dsn,this._options._metadata,this._options.tunnel);for(const i of n.attachments||[])s=ts(s,Wo(i));r&&(s=ts(s,r)),this.sendEnvelope(s).then(i=>this.emit("afterSendEvent",e,i))}sendSession(e){const{release:n,environment:r=Sn}=this._options;if("aggregates"in e){const i=e.attrs||{};if(!i.release&&!n){y&&b.warn(xs);return}i.release=i.release||n,i.environment=i.environment||r,e.attrs=i}else{if(!e.release&&!n){y&&b.warn(xs);return}e.release=e.release||n,e.environment=e.environment||r}this.emit("beforeSendSession",e);const s=Ko(e,this._dsn,this._options._metadata,this._options.tunnel);this.sendEnvelope(s)}recordDroppedEvent(e,n,r=1){if(this._options.sendClientReports){const s=`${e}:${n}`;y&&b.log(`Recording outcome: "${s}"${r>1?` (${r} times)`:""}`),this._outcomes[s]=(this._outcomes[s]||0)+r}}on(e,n){const r=this._hooks[e]=this._hooks[e]||new Set,s=(...i)=>n(...i);return r.add(s),()=>{r.delete(s)}}emit(e,...n){const r=this._hooks[e];r&&r.forEach(s=>s(...n))}sendEnvelope(e){return z(this,null,function*(){if(this.emit("beforeEnvelope",e),this._isEnabled()&&this._transport)try{return yield this._transport.send(e)}catch(n){return y&&b.error("Error while sending envelope:",n),{}}return y&&b.error("Transport disabled"),{}})}registerCleanup(e){}dispose(){}_setupIntegrations(){const{integrations:e}=this._options;this._integrations=Wa(this,e),ws(this,e)}_updateSessionFromEvent(e,n){var a,u;let r=n.level==="fatal",s=!1;const i=(a=n.exception)==null?void 0:a.values;if(i){s=!0,r=!1;for(const p of i)if(((u=p.mechanism)==null?void 0:u.handled)===!1){r=!0;break}}const o=e.status==="ok";(o&&e.errors===0||o&&r)&&(se(e,I(l({},r&&{status:"crashed"}),{errors:e.errors||Number(s||r)})),this.captureSession(e))}_isClientDoneProcessing(e){return z(this,null,function*(){let n=0;for(;!e||n<e;){if(yield new Promise(r=>setTimeout(r,1)),!this._numProcessing)return!0;n++}return!1})}_isEnabled(){return this.getOptions().enabled!==!1&&this._transport!==void 0}_prepareEvent(e,n,r,s){const i=this.getOptions(),o=this.getIntegrationNames();return!n.integrations&&o.length&&(n.integrations=o),this.emit("preprocessEvent",e,n),e.type||s.setLastEventId(e.event_id||n.event_id),Sa(i,e,n,r,this,s).then(c=>{var u;if(c===null)return c;this.emit("postprocessEvent",c,n),c.contexts=l({trace:l(l({},(u=c.contexts)==null?void 0:u.trace),Rr(r))},c.contexts);const a=lt(this,r);return c.sdkProcessingMetadata=l({dynamicSamplingContext:a},c.sdkProcessingMetadata),c})}_captureEvent(e,n={},r=T(),s=F()){return y&&jn(e)&&b.log(`Captured error event \`${Os(e)[0]||"<unknown>"}\``),this._processEvent(e,n,r,s).then(i=>i.event_id,i=>{y&&(js(i)?b.log(i.message):$s(i)?b.warn(i.message):b.warn(i))})}_processEvent(e,n,r,s){const i=this.getOptions(),{sampleRate:o}=i,c=Hs(e),a=jn(e),p=`before send for type \`${e.type||"error"}\``,f=typeof o=="undefined"?void 0:_n(o);if(a&&typeof f=="number"&&Ze()>f)return this.recordDroppedEvent("sample_rate","error"),Rn($n(`Discarding event because it's not included in the random sample (sampling rate = ${o})`));const _=Us(e.type);return this._prepareEvent(e,n,r,s).then(d=>{var S;if(d===null)throw this.recordDroppedEvent("event_processor",_),$n("An event processor returned `null`, will not send event.");if(((S=n.data)==null?void 0:S.__sentry__)===!0)return d;const m=kc(this,i,d,n);return wc(m,p)}).then(d=>{var S;if(d===null){if(this.recordDroppedEvent("before_send",_),c){const N=1+(e.spans||[]).length;this.recordDroppedEvent("before_send","span",N)}throw $n(`${p} returned \`null\`, will not send event.`)}const h=r.getSession()||s.getSession();if(a&&h&&this._updateSessionFromEvent(h,d),c){const v=((S=d.sdkProcessingMetadata)==null?void 0:S.spanCountBeforeProcessing)||0,N=d.spans?d.spans.length:0,x=v-N;x>0&&this.recordDroppedEvent("before_send","span",x)}const m=d.transaction_info;if(c&&m&&d.transaction!==e.transaction){const v="custom";d.transaction_info=I(l({},m),{source:v})}return this.sendEvent(d,n),d}).then(null,d=>{throw js(d)||$s(d)?d:(this.captureException(d,{mechanism:{handled:!1,type:"internal"},data:{__sentry__:!0},originalException:d}),_t(`Event processing pipeline threw an error, original event will not be sent. Details have been sent as a new event.
-Reason: ${d}`))})}_process(e,n){this._numProcessing++,this._promiseBuffer.add(e).then(r=>(this._numProcessing--,r),r=>(this._numProcessing--,r===Ln&&this.recordDroppedEvent("queue_overflow",n),r))}_clearOutcomes(){const e=this._outcomes;return this._outcomes={},Object.entries(e).map(([n,r])=>{const[s,i]=n.split(":");return{reason:s,category:i,quantity:r}})}_flushOutcomes(){y&&b.log("Flushing outcomes...");const e=this._clearOutcomes();if(e.length===0){y&&b.log("No outcomes to send");return}if(!this._dsn){y&&b.log("No dsn provided, will not send outcomes");return}y&&b.log("Sending outcomes:",e);const n=gc(e,this._options.tunnel&&Ee(this._dsn));this.sendEnvelope(n)}}function Us(t){return t==="replay_event"?"replay":t||"error"}function wc(t,e){const n=`${e} must return \`null\` or a valid event.`;if(de(t))return t.then(r=>{if(!sn(r)&&r!==null)throw _t(n);return r},r=>{throw _t(`${e} rejected with ${r}`)});if(!sn(t)&&t!==null)throw _t(n);return t}function kc(t,e,n,r){const{beforeSend:s,beforeSendTransaction:i,ignoreSpans:o}=e,c=!es(e.beforeSendSpan)&&e.beforeSendSpan;let a=n;if(jn(a)&&s)return s(a,r);if(Hs(a)){if(c||o){const u=yc(a);if(o!=null&&o.length&&ct({description:u.description,op:u.op,attributes:u.data},o))return null;if(c){const p=c(u);p?a=Me(n,Sc(p)):yn()}if(a.spans){const p=[],f=a.spans;for(const d of f){if(o!=null&&o.length&&ct({description:d.description,op:d.op,attributes:d.data},o)){Bo(f,d);continue}if(c){const h=c(d);h?p.push(h):(yn(),p.push(d))}else p.push(d)}const _=a.spans.length-p.length;_&&t.recordDroppedEvent("before_send","span",_),a.spans=p}}if(i){if(a.spans){const u=a.spans.length;a.sdkProcessingMetadata=I(l({},n.sdkProcessingMetadata),{spanCountBeforeProcessing:u})}return i(a,r)}}return a}function jn(t){return t.type===void 0}function Hs(t){return t.type==="transaction"}function Ac(t){let e=0;return t.name&&(e+=t.name.length*2),e+=8,e+zs(t.attributes)}function Rc(t){let e=0;return t.message&&(e+=t.message.length*2),e+zs(t.attributes)}function zs(t){if(!t)return 0;let e=0;return Object.values(t).forEach(n=>{Array.isArray(n)?e+=n.length*qs(n[0]):hr(n)?e+=qs(n):e+=100}),e}function qs(t){return typeof t=="string"?t.length*2:typeof t=="number"?8:typeof t=="boolean"?4:0}function Cc(t,e){e.debug===!0&&(y?b.enable():fe(()=>{console.warn("[Sentry] Cannot initialize SDK with `debug` option using a non-debug bundle.")})),T().update(e.initialScope);const r=new t(e);return Pc(r),r.init(),r}function Pc(t){T().setClient(t)}const Bn=100,Un=5e3,Nc=36e5;function Ws(t){function e(...n){y&&b.log("[Offline]:",...n)}return n=>{const r=t(n);if(!n.createStore)throw new Error("No `createStore` function was provided");const s=n.createStore(n);let i=Un,o;function c(f,_,d){return En(f,["client_report"])?!1:n.shouldStore?n.shouldStore(f,_,d):!0}function a(f){o&&clearTimeout(o),o=Nn(setTimeout(()=>z(null,null,function*(){o=void 0;const _=yield s.shift();_&&(e("Attempting to send previously queued event"),_[0].sent_at=new Date(ee()).toISOString(),p(_,!0).catch(d=>{e("Failed to retry sending",d)}))}),f))}function u(){o||(a(i),i=Math.min(i*2,Nc))}function p(f,_=!1){return z(this,null,function*(){var d,h;if(!_&&En(f,["replay_event","replay_recording"]))return yield s.push(f),a(Bn),{};try{if(n.shouldSend&&(yield n.shouldSend(f))===!1)throw new Error("Envelope not sent because `shouldSend` callback returned false");const m=yield r.send(f);let S=Bn;if(m){if((d=m.headers)!=null&&d["retry-after"])S=Ns(m.headers["retry-after"]);else if((h=m.headers)!=null&&h["x-sentry-rate-limits"])S=6e4;else if((m.statusCode||0)>=400)return m}return a(S),i=Un,m}catch(m){if(yield c(f,m,i))return _?yield s.unshift(f):yield s.push(f),u(),e("Error sending. Event queued.",m),{};throw m}})}return n.flushAtStartup&&u(),{send:p,flush:f=>(f===void 0&&(i=Un,a(Bn)),r.flush(f))}}}function Dc(t,...e){const n=new String(String.raw(t,...e));return n.__sentry_template_string__=t.join("\0").replace(/%/g,"%%").replace(/\0/g,"%s"),n.__sentry_template_values__=e,n}const Oc=Dc;function Mc(t={}){const e=t.client||P();if(!Ts()||!e)return{};const n=O(),r=Se(n);if(r.getTraceData)return r.getTraceData(t);const s=t.scope||T(),i=t.span||gn(),o=Be(i)&&!je(e.getOptions()),c=i&&!o?Co(i):xc(s),a=i?ve(i):lt(e,s),u=mo(a);if(!Ao.test(c))return b.warn("Invalid sentry-trace data. Cannot generate trace data"),{};const f={"sentry-trace":c,baggage:u};return t.propagateTraceparent&&(f.traceparent=i&&!o?Po(i):Fc(s)),f}function xc(t){const{traceId:e,sampled:n,propagationSpanId:r}=t.getPropagationContext();return Br(e,r,n)}function Fc(t){const{traceId:e,sampled:n,propagationSpanId:r}=t.getPropagationContext();return Ur(e,r,n)}const Lc=100;function M(t,e){const n=P(),r=F();if(!n)return;const{beforeBreadcrumb:s=null,maxBreadcrumbs:i=Lc}=n.getOptions();if(i<=0)return;const o=Oe(),c=l({timestamp:o},t),a=s?fe(()=>s(c,e)):c;a!==null&&(n.emit&&n.emit("beforeAddBreadcrumb",a,e),r.addBreadcrumb(a,i))}const $c=[/^Script error\.?$/,/^Javascript error: Script error\.? on line 0$/,/^ResizeObserver loop completed with undelivered notifications.$/,/^Cannot redefine property: googletag$/,/^Can't find variable: gmo$/,/^undefined is not an object \(evaluating 'a\.[A-Z]'\)$/,/can't redefine non-configurable property "solana"/,/vv\(\)\.getRestrictions is not a function/,/Can't find variable: _AutofillCallbackHandler/,/Object Not Found Matching Id:\d+, MethodName:simulateEvent/,/^Java exception was raised during method invocation$/],jc="EventFilters",Bc=(t={})=>{let e;return{name:jc,setup(n){const r=n.getOptions();e=Gs(t,r)},processEvent(n,r,s){if(!e){const i=s.getOptions();e=Gs(t,i)}return Uc(n,e)?null:n}}};function Gs(t={},e={}){return{allowUrls:[...t.allowUrls||[],...e.allowUrls||[]],denyUrls:[...t.denyUrls||[],...e.denyUrls||[]],ignoreErrors:[...t.ignoreErrors||[],...e.ignoreErrors||[],...t.disableErrorDefaults?[]:$c],ignoreTransactions:[...t.ignoreTransactions||[],...e.ignoreTransactions||[]]}}function Uc(t,e){if(t.type){if(t.type==="transaction"&&zc(t,e.ignoreTransactions))return y&&b.warn(`Event dropped due to being matched by \`ignoreTransactions\` option.
+  `)}`)}function Zo(t){if(!y)return;const{description:e="< unknown name >",op:n="< unknown op >"}=B(t),{spanId:r}=t.spanContext(),i=K(t)===t,o=`[Tracing] Finishing "${n}" ${i?"root ":""}span "${e}" with ID ${r}`;b.log(o)}function Ie(t,e,n,r=gn()){const s=r&&K(r);s&&(y&&b.log(`[Measurement] Setting measurement on root span: ${t} = ${e} ${n}`),s.addEvent(t,{[Nr]:e,[Pr]:n}))}function is(t){if(!t||t.length===0)return;const e={};return t.forEach(n=>{const r=n.attributes||{},s=r[Pr],i=r[Nr];typeof s=="string"&&typeof i=="number"&&(e[n.name]={value:i,unit:s})}),e}function os(){return X(D()).segmentSpanCaptureStrategy}function Ue(t){return t.getOptions().traceLifecycle==="stream"}const as=1e3;class vn{constructor(e={}){this._traceId=e.traceId||me(),this._spanId=e.spanId||ge(),this._startTime=e.startTimestamp||_e(),this._links=e.links,this._attributes={},this.setAttributes(l({[fn]:"manual",[st]:e.op},e.attributes)),this._name=e.name,e.parentSpanId&&(this._parentSpanId=e.parentSpanId),"sampled"in e&&(this._sampled=e.sampled),e.endTimestamp&&(this._endTime=e.endTimestamp),this._events=[],this._isStandaloneSpan=e.isStandalone,this._endTime&&this._onSpanEnded()}addLink(e){return this._frozen?this:(this._links?this._links.push(e):this._links=[e],this)}addLinks(e){return this._frozen?this:(this._links?this._links.push(...e):this._links=e,this)}recordException(e,n){}spanContext(){const{_spanId:e,_traceId:n,_sampled:r}=this;return{spanId:e,traceId:n,traceFlags:r?ot:Hr}}setAttribute(e,n){return this._frozen?this:(n===void 0?delete this._attributes[e]:this._attributes[e]=n,e===Fe&&n!==void 0&&Fr(this)&&po(this),this)}setAttributes(e){return Object.keys(e).forEach(n=>this.setAttribute(n,e[n])),this}updateStartTime(e){this._frozen||(this._startTime=Te(e))}setStatus(e){return this._frozen?this:(this._status=e,this)}updateName(e){return this._frozen?this:(this._name=e,Fr(this)||this.setAttribute(Fe,"custom"),this)}end(e){if(this._endTime){this._frozen=Lr(this);return}this._endTime=Te(e),Zo(this),this._onSpanEnded(),this._frozen=Lr(this)}getSpanJSON(){return{data:this._attributes,description:this._name,op:this._attributes[st],parent_span_id:this._parentSpanId,span_id:this._spanId,start_timestamp:this._startTime,status:Jr(this._status),timestamp:this._endTime,trace_id:this._traceId,origin:this._attributes[fn],profile_id:this._attributes[dn],exclusive_time:this._attributes[pn],measurements:is(this._events),is_segment:this._isStandaloneSpan&&K(this)===this||void 0,segment_id:this._isStandaloneSpan?K(this).spanContext().spanId:void 0,links:Wr(this._links)}}getStreamedSpanJSON(){var e,n;return{name:(e=this._name)!=null?e:"",span_id:this._spanId,trace_id:this._traceId,parent_span_id:this._parentSpanId,start_timestamp:this._startTime,end_timestamp:(n=this._endTime)!=null?n:this._startTime,is_segment:this._isStandaloneSpan||this===K(this),status:Yr(this._status),attributes:Fo(this._attributes,this._status),links:No(this._links)}}isRecording(){return!this._endTime&&!!this._sampled}addEvent(e,n,r){if(this._frozen)return this;y&&b.log("[Tracing] Adding an event to span:",e);const s=cs(n)?n:r||_e(),i=cs(n)?{}:n||{},o={name:e,time:Te(s),attributes:i};return this._events.push(o),this}isStandaloneSpan(){return!!this._isStandaloneSpan}_onSpanEnded(){const e=P();e&&(e.emit("spanEnd",this),this._isStandaloneSpan||e.emit("afterSpanEnd",this));const n=K(this),r=this._isStandaloneSpan||this===n;if(this._isStandaloneSpan){this._sampled?ta(Xo([this],e)):(y&&b.log("[Tracing] Discarding standalone span because its trace was not chosen to be sampled."),e&&e.recordDroppedEvent("sample_rate","span"));return}if(!r){const o=os();if(o){const c=be(this).scope||T();o.onChildSpanEnded(this,n,a=>this._convertSpanToTransaction(a),c)}return}if(e&&Ue(e)){e.emit("afterSegmentSpanEnd",this);return}const s=be(this).scope||T(),i=os();if(i)i.onSegmentSpanEnded(o=>this._convertSpanToTransaction(o),s);else{const o=this._convertSpanToTransaction();o&&s.captureEvent(o)}}_convertSpanToTransaction(e={}){var f,_,d,h;if(!us(B(this)))return;this._name||(y&&b.warn("Transaction has no name, falling back to `<unlabeled transaction>`."),this._name="<unlabeled transaction>");const{scope:n,isolationScope:r}=be(this),s=(f=n==null?void 0:n.getScopeData().sdkProcessingMetadata)==null?void 0:f.normalizedRequest;if(this._sampled!==!0)return;(_=e.onSpanCaptured)==null||_.call(e,this);const i=[];for(const m of Lo(this)){if(m===this||ea(m)||(d=e.isSpanAlreadyCaptured)!=null&&d.call(e,m))continue;const S=B(m);us(S)&&((h=e.onSpanCaptured)==null||h.call(e,m),i.push(S))}const o=this._attributes[Fe];delete this._attributes[Or];let c=!1;i.forEach(m=>{var S;delete m.data[Or],(S=m.op)!=null&&S.startsWith("gen_ai.")&&(c=!0)});const a=l({contexts:{trace:Ro(this)},spans:i.length>as?i.sort((m,S)=>m.start_timestamp-S.start_timestamp).slice(0,as):i,start_timestamp:this._startTime,timestamp:this._endTime,transaction:this._name,type:"transaction",sdkProcessingMetadata:{capturedSpanScope:n,capturedSpanIsolationScope:r,dynamicSamplingContext:ve(this),hasGenAiSpans:c},request:s},o&&{transaction_info:{source:o}}),u=is(this._events);return u&&Object.keys(u).length&&(y&&b.log("[Measurements] Adding measurements to transaction event",JSON.stringify(u,void 0,2)),a.measurements=u),a}}function cs(t){return t&&typeof t=="number"||t instanceof Date||Array.isArray(t)}function us(t){return!!t.start_timestamp&&!!t.timestamp&&!!t.span_id&&!!t.trace_id}function ea(t){return t instanceof vn&&t.isStandaloneSpan()}function ta(t){const e=P();if(!e)return;const n=t[1];if(!n||n.length===0){e.recordDroppedEvent("before_send","span");return}e.sendEnvelope(t)}function na(t,e,n=()=>{},r=()=>{}){let s;try{s=t()}catch(i){throw e(i),n(),i}return ra(s,e,n,r)}function ra(t,e,n,r){return de(t)?kr(t,s=>{n(),r(s)},s=>{e(s),n()}):(n(),r(t),t)}function sa(t,e,n){if(!je(t))return[!1];let r,s;typeof t.tracesSampler=="function"?(s=t.tracesSampler(I(l({},e),{inheritOrSampleWith:c=>typeof e.parentSampleRate=="number"?e.parentSampleRate:typeof e.parentSampled=="boolean"?Number(e.parentSampled):c})),r=!0):e.parentSampled!==void 0?s=e.parentSampled:typeof t.tracesSampleRate!="undefined"&&(s=t.tracesSampleRate,r=!0);const i=_n(s);if(i===void 0)return y&&b.warn(`[Tracing] Discarding root span because of invalid sample rate. Sample rate must be a boolean or a number between 0 and 1. Got ${JSON.stringify(s)} of type ${JSON.stringify(typeof s)}.`),[!1];if(!i)return y&&b.log(`[Tracing] Discarding transaction because ${typeof t.tracesSampler=="function"?"tracesSampler returned 0 or false":"a negative sampling decision was inherited or tracesSampleRate is set to 0"}`),[!1,i,r];const o=n<i;return o||y&&b.log(`[Tracing] Discarding transaction because it's not included in the random sample (sampling rate = ${Number(s)})`),[o,i,r]}const ia="__SENTRY_SUPPRESS_TRACING__";function He(t,e){const n=dt();if(n.startSpan)return n.startSpan(t,e);const r=ps(t),{forceTransaction:s,parentSpan:i,scope:o}=t,c=o==null?void 0:o.clone();return ie(c,()=>ca(i)(()=>{const u=T(),p=_s(u,i),f=P(),d=t.onlyIfParent&&!p?fs(u,f):ds({parentSpan:p,spanArguments:r,forceTransaction:s,scope:u});return(!la(d)||!p)&&nt(u,d),na(()=>e(d),()=>{const{status:h}=B(d);d.isRecording()&&(!h||h==="ok")&&d.setStatus({code:j,message:"internal_error"})},()=>{d.end()})}))}function ft(t){const e=dt();return e.startInactiveSpan?e.startInactiveSpan(t):oa(t)}function oa(t){const e=ps(t),{forceTransaction:n,parentSpan:r}=t;return(t.scope?i=>ie(t.scope,i):r!==void 0?i=>ls(r,i):i=>i())(()=>{const i=T(),o=_s(i,r),c=P();return t.onlyIfParent&&!o?fs(i,c):ds({parentSpan:o,spanArguments:e,forceTransaction:n,scope:i})})}function ls(t,e){const n=dt();return n.withActiveSpan?n.withActiveSpan(t,e):ie(r=>(nt(r,t||void 0),e(r)))}function In(t=T()){const e=dt();return e.isTracingSuppressed?e.isTracingSuppressed(t):t.getScopeData().sdkProcessingMetadata[ia]===!0}function fs(t,e){e==null||e.recordDroppedEvent("no_parent_span","span");const n=new ut({traceId:t.getPropagationContext().traceId});return Le(n,t,F()),n}function ds({parentSpan:t,spanArguments:e,forceTransaction:n,scope:r}){var c;const s=F();if(!je()){const a=l(l({},s.getPropagationContext()),r.getPropagationContext()),u=t?t.spanContext().traceId:a.traceId,p=new ut({traceId:u});return t&&!n&&at(t,p),Le(p,r,s),p}const i=P();if(ua(i,e)){In(r)||i==null||i.recordDroppedEvent("ignored","span");const a=new ut({dropReason:"ignored",traceId:(c=t==null?void 0:t.spanContext().traceId)!=null?c:r.getPropagationContext().traceId});return t&&!n&&at(t,a),Le(a,r,s),a}let o;if(t&&!n)o=aa(t,r,e,s),at(t,o);else if(t){const a=ve(t),{traceId:u,spanId:p}=t.spanContext(),f=oe(t);o=hs(l({traceId:u,parentSpanId:p},e),r,s,f),Qr(o,a)}else{const{traceId:a,dsc:u,parentSpanId:p,sampled:f}=l(l({},s.getPropagationContext()),r.getPropagationContext());o=hs(l({traceId:a,parentSpanId:p},e),r,s,f),u&&Qr(o,u)}return Qo(o),o}function ps(t){const e=t.experimental||{},n=l({isStandalone:e.standalone},t);if(t.startTime){const r=l({},n);return r.startTimestamp=Te(t.startTime),delete r.startTime,r}return n}function dt(){const t=D();return Se(t)}function hs(t,e,n,r){var S,v;const s=P(),i=(s==null?void 0:s.getOptions())||{},{name:o=""}=t,c={spanAttributes:l({},t.attributes),spanName:o,parentSampled:r};s==null||s.emit("beforeSampling",c,{decision:!1});const a=(S=c.parentSampled)!=null?S:r,u=c.spanAttributes,p=e.getPropagationContext(),f=In(e),[_,d,h]=f?[!1]:sa(i,{name:o,parentSampled:a,attributes:u,normalizedRequest:n.getScopeData().sdkProcessingMetadata.normalizedRequest,parentSampleRate:_n((v=p.dsc)==null?void 0:v.sample_rate)},p.sampleRand),m=new vn(I(l({},t),{attributes:l({[Fe]:"custom",[Cr]:d!==void 0&&h?d:void 0},u),sampled:_}));return!_&&s&&!f&&(y&&b.log("[Tracing] Discarding root span because its trace was not chosen to be sampled."),s.recordDroppedEvent("sample_rate",Ue(s)?"span":"transaction")),Le(m,e,n),s&&s.emit("spanStart",m),m}function aa(t,e,n,r){const{spanId:s,traceId:i}=t.spanContext(),o=In(e),c=o?!1:oe(t),a=c?new vn(I(l({},n),{parentSpanId:s,traceId:i,sampled:c})):new ut({traceId:i});at(t,a),Le(a,e,r);const u=P();return u&&(Ue(u)&&Be(a)&&(Be(t)&&t.dropReason?(a.dropReason=t.dropReason,u.recordDroppedEvent(t.dropReason,"span")):o||(a.dropReason="sample_rate",u.recordDroppedEvent("sample_rate","span"))),u.emit("spanStart",a),n.endTimestamp&&(u.emit("spanEnd",a),u.emit("afterSpanEnd",a))),a}function _s(t,e){if(e)return e;if(e===null)return;const n=xe(t);if(!n)return;const r=P();return(r?r.getOptions():{}).parentSpanIsAlwaysRootSpan?K(n):n}function ca(t){return t!==void 0?e=>ls(t,e):e=>e()}function ua(t,e){var r;const n=t==null?void 0:t.getOptions().ignoreSpans;return!t||!Ue(t)||!(n!=null&&n.length)?!1:ct({description:e.name||"",op:((r=e.attributes)==null?void 0:r[st])||e.op,attributes:e.attributes},n)}function la(t){return Be(t)&&t.dropReason==="ignored"}function fa(t,e){const{fingerprint:n,span:r,breadcrumbs:s,sdkProcessingMetadata:i}=e;da(t,e),r&&_a(t,r),ma(t,n),pa(t,s),ha(t,i)}function ms(t,e){const{extra:n,tags:r,attributes:s,user:i,contexts:o,level:c,sdkProcessingMetadata:a,breadcrumbs:u,fingerprint:p,eventProcessors:f,attachments:_,propagationContext:d,transactionName:h,span:m}=e;ze(t,"extra",n),ze(t,"tags",r),ze(t,"attributes",s),ze(t,"user",i),ze(t,"contexts",o),t.sdkProcessingMetadata=Me(t.sdkProcessingMetadata,a,2),c&&(t.level=c),h&&(t.transactionName=h),m&&(t.span=m),u.length&&(t.breadcrumbs=[...t.breadcrumbs,...u]),p.length&&(t.fingerprint=[...t.fingerprint,...p]),f.length&&(t.eventProcessors=[...t.eventProcessors,...f]),_.length&&(t.attachments=[...t.attachments,..._]),t.propagationContext=l(l({},t.propagationContext),d)}function ze(t,e,n){t[e]=Me(t[e],n,1)}function wn(t,e){const n=so().getScopeData();return t&&ms(n,t.getScopeData()),e&&ms(n,e.getScopeData()),n}function da(t,e){const{extra:n,tags:r,user:s,contexts:i,level:o,transactionName:c}=e;Object.keys(n).length&&(t.extra=l(l({},n),t.extra)),Object.keys(r).length&&(t.tags=l(l({},r),t.tags)),Object.keys(s).length&&(t.user=l(l({},s),t.user)),Object.keys(i).length&&(t.contexts=l(l({},i),t.contexts)),o&&(t.level=o),c&&t.type!=="transaction"&&(t.transaction=c)}function pa(t,e){const n=[...t.breadcrumbs||[],...e];t.breadcrumbs=n.length?n:void 0}function ha(t,e){t.sdkProcessingMetadata=l(l({},t.sdkProcessingMetadata),e)}function _a(t,e){t.contexts=l({trace:qr(e)},t.contexts),t.sdkProcessingMetadata=l({dynamicSamplingContext:ve(e)},t.sdkProcessingMetadata);const n=K(e),r=B(n).description;r&&!t.transaction&&t.type==="transaction"&&(t.transaction=r)}function ma(t,e){t.fingerprint=t.fingerprint?Array.isArray(t.fingerprint)?t.fingerprint:[t.fingerprint]:[],e&&(t.fingerprint=t.fingerprint.concat(e)),t.fingerprint.length||delete t.fingerprint}const kn=0,gs=1,ys=2;function An(t){return new qe(e=>{e(t)})}function Rn(t){return new qe((e,n)=>{n(t)})}class qe{constructor(e){this._state=kn,this._handlers=[],this._runExecutor(e)}then(e,n){return new qe((r,s)=>{this._handlers.push([!1,i=>{if(!e)r(i);else try{r(e(i))}catch(o){s(o)}},i=>{if(!n)s(i);else try{r(n(i))}catch(o){s(o)}}]),this._executeHandlers()})}catch(e){return this.then(n=>n,e)}finally(e){return new qe((n,r)=>{let s,i;return this.then(o=>{i=!1,s=o,e&&e()},o=>{i=!0,s=o,e&&e()}).then(()=>{if(i){r(s);return}n(s)})})}_executeHandlers(){if(this._state===kn)return;const e=this._handlers.slice();this._handlers=[],e.forEach(n=>{n[0]||(this._state===gs&&n[1](this._value),this._state===ys&&n[2](this._value),n[0]=!0)})}_runExecutor(e){const n=(i,o)=>{if(this._state===kn){if(de(o)){o.then(r,s);return}this._state=i,this._value=o,this._executeHandlers()}},r=i=>{n(gs,i)},s=i=>{n(ys,i)};try{e(r,s)}catch(i){s(i)}}}function ga(t,e,n,r=0){try{const s=Cn(e,n,t,r);return de(s)?s:An(s)}catch(s){return Rn(s)}}function Cn(t,e,n,r){const s=n[r];if(!t||!s)return t;const i=s(l({},t),e);return y&&i===null&&b.log(`Event processor "${s.id||"?"}" dropped event`),de(i)?i.then(o=>Cn(o,e,n,r+1)):Cn(i,e,n,r+1)}let Z,Ss,bs,ne;function ya(t){const e=C._sentryDebugIds,n=C._debugIds;if(!e&&!n)return{};const r=e?Object.keys(e):[],s=n?Object.keys(n):[];if(ne&&r.length===Ss&&s.length===bs)return ne;Ss=r.length,bs=s.length,ne={},Z||(Z={});const i=(o,c)=>{for(const a of o){const u=c[a],p=Z==null?void 0:Z[a];if(p&&ne&&u)ne[p[0]]=u,Z&&(Z[a]=[p[0],u]);else if(u){const f=t(a);for(let _=f.length-1;_>=0;_--){const d=f[_],h=d==null?void 0:d.filename;if(h&&ne&&Z){ne[h]=u,Z[a]=[h,u];break}}}}};return e&&i(r,e),n&&i(s,n),ne}function Sa(t,e,n,r,s,i){const{normalizeDepth:o=3,normalizeMaxBreadth:c=1e3}=t,a=I(l({},e),{event_id:e.event_id||n.event_id||W(),timestamp:e.timestamp||De()}),u=n.integrations||t.integrations.map(v=>v.name);ba(a,t),va(a,u),s&&s.emit("applyFrameMetadata",e),e.type===void 0&&Ea(a,t.stackParser);const p=wa(r,n.captureContext);n.mechanism&&zi(a,n.mechanism);const f=s?s.getEventProcessors():[],_=wn(i,p),d=[...n.attachments||[],..._.attachments];d.length&&(n.attachments=d),fa(a,_);const h=[...f,..._.eventProcessors];return(n.data&&n.data.__sentry__===!0?An(a):ga(h,a,n)).then(v=>(v&&Ta(v),typeof o=="number"&&o>0?Ia(v,o,c):v))}function ba(t,e){var c,a;const{environment:n,release:r,dist:s,maxValueLength:i}=e;t.environment=t.environment||n||Sn,!t.release&&r&&(t.release=r),!t.dist&&s&&(t.dist=s);const o=t.request;o!=null&&o.url&&i&&(o.url=cn(o.url,i)),i&&((a=(c=t.exception)==null?void 0:c.values)==null||a.forEach(u=>{u.value&&(u.value=cn(u.value,i))}))}function Ea(t,e){var r,s;const n=ya(e);(s=(r=t.exception)==null?void 0:r.values)==null||s.forEach(i=>{var o,c;(c=(o=i.stacktrace)==null?void 0:o.frames)==null||c.forEach(a=>{a.filename&&(a.debug_id=n[a.filename])})})}function Ta(t){var r,s;const e={};if((s=(r=t.exception)==null?void 0:r.values)==null||s.forEach(i=>{var o,c;(c=(o=i.stacktrace)==null?void 0:o.frames)==null||c.forEach(a=>{a.debug_id&&(a.abs_path?e[a.abs_path]=a.debug_id:a.filename&&(e[a.filename]=a.debug_id),delete a.debug_id)})}),Object.keys(e).length===0)return;t.debug_meta=t.debug_meta||{},t.debug_meta.images=t.debug_meta.images||[];const n=t.debug_meta.images;Object.entries(e).forEach(([i,o])=>{n.push({type:"sourcemap",code_file:i,debug_id:o})})}function va(t,e){e.length>0&&(t.sdk=t.sdk||{},t.sdk.integrations=[...t.sdk.integrations||[],...e])}function Ia(t,e,n){var s,i;if(!t)return null;const r=l(l(l(l(l({},t),t.breadcrumbs&&{breadcrumbs:t.breadcrumbs.map(o=>l(l({},o),o.data&&{data:te(o.data,e,n)}))}),t.user&&{user:te(t.user,e,n)}),t.contexts&&{contexts:te(t.contexts,e,n)}),t.extra&&{extra:te(t.extra,e,n)});return(s=t.contexts)!=null&&s.trace&&r.contexts&&(r.contexts.trace=t.contexts.trace,t.contexts.trace.data&&(r.contexts.trace.data=te(t.contexts.trace.data,e,n))),t.spans&&(r.spans=t.spans.map(o=>l(l({},o),o.data&&{data:te(o.data,e,n)}))),(i=t.contexts)!=null&&i.flags&&r.contexts&&(r.contexts.flags=te(t.contexts.flags,3,n)),r}function wa(t,e){if(!e)return t;const n=t?t.clone():new Y;return n.update(e),n}function ka(t){if(t)return Aa(t)?{captureContext:t}:Ca(t)?{captureContext:t}:t}function Aa(t){return t instanceof Y||typeof t=="function"}const Ra=["user","level","extra","contexts","tags","fingerprint","propagationContext"];function Ca(t){return Object.keys(t).some(e=>Ra.includes(e))}function we(t,e){return T().captureException(t,ka(e))}function Pa(t,e){const n=typeof e=="string"?e:void 0,r=typeof e!="string"?{captureContext:e}:void 0;return T().captureMessage(t,n,r)}function Na(t,e){return T().captureEvent(t,e)}function re(t,e){F().setContext(t,e)}function Oa(t){F().setExtras(t)}function Da(t,e){F().setExtra(t,e)}function Ma(t){F().setTags(t)}function xa(t,e){F().setTag(t,e)}function Fa(t){F().setUser(t)}function La(){return F().lastEventId()}function Es(t){return z(this,null,function*(){const e=P();return e?e.flush(t):(y&&b.warn("Cannot flush events. No client defined."),Promise.resolve(!1))})}function $a(t){return z(this,null,function*(){const e=P();return e?e.close(t):(y&&b.warn("Cannot flush events and disable SDK. No client defined."),Promise.resolve(!1))})}function Ts(){const t=P();return(t==null?void 0:t.getOptions().enabled)!==!1&&!!(t!=null&&t.getTransport())}function ja(t){F().addEventProcessor(t)}function vs(t){const e=F(),{user:n}=wn(e,T()),{userAgent:r}=C.navigator||{},s=br(l(l({user:n},r&&{userAgent:r}),t)),i=e.getSession();return(i==null?void 0:i.status)==="ok"&&se(i,{status:"exited"}),pt(),e.setSession(s),s}function pt(){const t=F(),n=T().getSession()||t.getSession();n&&Er(n),Is(),t.setSession()}function Is(){const t=F(),e=P(),n=t.getSession();n&&e&&e.captureSession(n)}function Pn(t=!1){if(t){pt();return}Is()}function Nn(t){return typeof t=="object"&&typeof t.unref=="function"&&t.unref(),t}const Ba="7";function Ua(t){const e=t.protocol?`${t.protocol}:`:"",n=t.port?`:${t.port}`:"";return`${e}//${t.host}${n}${t.path?`/${t.path}`:""}/api/`}function Ha(t){return`${Ua(t)}${t.projectId}/envelope/`}function za(t,e){const n={sentry_version:Ba};return t.publicKey&&(n.sentry_key=t.publicKey),e&&(n.sentry_client=`${e.name}/${e.version}`),new URLSearchParams(n).toString()}function qa(t,e,n){return e||`${Ha(t)}?${za(t,n)}`}const We=[];function Wa(t,e){const n={};return e.forEach(r=>{r!=null&&r.beforeSetup&&r.beforeSetup(t)}),e.forEach(r=>{r&&ks(t,r,n)}),n}function ws(t,e){for(const n of e)n!=null&&n.afterAllSetup&&n.afterAllSetup(t)}function ks(t,e,n){if(n[e.name]){y&&b.log(`Integration skipped because it was already installed: ${e.name}`);return}if(n[e.name]=e,!We.includes(e.name)&&typeof e.setupOnce=="function"&&(e.setupOnce(),We.push(e.name)),e.setup&&typeof e.setup=="function"&&e.setup(t),typeof e.preprocessEvent=="function"){const r=e.preprocessEvent.bind(e);t.on("preprocessEvent",(s,i)=>r(s,i,t))}if(typeof e.processEvent=="function"){const r=e.processEvent.bind(e),s=Object.assign((i,o)=>r(i,o,t),{id:e.name});t.addEventProcessor(s)}["processSpan","processSegmentSpan"].forEach(r=>{const s=e[r];typeof s=="function"&&t.on(r,i=>s.call(e,i,t))}),y&&b.log(`Integration installed: ${e.name}`)}function Ga(t){const e=P();if(!e){y&&b.warn(`Cannot add integration "${t.name}" because no SDK Client is available.`);return}e.addIntegration(t)}function El(t){return t}const Ja="sentry.timestamp.sequence";let On=0,Dn;function Ya(t){const e=Math.floor(t*1e3);Dn!==void 0&&e!==Dn&&(On=0);const n=On;return On++,Dn=e,{key:Ja,value:{value:n,type:"integer"}}}function Ka(t,e){return e?ie(e,()=>{const n=gn(),r=n?qr(n):Rr(e);return[n?ve(n):lt(t,e),r]}):[void 0,void 0]}const Va={trace:1,debug:5,info:9,warn:13,error:17,fatal:21};function Xa(){return typeof __SENTRY_BROWSER_BUNDLE__!="undefined"&&!!__SENTRY_BROWSER_BUNDLE__}function Qa(){return!Xa()&&Object.prototype.toString.call(typeof process!="undefined"?process:0)==="[object process]"}function Mn(){return typeof window!="undefined"&&(!Qa()||Za())}function Za(){const t=C.process;return(t==null?void 0:t.type)==="renderer"}function ec(t,e){const n=e?"auto":"never";return[{type:"log",item_count:t.length,content_type:"application/vnd.sentry.items.log+json"},I(l({version:2},Mn()&&{ingest_settings:{infer_ip:n,infer_user_agent:n}}),{items:t})]}function tc(t,e,n,r,s){const i={};return e!=null&&e.sdk&&(i.sdk={name:e.sdk.name,version:e.sdk.version}),n&&r&&(i.dsn=Ee(r)),ae(i,[ec(t,s)])}const nc=100;function V(t,e,n,r=!0){n&&(!t[e]||r)&&(t[e]=n)}function rc(t,e){const n=xn(),r=Rs(t);r===void 0?n.set(t,[e]):r.length>=nc?(As(t,r),n.set(t,[e])):n.set(t,[...r,e])}function sc(t,e=T(),n=rc){var Kt,Pe,_i;const r=(Kt=e==null?void 0:e.getClient())!=null?Kt:P();if(!r){y&&b.warn("No client available to capture log.");return}const{release:s,environment:i,enableLogs:o=!1,beforeSendLog:c}=r.getOptions();if(!o){y&&b.warn("logging option not enabled, log will not be captured.");return}const[,a]=Ka(r,e),u=l({},t.attributes),{user:{id:p,email:f,username:_},attributes:d={}}=wn(F(),e);V(u,"user.id",p,!1),V(u,"user.email",f,!1),V(u,"user.name",_,!1),V(u,"sentry.release",s),V(u,"sentry.environment",i);const{name:h,version:m}=(_i=(Pe=r.getSdkMetadata())==null?void 0:Pe.sdk)!=null?_i:{};V(u,"sentry.sdk.name",h),V(u,"sentry.sdk.version",m);const S=r.getIntegrationByName("Replay"),v=S==null?void 0:S.getReplayId(!0);V(u,"sentry.replay_id",v),v&&(S==null?void 0:S.getRecordingMode())==="buffer"&&V(u,"sentry._internal.replay_is_buffering",!0);const N=t.message;if(rn(N)){const{__sentry_template_string__:pl,__sentry_template_values__:Vt=[]}=N;Vt!=null&&Vt.length&&(u["sentry.message.template"]=pl),Vt.forEach((hl,_l)=>{u[`sentry.message.parameter.${_l}`]=hl})}const x=xe(e);V(u,"sentry.trace.parent_span_id",x==null?void 0:x.spanContext().spanId);const $=I(l({},t),{attributes:u});r.emit("beforeCaptureLog",$);const k=c?fe(()=>c($)):$;if(!k){r.recordDroppedEvent("before_send","log_item",1),y&&b.warn("beforeSendLog returned null, log will not be captured.");return}const{level:J,message:ce,attributes:Yt={},severityNumber:A}=k,L=_e(),H=Ya(L),ue={timestamp:L,level:J,body:Fn(String(ce)),trace_id:a==null?void 0:a.trace_id,severity_number:A!=null?A:Va[J],attributes:ic(I(l(l({},rt(d)),rt(Yt,!0)),{[H.key]:H.value}))};n(r,ue),r.emit("afterCaptureLog",k)}function As(t,e){var i;const n=(i=e!=null?e:Rs(t))!=null?i:[];if(n.length===0)return;const r=t.getOptions(),s=tc(n,r._metadata,r.tunnel,t.getDsn(),t.getDataCollectionOptions().userInfo);xn().set(t,[]),t.emit("flushLogs"),t.sendEnvelope(s)}function Rs(t){return xn().get(t)}function xn(){return le("clientToLogBufferMap",()=>new WeakMap)}function ic(t){const e={};for(const[n,r]of Object.entries(t)){const s=Fn(n);r.type==="string"?e[s]=I(l({},r),{value:Fn(r.value)}):e[s]=r}return e}function Fn(t){const e=Object(t),n=e.isWellFormed,r=e.toWellFormed;return typeof n=="function"&&typeof r=="function"?n.call(t)?t:r.call(t):t}function oc(t,e){const n=e?"auto":"never";return[{type:"trace_metric",item_count:t.length,content_type:"application/vnd.sentry.items.trace-metric+json"},I(l({version:2},Mn()&&{ingest_settings:{infer_ip:n,infer_user_agent:n}}),{items:t})]}function ac(t,e,n,r,s){const i={};return e!=null&&e.sdk&&(i.sdk={name:e.sdk.name,version:e.sdk.version}),n&&r&&(i.dsn=Ee(r)),ae(i,[oc(t,s)])}function cc(t,e){var i;const n=(i=e!=null?e:uc(t))!=null?i:[];if(n.length===0)return;const r=t.getOptions(),s=ac(n,r._metadata,r.tunnel,t.getDsn(),t.getDataCollectionOptions().userInfo);Cs().set(t,[]),t.emit("flushMetrics"),t.sendEnvelope(s)}function uc(t){return Cs().get(t)}function Cs(){return le("clientToMetricBufferMap",()=>new WeakMap)}function lc(t){const e={trace_id:t.trace_id,span_id:t.span_id,parent_span_id:t.parent_span_id,name:t.description||"",start_timestamp:t.start_timestamp,end_timestamp:t.timestamp||t.start_timestamp,status:!t.status||t.status==="ok"||t.status==="cancelled"?"ok":"error",is_segment:!1,attributes:l({},t.data),links:t.links};return Do(e)}function fc(t,e){var i,o,c;if(t.type!=="transaction"||!((i=t.spans)!=null&&i.length)||!((o=t.sdkProcessingMetadata)!=null&&o.hasGenAiSpans)||e.getOptions().streamGenAiSpans===!1||Ue(e))return;const n=[],r=[];for(const a of t.spans)(c=a.op)!=null&&c.startsWith("gen_ai.")?n.push(lc(a)):r.push(a);if(n.length===0)return;t.spans=r;const s=e.getDataCollectionOptions().userInfo?"auto":"never";return[{type:"span",item_count:n.length,content_type:"application/vnd.sentry.items.span.v2+json"},I(l({version:2},Mn()&&{ingest_settings:{infer_ip:s,infer_user_agent:s}}),{items:n})]}const Ln=Symbol.for("SentryBufferFullError");function Ps(t=100){const e=new Set;function n(){return e.size<t}function r(o){e.delete(o)}function s(o){if(!n())return Rn(Ln);const c=o();return e.add(c),c.then(()=>r(c),()=>r(c)),c}function i(o){if(!e.size)return An(!0);const c=Promise.allSettled(Array.from(e)).then(()=>!0);if(!o)return c;const a=[c,new Promise(u=>Nn(setTimeout(()=>u(!1),o)))];return Promise.race(a)}return{get $(){return Array.from(e)},add:s,drain:i}}const dc=60*1e3;function Ns(t,e=ee()){const n=parseInt(`${t}`,10);if(!isNaN(n))return n*1e3;const r=Date.parse(`${t}`);return isNaN(r)?dc:r-e}function pc(t,e){return t[e]||t.all||0}function hc(t,e,n=ee()){return pc(t,e)>n}function _c(t,{statusCode:e,headers:n},r=ee()){const s=l({},t),i=n==null?void 0:n["x-sentry-rate-limits"],o=n==null?void 0:n["retry-after"];if(i)for(const c of i.trim().split(",")){const[a,u,,,p]=c.split(":",5),f=parseInt(a,10),_=(isNaN(f)?60:f)*1e3;if(!u)s.all=r+_;else for(const d of u.split(";"))d==="metric_bucket"?(!p||p.split(";").includes("custom"))&&(s[d]=r+_):s[d]=r+_}else o?s.all=r+Ns(o,r):e===429&&(s.all=r+60*1e3);return s}const Os=64;function mc(t,e,n=Ps(t.bufferSize||Os)){let r={};const s=o=>n.drain(o);function i(o){const c=[];if(bn(o,(f,_)=>{const d=rs(_);hc(r,d)?t.recordDroppedEvent("ratelimit_backoff",d):c.push(f)}),c.length===0)return Promise.resolve({});const a=ae(o[0],c),u=f=>{if(En(a,["client_report"])){y&&b.warn(`Dropping client report. Will not send outcomes (reason: ${f}).`);return}bn(a,(_,d)=>{t.recordDroppedEvent(f,rs(d))})},p=()=>e({body:Ho(a)}).then(f=>f.statusCode===413?(y&&b.error("Sentry responded with status code 413. Envelope was discarded due to exceeding size limits."),u("send_error"),f):(y&&f.statusCode!==void 0&&(f.statusCode<200||f.statusCode>=300)&&b.warn(`Sentry responded with status code ${f.statusCode} to sent event.`),r=_c(r,f),f),f=>{throw u("network_error"),y&&b.error("Encountered error running transport request:",f),f});return n.add(p).then(f=>f,f=>{if(f===Ln)return y&&b.error("Skipped sending event because buffer is full."),u("queue_overflow"),Promise.resolve({});throw f})}return{send:i,flush:s}}function gc(t,e,n){const r=[{type:"client_report"},{timestamp:De(),discarded_events:t}];return ae(e?{dsn:e}:{},[r])}function Ds(t){const e=[];t.message&&e.push(t.message);try{const n=t.exception.values[t.exception.values.length-1];n!=null&&n.value&&(e.push(n.value),n.type&&e.push(`${n.type}: ${n.value}`))}catch(n){}return e}function yc(t){var a,u,p;const{trace_id:e,parent_span_id:n,span_id:r,status:s,origin:i,data:o,op:c}=(u=(a=t.contexts)==null?void 0:a.trace)!=null?u:{};return{data:o!=null?o:{},description:t.transaction,op:c,parent_span_id:n,span_id:r!=null?r:"",start_timestamp:(p=t.start_timestamp)!=null?p:0,status:s,timestamp:t.timestamp,trace_id:e!=null?e:"",origin:i,profile_id:o==null?void 0:o[dn],exclusive_time:o==null?void 0:o[pn],measurements:t.measurements,is_segment:!0}}function Sc(t){return{type:"transaction",timestamp:t.timestamp,start_timestamp:t.start_timestamp,transaction:t.description,contexts:{trace:{trace_id:t.trace_id,span_id:t.span_id,parent_span_id:t.parent_span_id,op:t.op,status:t.status,origin:t.origin,data:l(l(l({},t.data),t.profile_id&&{[dn]:t.profile_id}),t.exclusive_time&&{[pn]:t.exclusive_time})}},measurements:t.measurements}}const ht=["forwarded","-ip","remote-","via","-user"];function bc(t){return t===!0?{userInfo:!0,cookies:!0,httpHeaders:{request:!0,response:!0},httpBodies:["incomingRequest","outgoingRequest","incomingResponse","outgoingResponse"],urlQueryParams:!0,graphQL:{document:!0,variables:!0},genAI:{inputs:!0,outputs:!0},databaseQueryData:!0,stackFrameVariables:!0,frameContextLines:7}:{userInfo:!1,cookies:{deny:ht},httpHeaders:{request:{deny:ht},response:{deny:ht}},httpBodies:[],urlQueryParams:{deny:ht},graphQL:{document:!0,variables:!0},genAI:{inputs:!1,outputs:!1},databaseQueryData:!1,stackFrameVariables:!0,frameContextLines:7}}const Ec={userInfo:!0,cookies:!0,httpHeaders:{request:!0,response:!0},httpBodies:["incomingRequest","outgoingRequest","incomingResponse","outgoingResponse"],urlQueryParams:!0,graphQL:{document:!0,variables:!0},genAI:{inputs:!0,outputs:!0},databaseQueryData:!0,stackFrameVariables:!0,frameContextLines:5};function Tc(t){var r,s,i,o,c,a,u,p,f,_,d,h,m,S,v,N,x,$,k,J,ce;const e=t.dataCollection!=null?Ec:bc(t.sendDefaultPii),n=(r=t.dataCollection)!=null?r:{};return{userInfo:(s=n.userInfo)!=null?s:e.userInfo,cookies:(i=n.cookies)!=null?i:e.cookies,httpHeaders:{request:(c=(o=n.httpHeaders)==null?void 0:o.request)!=null?c:e.httpHeaders.request,response:(u=(a=n.httpHeaders)==null?void 0:a.response)!=null?u:e.httpHeaders.response},httpBodies:(p=n.httpBodies)!=null?p:e.httpBodies,urlQueryParams:(_=(f=n.urlQueryParams)!=null?f:n.queryParams)!=null?_:e.urlQueryParams,graphQL:{document:(h=(d=n.graphQL)==null?void 0:d.document)!=null?h:e.graphQL.document,variables:(S=(m=n.graphQL)==null?void 0:m.variables)!=null?S:e.graphQL.variables},genAI:{inputs:(N=(v=n.genAI)==null?void 0:v.inputs)!=null?N:e.genAI.inputs,outputs:($=(x=n.genAI)==null?void 0:x.outputs)!=null?$:e.genAI.outputs},databaseQueryData:(k=n.databaseQueryData)!=null?k:e.databaseQueryData,stackFrameVariables:(J=n.stackFrameVariables)!=null?J:e.stackFrameVariables,frameContextLines:(ce=n.frameContextLines)!=null?ce:e.frameContextLines}}const Ms="Not capturing exception because it's already been captured.",xs="Discarded session because of missing or non-string release",Fs=Symbol.for("SentryInternalError"),Ls=Symbol.for("SentryDoNotSendEventError"),vc=5e3;function _t(t){return{message:t,[Fs]:!0}}function $n(t){return{message:t,[Ls]:!0}}function $s(t){return on(t)&&Fs in t}function js(t){return on(t)&&Ls in t}function Bs(t,e,n,r,s){let i=0,o,c=!1;t.on(n,()=>{i=0,clearTimeout(o),c=!1}),t.on(e,a=>{var u;if(i+=r(a),i>=8e5)s(t);else if(!c){const p=(u=t.getOptions()._flushInterval)!=null?u:vc;p>0&&(c=!0,o=Nn(setTimeout(()=>{s(t)},p)))}}),t.on("flush",()=>{s(t)})}class Ic{constructor(e){var r,s,i,o,c,a,u;if(this._options=e,this._integrations={},this._numProcessing=0,this._outcomes={},this._hooks={},this._eventProcessors=[],this._promiseBuffer=Ps((s=(r=e.transportOptions)==null?void 0:r.bufferSize)!=null?s:Os),this._dataCollection=Tc(e),e.dsn?this._dsn=ko(e.dsn):y&&b.warn("No DSN provided, client will not send events."),this._dsn){const p=qa(this._dsn,e.tunnel,e._metadata?e._metadata.sdk:void 0);this._transport=e.transport(I(l({tunnel:this._options.tunnel,recordDroppedEvent:this.recordDroppedEvent.bind(this)},e.transportOptions),{url:p}))}this._options.enableLogs=(o=this._options.enableLogs)!=null?o:(i=this._options._experiments)==null?void 0:i.enableLogs,this._options.enableLogs&&Bs(this,"afterCaptureLog","flushLogs",Rc,As),((u=(a=this._options.enableMetrics)!=null?a:(c=this._options._experiments)==null?void 0:c.enableMetrics)!=null?u:!0)&&Bs(this,"afterCaptureMetric","flushMetrics",Ac,cc)}captureException(e,n,r){const s=W();if(yr(e))return y&&b.log(Ms),s;const i=l({event_id:s},n);return this._process(()=>this.eventFromException(e,i).then(o=>this._captureEvent(o,i,r)).then(o=>o),"error"),i.event_id}captureMessage(e,n,r,s){const i=l({event_id:W()},r),o=rn(e)?e:String(e),c=hr(e),a=c?this.eventFromMessage(o,n,i):this.eventFromException(e,i);return this._process(()=>a.then(u=>this._captureEvent(u,i,s)),c?"unknown":"error"),i.event_id}captureEvent(e,n,r){const s=W();if(n!=null&&n.originalException&&yr(n.originalException))return y&&b.log(Ms),s;const i=l({event_id:s},n),o=e.sdkProcessingMetadata||{},c=o.capturedSpanScope,a=o.capturedSpanIsolationScope,u=Us(e.type);return this._process(()=>this._captureEvent(e,i,c||r,a),u),i.event_id}captureSession(e){this.sendSession(e),se(e,{init:!1})}getDsn(){return this._dsn}getOptions(){return this._options}getDataCollectionOptions(){return this._dataCollection}getSdkMetadata(){return this._options._metadata}getTransport(){return this._transport}flush(e){return z(this,null,function*(){const n=this._transport;if(this.emit("flush"),!n)return!0;const r=yield this._isClientDoneProcessing(e),s=yield n.flush(e);return r&&s})}close(e){return z(this,null,function*(){const n=yield this.flush(e);return this.getOptions().enabled=!1,this.emit("close"),n})}getEventProcessors(){return this._eventProcessors}addEventProcessor(e){this._eventProcessors.push(e)}init(){(this._isEnabled()||this._options.integrations.some(({name:e})=>e.startsWith("Spotlight")))&&this._setupIntegrations()}getIntegrationByName(e){return this._integrations[e]}getIntegrationNames(){return Object.keys(this._integrations)}addIntegration(e){const n=this._integrations[e.name];!n&&e.beforeSetup&&e.beforeSetup(this),ks(this,e,this._integrations),n||ws(this,[e])}sendEvent(e,n={}){this.emit("beforeSendEvent",e,n);const r=fc(e,this);let s=Vo(e,this._dsn,this._options._metadata,this._options.tunnel);for(const i of n.attachments||[])s=ts(s,Wo(i));r&&(s=ts(s,r)),this.sendEnvelope(s).then(i=>this.emit("afterSendEvent",e,i))}sendSession(e){const{release:n,environment:r=Sn}=this._options;if("aggregates"in e){const i=e.attrs||{};if(!i.release&&!n){y&&b.warn(xs);return}i.release=i.release||n,i.environment=i.environment||r,e.attrs=i}else{if(!e.release&&!n){y&&b.warn(xs);return}e.release=e.release||n,e.environment=e.environment||r}this.emit("beforeSendSession",e);const s=Ko(e,this._dsn,this._options._metadata,this._options.tunnel);this.sendEnvelope(s)}recordDroppedEvent(e,n,r=1){if(this._options.sendClientReports){const s=`${e}:${n}`;y&&b.log(`Recording outcome: "${s}"${r>1?` (${r} times)`:""}`),this._outcomes[s]=(this._outcomes[s]||0)+r}}on(e,n){const r=this._hooks[e]=this._hooks[e]||new Set,s=(...i)=>n(...i);return r.add(s),()=>{r.delete(s)}}emit(e,...n){const r=this._hooks[e];r&&r.forEach(s=>s(...n))}sendEnvelope(e){return z(this,null,function*(){if(this.emit("beforeEnvelope",e),this._isEnabled()&&this._transport)try{return yield this._transport.send(e)}catch(n){return y&&b.error("Error while sending envelope:",n),{}}return y&&b.error("Transport disabled"),{}})}registerCleanup(e){}dispose(){}_setupIntegrations(){const{integrations:e}=this._options;this._integrations=Wa(this,e),ws(this,e)}_updateSessionFromEvent(e,n){var a,u;let r=n.level==="fatal",s=!1;const i=(a=n.exception)==null?void 0:a.values;if(i){s=!0,r=!1;for(const p of i)if(((u=p.mechanism)==null?void 0:u.handled)===!1){r=!0;break}}const o=e.status==="ok";(o&&e.errors===0||o&&r)&&(se(e,I(l({},r&&{status:"crashed"}),{errors:e.errors||Number(s||r)})),this.captureSession(e))}_isClientDoneProcessing(e){return z(this,null,function*(){let n=0;for(;!e||n<e;){if(yield new Promise(r=>setTimeout(r,1)),!this._numProcessing)return!0;n++}return!1})}_isEnabled(){return this.getOptions().enabled!==!1&&this._transport!==void 0}_prepareEvent(e,n,r,s){const i=this.getOptions(),o=this.getIntegrationNames();return!n.integrations&&o.length&&(n.integrations=o),this.emit("preprocessEvent",e,n),e.type||s.setLastEventId(e.event_id||n.event_id),Sa(i,e,n,r,this,s).then(c=>{var u;if(c===null)return c;this.emit("postprocessEvent",c,n),c.contexts=l({trace:l(l({},(u=c.contexts)==null?void 0:u.trace),Rr(r))},c.contexts);const a=lt(this,r);return c.sdkProcessingMetadata=l({dynamicSamplingContext:a},c.sdkProcessingMetadata),c})}_captureEvent(e,n={},r=T(),s=F()){return y&&jn(e)&&b.log(`Captured error event \`${Ds(e)[0]||"<unknown>"}\``),this._processEvent(e,n,r,s).then(i=>i.event_id,i=>{y&&(js(i)?b.log(i.message):$s(i)?b.warn(i.message):b.warn(i))})}_processEvent(e,n,r,s){const i=this.getOptions(),{sampleRate:o}=i,c=Hs(e),a=jn(e),p=`before send for type \`${e.type||"error"}\``,f=typeof o=="undefined"?void 0:_n(o);if(a&&typeof f=="number"&&Ze()>f)return this.recordDroppedEvent("sample_rate","error"),Rn($n(`Discarding event because it's not included in the random sample (sampling rate = ${o})`));const _=Us(e.type);return this._prepareEvent(e,n,r,s).then(d=>{var S;if(d===null)throw this.recordDroppedEvent("event_processor",_),$n("An event processor returned `null`, will not send event.");if(((S=n.data)==null?void 0:S.__sentry__)===!0)return d;const m=kc(this,i,d,n);return wc(m,p)}).then(d=>{var S;if(d===null){if(this.recordDroppedEvent("before_send",_),c){const N=1+(e.spans||[]).length;this.recordDroppedEvent("before_send","span",N)}throw $n(`${p} returned \`null\`, will not send event.`)}const h=r.getSession()||s.getSession();if(a&&h&&this._updateSessionFromEvent(h,d),c){const v=((S=d.sdkProcessingMetadata)==null?void 0:S.spanCountBeforeProcessing)||0,N=d.spans?d.spans.length:0,x=v-N;x>0&&this.recordDroppedEvent("before_send","span",x)}const m=d.transaction_info;if(c&&m&&d.transaction!==e.transaction){const v="custom";d.transaction_info=I(l({},m),{source:v})}return this.sendEvent(d,n),d}).then(null,d=>{throw js(d)||$s(d)?d:(this.captureException(d,{mechanism:{handled:!1,type:"internal"},data:{__sentry__:!0},originalException:d}),_t(`Event processing pipeline threw an error, original event will not be sent. Details have been sent as a new event.
+Reason: ${d}`))})}_process(e,n){this._numProcessing++,this._promiseBuffer.add(e).then(r=>(this._numProcessing--,r),r=>(this._numProcessing--,r===Ln&&this.recordDroppedEvent("queue_overflow",n),r))}_clearOutcomes(){const e=this._outcomes;return this._outcomes={},Object.entries(e).map(([n,r])=>{const[s,i]=n.split(":");return{reason:s,category:i,quantity:r}})}_flushOutcomes(){y&&b.log("Flushing outcomes...");const e=this._clearOutcomes();if(e.length===0){y&&b.log("No outcomes to send");return}if(!this._dsn){y&&b.log("No dsn provided, will not send outcomes");return}y&&b.log("Sending outcomes:",e);const n=gc(e,this._options.tunnel&&Ee(this._dsn));this.sendEnvelope(n)}}function Us(t){return t==="replay_event"?"replay":t||"error"}function wc(t,e){const n=`${e} must return \`null\` or a valid event.`;if(de(t))return t.then(r=>{if(!sn(r)&&r!==null)throw _t(n);return r},r=>{throw _t(`${e} rejected with ${r}`)});if(!sn(t)&&t!==null)throw _t(n);return t}function kc(t,e,n,r){const{beforeSend:s,beforeSendTransaction:i,ignoreSpans:o}=e,c=!es(e.beforeSendSpan)&&e.beforeSendSpan;let a=n;if(jn(a)&&s)return s(a,r);if(Hs(a)){if(c||o){const u=yc(a);if(o!=null&&o.length&&ct({description:u.description,op:u.op,attributes:u.data},o))return null;if(c){const p=c(u);p?a=Me(n,Sc(p)):yn()}if(a.spans){const p=[],f=a.spans;for(const d of f){if(o!=null&&o.length&&ct({description:d.description,op:d.op,attributes:d.data},o)){Bo(f,d);continue}if(c){const h=c(d);h?p.push(h):(yn(),p.push(d))}else p.push(d)}const _=a.spans.length-p.length;_&&t.recordDroppedEvent("before_send","span",_),a.spans=p}}if(i){if(a.spans){const u=a.spans.length;a.sdkProcessingMetadata=I(l({},n.sdkProcessingMetadata),{spanCountBeforeProcessing:u})}return i(a,r)}}return a}function jn(t){return t.type===void 0}function Hs(t){return t.type==="transaction"}function Ac(t){let e=0;return t.name&&(e+=t.name.length*2),e+=8,e+zs(t.attributes)}function Rc(t){let e=0;return t.message&&(e+=t.message.length*2),e+zs(t.attributes)}function zs(t){if(!t)return 0;let e=0;return Object.values(t).forEach(n=>{Array.isArray(n)?e+=n.length*qs(n[0]):hr(n)?e+=qs(n):e+=100}),e}function qs(t){return typeof t=="string"?t.length*2:typeof t=="number"?8:typeof t=="boolean"?4:0}function Cc(t,e){e.debug===!0&&(y?b.enable():fe(()=>{console.warn("[Sentry] Cannot initialize SDK with `debug` option using a non-debug bundle.")})),T().update(e.initialScope);const r=new t(e);return Pc(r),r.init(),r}function Pc(t){T().setClient(t)}const Bn=100,Un=5e3,Nc=36e5;function Ws(t){function e(...n){y&&b.log("[Offline]:",...n)}return n=>{const r=t(n);if(!n.createStore)throw new Error("No `createStore` function was provided");const s=n.createStore(n);let i=Un,o;function c(f,_,d){return En(f,["client_report"])?!1:n.shouldStore?n.shouldStore(f,_,d):!0}function a(f){o&&clearTimeout(o),o=Nn(setTimeout(()=>z(null,null,function*(){o=void 0;const _=yield s.shift();_&&(e("Attempting to send previously queued event"),_[0].sent_at=new Date(ee()).toISOString(),p(_,!0).catch(d=>{e("Failed to retry sending",d)}))}),f))}function u(){o||(a(i),i=Math.min(i*2,Nc))}function p(f,_=!1){return z(this,null,function*(){var d,h;if(!_&&En(f,["replay_event","replay_recording"]))return yield s.push(f),a(Bn),{};try{if(n.shouldSend&&(yield n.shouldSend(f))===!1)throw new Error("Envelope not sent because `shouldSend` callback returned false");const m=yield r.send(f);let S=Bn;if(m){if((d=m.headers)!=null&&d["retry-after"])S=Ns(m.headers["retry-after"]);else if((h=m.headers)!=null&&h["x-sentry-rate-limits"])S=6e4;else if((m.statusCode||0)>=400)return m}return a(S),i=Un,m}catch(m){if(yield c(f,m,i))return _?yield s.unshift(f):yield s.push(f),u(),e("Error sending. Event queued.",m),{};throw m}})}return n.flushAtStartup&&u(),{send:p,flush:f=>(f===void 0&&(i=Un,a(Bn)),r.flush(f))}}}function Oc(t,...e){const n=new String(String.raw(t,...e));return n.__sentry_template_string__=t.join("\0").replace(/%/g,"%%").replace(/\0/g,"%s"),n.__sentry_template_values__=e,n}const Dc=Oc;function Mc(t={}){const e=t.client||P();if(!Ts()||!e)return{};const n=D(),r=Se(n);if(r.getTraceData)return r.getTraceData(t);const s=t.scope||T(),i=t.span||gn(),o=Be(i)&&!je(e.getOptions()),c=i&&!o?Co(i):xc(s),a=i?ve(i):lt(e,s),u=mo(a);if(!Ao.test(c))return b.warn("Invalid sentry-trace data. Cannot generate trace data"),{};const f={"sentry-trace":c,baggage:u};return t.propagateTraceparent&&(f.traceparent=i&&!o?Po(i):Fc(s)),f}function xc(t){const{traceId:e,sampled:n,propagationSpanId:r}=t.getPropagationContext();return Br(e,r,n)}function Fc(t){const{traceId:e,sampled:n,propagationSpanId:r}=t.getPropagationContext();return Ur(e,r,n)}const Lc=100;function M(t,e){const n=P(),r=F();if(!n)return;const{beforeBreadcrumb:s=null,maxBreadcrumbs:i=Lc}=n.getOptions();if(i<=0)return;const o=De(),c=l({timestamp:o},t),a=s?fe(()=>s(c,e)):c;a!==null&&(n.emit&&n.emit("beforeAddBreadcrumb",a,e),r.addBreadcrumb(a,i))}const $c=[/^Script error\.?$/,/^Javascript error: Script error\.? on line 0$/,/^ResizeObserver loop completed with undelivered notifications.$/,/^Cannot redefine property: googletag$/,/^Can't find variable: gmo$/,/^undefined is not an object \(evaluating 'a\.[A-Z]'\)$/,/can't redefine non-configurable property "solana"/,/vv\(\)\.getRestrictions is not a function/,/Can't find variable: _AutofillCallbackHandler/,/Object Not Found Matching Id:\d+, MethodName:simulateEvent/,/^Java exception was raised during method invocation$/],jc="EventFilters",Bc=(t={})=>{let e;return{name:jc,setup(n){const r=n.getOptions();e=Gs(t,r)},processEvent(n,r,s){if(!e){const i=s.getOptions();e=Gs(t,i)}return Uc(n,e)?null:n}}};function Gs(t={},e={}){return{allowUrls:[...t.allowUrls||[],...e.allowUrls||[]],denyUrls:[...t.denyUrls||[],...e.denyUrls||[]],ignoreErrors:[...t.ignoreErrors||[],...e.ignoreErrors||[],...t.disableErrorDefaults?[]:$c],ignoreTransactions:[...t.ignoreTransactions||[],...e.ignoreTransactions||[]]}}function Uc(t,e){if(t.type){if(t.type==="transaction"&&zc(t,e.ignoreTransactions))return y&&b.warn(`Event dropped due to being matched by \`ignoreTransactions\` option.
 Event: ${he(t)}`),!0}else{if(Hc(t,e.ignoreErrors))return y&&b.warn(`Event dropped due to being matched by \`ignoreErrors\` option.
 Event: ${he(t)}`),!0;if(Jc(t))return y&&b.warn(`Event dropped due to not having an error message, error type or stacktrace.
 Event: ${he(t)}`),!0;if(qc(t,e.denyUrls))return y&&b.warn(`Event dropped due to being matched by \`denyUrls\` option.
 Event: ${he(t)}.
 Url: ${mt(t)}`),!0;if(!Wc(t,e.allowUrls))return y&&b.warn(`Event dropped due to not being matched by \`allowUrls\` option.
 Event: ${he(t)}.
-Url: ${mt(t)}`),!0}return!1}function Hc(t,e){return e!=null&&e.length?Os(t).some(n=>et(n,e)):!1}function zc(t,e){if(!(e!=null&&e.length))return!1;const n=t.transaction;return n?et(n,e):!1}function qc(t,e){if(!(e!=null&&e.length))return!1;const n=mt(t);return n?et(n,e):!1}function Wc(t,e){if(!(e!=null&&e.length))return!0;const n=mt(t);return n?et(n,e):!0}function Gc(t=[]){for(let e=t.length-1;e>=0;e--){const n=t[e];if(n&&n.filename!=="<anonymous>"&&n.filename!=="[native code]")return n.filename||null}return null}function mt(t){var e,n,r;try{const s=[...(n=(e=t.exception)==null?void 0:e.values)!=null?n:[]].reverse().find(o=>{var c,a,u;return((c=o.mechanism)==null?void 0:c.parent_id)===void 0&&((u=(a=o.stacktrace)==null?void 0:a.frames)==null?void 0:u.length)}),i=(r=s==null?void 0:s.stacktrace)==null?void 0:r.frames;return i?Gc(i):null}catch(s){return y&&b.error(`Cannot extract url for event ${he(t)}`),null}}function Jc(t){var e,n;return(n=(e=t.exception)==null?void 0:e.values)!=null&&n.length?!t.message&&!t.exception.values.some(r=>r.stacktrace||r.type&&r.type!=="Error"||r.value):!1}function Yc(t,e){return t(e.stack||"",1)}function Kc(t){return pr(t)&&"__sentry_fetch_url_host__"in t&&typeof t.__sentry_fetch_url_host__=="string"}function Vc(t){return Kc(t)?`${t.message} (${t.__sentry_fetch_url_host__})`:t.message}function Xc(t,e){const n={type:e.name||e.constructor.name,value:Vc(e)},r=Yc(t,e);return r.length&&(n.stacktrace={frames:r}),n}function ke(t,e,n,r,s){sc({level:t,message:e,attributes:n,severityNumber:s},r)}function Qc(t,e,{scope:n}={}){ke("trace",t,e,n)}function Zc(t,e,{scope:n}={}){ke("debug",t,e,n)}function eu(t,e,{scope:n}={}){ke("info",t,e,n)}function tu(t,e,{scope:n}={}){ke("warn",t,e,n)}function nu(t,e,{scope:n}={}){ke("error",t,e,n)}function ru(t,e,{scope:n}={}){ke("fatal",t,e,n)}const su=Object.freeze(Object.defineProperty({__proto__:null,debug:Zc,error:nu,fatal:ru,fmt:Oc,info:eu,trace:Qc,warn:tu},Symbol.toStringTag,{value:"Module"})),gt="1.15.0",Js="sentry.javascript.miniapp",iu=/^\s*at\s+(?:(.*?)\s*\((.+):(\d+):(\d+)\)|(.+):(\d+):(\d+))\s*$/,ou=/^\s*([^@]*)@(?:(.+):(\d+):(\d+)|(.+):(\d+))\s*$/,au=/^\s*(.+):(\d+):(\d+)\s*$/;function Ae(t){if(!t)return 0;const e=parseInt(t,10);return isNaN(e)?0:e}function Ys(t){return t?/^[a-z][a-z0-9+.-]*:\/\//i.test(t)||t.startsWith("/")||t.includes("/")||/\.(?:cjs|js|jsx|mjs|ts|tsx|vue|json|wxml|axml|swan|ttml)$/i.test(t):!1}function cu(t){const e=iu.exec(t);if(!e)return;const[,n,r,s,i,o,c,a]=e,u=r||o,p={filename:u||"<anonymous>",function:r&&n||Ne,in_app:Hn(u)},f=s||c,_=i||a,d=Ae(f),h=Ae(_);return d&&(p.lineno=d),h&&(p.colno=h),p}function uu(t){const e=ou.exec(t);if(!e)return;const[,n,r,s,i,o,c]=e,a=r||o;if(!Ys(a))return;const u=s||c,p={filename:a||"<anonymous>",function:n||Ne,in_app:Hn(a)},f=Ae(u),_=Ae(i);return f&&(p.lineno=f),_&&(p.colno=_),p}function lu(t){const e=t.trim();if(e.startsWith("at ")||e.includes("@"))return;const n=au.exec(t);if(!n)return;const[,r,s,i]=n;if(!Ys(r))return;const o={filename:r||"<anonymous>",function:Ne,in_app:Hn(r)},c=Ae(s),a=Ae(i);return c&&(o.lineno=c),a&&(o.colno=a),o}function Hn(t){return t?!(t.includes("sentry-miniapp")||t.includes("@sentry")||t.includes("WAService")||t.includes("WASubContext")||t.includes("aframeworkx")||t.includes("__dev__")):!0}const Ks=Ai([90,cu],[80,uu],[70,lu]);let Ge={required:!1},zn=!1;function fu(t){Ge=l({},t),zn=!t.required}function du(t){zn=t}function Vs(){return!Ge.required||zn}function pu(t,e){if(e>0&&Ge.required&&typeof Ge.onDrop=="function")try{Ge.onDrop({reason:t,dropped:e})}catch(n){}}const hu=()=>typeof window!="undefined"&&window!==null&&typeof document!="undefined"&&document!==null,_u=[{global:"wx",name:"wechat"},{global:"my",name:"alipay"},{global:"tt",name:"bytedance"},{global:"dd",name:"dingtalk"},{global:"qq",name:"qq"},{global:"swan",name:"swan"},{global:"ks",name:"kuaishou"}],Xs=()=>{const t=globalThis;for(const e of _u){const n=t[e.global];if(typeof n=="object"&&n!==null)return{sdk:n,name:e.name}}return null},qn=()=>{const t=Xs();if(!t)return hu()?console.warn(`[sentry-miniapp] 检测到当前运行在浏览器/H5 环境（如 uni-app H5、Taro H5）。
+Url: ${mt(t)}`),!0}return!1}function Hc(t,e){return e!=null&&e.length?Ds(t).some(n=>et(n,e)):!1}function zc(t,e){if(!(e!=null&&e.length))return!1;const n=t.transaction;return n?et(n,e):!1}function qc(t,e){if(!(e!=null&&e.length))return!1;const n=mt(t);return n?et(n,e):!1}function Wc(t,e){if(!(e!=null&&e.length))return!0;const n=mt(t);return n?et(n,e):!0}function Gc(t=[]){for(let e=t.length-1;e>=0;e--){const n=t[e];if(n&&n.filename!=="<anonymous>"&&n.filename!=="[native code]")return n.filename||null}return null}function mt(t){var e,n,r;try{const s=[...(n=(e=t.exception)==null?void 0:e.values)!=null?n:[]].reverse().find(o=>{var c,a,u;return((c=o.mechanism)==null?void 0:c.parent_id)===void 0&&((u=(a=o.stacktrace)==null?void 0:a.frames)==null?void 0:u.length)}),i=(r=s==null?void 0:s.stacktrace)==null?void 0:r.frames;return i?Gc(i):null}catch(s){return y&&b.error(`Cannot extract url for event ${he(t)}`),null}}function Jc(t){var e,n;return(n=(e=t.exception)==null?void 0:e.values)!=null&&n.length?!t.message&&!t.exception.values.some(r=>r.stacktrace||r.type&&r.type!=="Error"||r.value):!1}function Yc(t,e){return t(e.stack||"",1)}function Kc(t){return pr(t)&&"__sentry_fetch_url_host__"in t&&typeof t.__sentry_fetch_url_host__=="string"}function Vc(t){return Kc(t)?`${t.message} (${t.__sentry_fetch_url_host__})`:t.message}function Xc(t,e){const n={type:e.name||e.constructor.name,value:Vc(e)},r=Yc(t,e);return r.length&&(n.stacktrace={frames:r}),n}function ke(t,e,n,r,s){sc({level:t,message:e,attributes:n,severityNumber:s},r)}function Qc(t,e,{scope:n}={}){ke("trace",t,e,n)}function Zc(t,e,{scope:n}={}){ke("debug",t,e,n)}function eu(t,e,{scope:n}={}){ke("info",t,e,n)}function tu(t,e,{scope:n}={}){ke("warn",t,e,n)}function nu(t,e,{scope:n}={}){ke("error",t,e,n)}function ru(t,e,{scope:n}={}){ke("fatal",t,e,n)}const su=Object.freeze(Object.defineProperty({__proto__:null,debug:Zc,error:nu,fatal:ru,fmt:Dc,info:eu,trace:Qc,warn:tu},Symbol.toStringTag,{value:"Module"})),gt="1.15.0",Js="sentry.javascript.miniapp",iu=/^\s*at\s+(?:(.*?)\s*\((.+):(\d+):(\d+)\)|(.+):(\d+):(\d+))\s*$/,ou=/^\s*([^@]*)@(?:(.+):(\d+):(\d+)|(.+):(\d+))\s*$/,au=/^\s*(.+):(\d+):(\d+)\s*$/;function Ae(t){if(!t)return 0;const e=parseInt(t,10);return isNaN(e)?0:e}function Ys(t){return t?/^[a-z][a-z0-9+.-]*:\/\//i.test(t)||t.startsWith("/")||t.includes("/")||/\.(?:cjs|js|jsx|mjs|ts|tsx|vue|json|wxml|axml|swan|ttml)$/i.test(t):!1}function cu(t){const e=iu.exec(t);if(!e)return;const[,n,r,s,i,o,c,a]=e,u=r||o,p={filename:u||"<anonymous>",function:r&&n||Ne,in_app:Hn(u)},f=s||c,_=i||a,d=Ae(f),h=Ae(_);return d&&(p.lineno=d),h&&(p.colno=h),p}function uu(t){const e=ou.exec(t);if(!e)return;const[,n,r,s,i,o,c]=e,a=r||o;if(!Ys(a))return;const u=s||c,p={filename:a||"<anonymous>",function:n||Ne,in_app:Hn(a)},f=Ae(u),_=Ae(i);return f&&(p.lineno=f),_&&(p.colno=_),p}function lu(t){const e=t.trim();if(e.startsWith("at ")||e.includes("@"))return;const n=au.exec(t);if(!n)return;const[,r,s,i]=n;if(!Ys(r))return;const o={filename:r||"<anonymous>",function:Ne,in_app:Hn(r)},c=Ae(s),a=Ae(i);return c&&(o.lineno=c),a&&(o.colno=a),o}function Hn(t){return t?!(t.includes("sentry-miniapp")||t.includes("@sentry")||t.includes("WAService")||t.includes("WASubContext")||t.includes("aframeworkx")||t.includes("__dev__")):!0}const Ks=Ai([90,cu],[80,uu],[70,lu]);let Ge={required:!1},zn=!1;function fu(t){Ge=l({},t),zn=!t.required}function du(t){zn=t}function Vs(){return!Ge.required||zn}function pu(t,e){if(e>0&&Ge.required&&typeof Ge.onDrop=="function")try{Ge.onDrop({reason:t,dropped:e})}catch(n){}}function yt(t,e,n){n===null||typeof n!="object"&&typeof n!="function"||e.has(n)||(e.add(n),t.push(n))}const hu=()=>{const t=[],e=new Set;return typeof window!="undefined"&&yt(t,e,window),typeof global!="undefined"&&yt(t,e,global),typeof globalThis!="undefined"&&yt(t,e,globalThis),typeof self!="undefined"&&yt(t,e,self),t},_u=()=>typeof window!="undefined"&&window!==null&&typeof document!="undefined"&&document!==null,mu=[{global:"wx",name:"wechat"},{global:"my",name:"alipay"},{global:"tt",name:"bytedance"},{global:"dd",name:"dingtalk"},{global:"qq",name:"qq"},{global:"swan",name:"swan"},{global:"ks",name:"kuaishou"}],Xs=()=>{const t=globalThis;for(const e of mu){const n=t[e.global];if(typeof n=="object"&&n!==null)return{sdk:n,name:e.name}}return null},qn=()=>{const t=Xs();if(!t)return _u()?console.warn(`[sentry-miniapp] 检测到当前运行在浏览器/H5 环境（如 uni-app H5、Taro H5）。
 本 SDK 仅适配各小程序平台，不支持浏览器原生信号（window.onerror、fetch/XHR 拦截、PerformanceObserver 等）。
 建议改用 Sentry 官方浏览器 SDK：@sentry/browser。
 若使用 uni-app/Taro，可结合条件编译按端引入：H5 用 @sentry/browser，小程序端用 sentry-miniapp。
-详情参考：https://docs.sentry.io/platforms/javascript/`):console.warn("[sentry-miniapp] 未检测到已支持的小程序平台，SDK 将以降级模式运行"),{request:()=>{},httpRequest:()=>{},getSystemInfoSync:()=>({}),URLSearchParams:()=>{}};const e=t.sdk;t.name==="alipay"&&!e.request&&e.httpRequest&&(e.request=e.httpRequest);const n=e;if((t.name==="alipay"||t.name==="dingtalk")&&!n.__sentryStorageAdapted){if(e.getStorageSync){const r=e.getStorageSync;e.getStorageSync=s=>{const i=r.call(e,{key:s});return i?i.data:null}}if(e.setStorageSync){const r=e.setStorageSync;e.setStorageSync=(s,i)=>{r.call(e,{key:s,data:i})}}if(e.removeStorageSync){const r=e.removeStorageSync;e.removeStorageSync=s=>{r.call(e,{key:s})}}n.__sentryStorageAdapted=!0}return e},Qs=()=>{var t,e;return(e=(t=Xs())==null?void 0:t.name)!=null?e:"unknown"},mu=()=>{try{const t=qn(),e={};let n=!1;if(t.getAppBaseInfo){const s=t.getAppBaseInfo();Object.assign(e,s),n=!0}if(t.getWindowInfo){const s=t.getWindowInfo();Object.assign(e,s),n=!0}if(t.getDeviceInfo){const s=t.getDeviceInfo();Object.assign(e,s),n=!0}if(t.getAppAuthorizeSetting){const s=t.getAppAuthorizeSetting();e.albumAuthorized=s.albumAuthorized==="authorized",e.cameraAuthorized=s.cameraAuthorized==="authorized",e.locationAuthorized=s.locationAuthorized==="authorized",e.microphoneAuthorized=s.microphoneAuthorized==="authorized",e.notificationAuthorized=s.notificationAuthorized==="authorized"}if(t.getSystemSetting){const s=t.getSystemSetting();Object.assign(e,s)}if(n&&!!(e.brand||e.model||e.system))return e;if(t.getSystemInfoSync){const s=t.getSystemInfoSync();return!s.SDKVersion&&s.version&&(s.SDKVersion=s.version),s}return n?e:null}catch(t){return console.warn("[sentry-miniapp] Failed to get system info:",t),null}};let Wn=null;const yt=()=>(Wn===null&&(Wn=mu()),Wn),gu=()=>{var t,e;try{const n=qn();if(n.getAccountInfoSync){const r=n.getAccountInfoSync();return{appId:((t=r==null?void 0:r.miniProgram)==null?void 0:t.appId)||"unknown",version:((e=r==null?void 0:r.miniProgram)==null?void 0:e.version)||"unknown"}}}catch(n){}return{appId:"unknown",version:"unknown"}};let Gn=null;const yu=()=>{if(Gn===null){const t=gu();return(t.appId!=="unknown"||t.version!=="unknown")&&(Gn=t),t}return Gn},Zs=()=>Qs()!=="unknown";let Jn=null,Yn=null;const w=()=>(Jn===null&&(Jn=qn()),Jn),ei=()=>(Yn===null&&(Yn=Qs()),Yn);let Kn=null;const Su=()=>{if(Kn===null){const t=globalThis,e=typeof t.GameGlobal!="undefined",n=typeof t.App!="function"&&typeof t.Page!="function"&&typeof t.getCurrentPages!="function";Kn=Zs()&&(e||n)}return Kn},ti=()=>{try{const t=w();if(t.getPerformance&&typeof t.getPerformance=="function")return t.getPerformance()}catch(t){console.warn("Failed to get performance manager:",t)}return null},Je=()=>Date.now(),Vn=()=>Date.now(),ni="application/x-sentry-envelope";function ri(t){const e=t.url,n=t.headers||{};function r(s){return new Promise((i,o)=>{var a,u,p,f;const c={url:e,method:"POST",data:s.body,header:l(l({"Content-Type":ni},n),s.headers),headers:l(l({"Content-Type":ni},n),s.headers),timeout:1e4,success:_=>{const d=_.statusCode||_.status,h=_.header||_.headers||{};i({statusCode:d,headers:{"x-sentry-rate-limits":h["x-sentry-rate-limits"],"retry-after":h["retry-after"]}})},fail:_=>{o(new Error(`Network request failed: ${_.errMsg||_.errorMessage||_.message||"Unknown error"}`))}};w().request?(u=(a=w()).request)==null||u.call(a,c):w().httpRequest?(f=(p=w()).httpRequest)==null||f.call(p,c):o(new Error("No request method available in current miniapp environment"))})}return mc(t,r)}const bu=30,Eu=864e5,Xn="sentry_offline_store",Tu=900*1024;function Qn(t){var p,f,_,d;const e=(p=t.offlineCacheLimit)!=null?p:bu,n=(f=t.offlineCacheMaxAge)!=null?f:Eu,r=(_=t.maxBytes)!=null?_:Tu,s=(d=t.evictionMode)!=null?d:"error-priority",i=t.onDrop;function o(h,m){if(m>0&&typeof i=="function")try{i(h,m)}catch(S){}}function c(h){const m=Date.now(),S=h.filter(v=>m-v.timestamp<n);return o("age",h.length-S.length),S}function a(h,m){let S=0;for(;h.length>e;){if(s==="preserve-oldest")h.pop();else if(m==="push"){const v=h.findIndex(N=>!Zn(N.envelope));v>=0?h.splice(v,1):h.shift()}else{const v=vu(h,N=>!Zn(N.envelope));v>=0?h.splice(v,1):h.pop()}S++}return S}function u(h){const m=w().setStorageSync;if(!m)return;let S=h,v=JSON.stringify(S),N=0;for(;S.length>0&&wu(v)>r;){if(s==="preserve-oldest")S=S.slice(0,S.length-1);else{const x=S.findIndex(k=>!Zn(k.envelope)),$=x>=0?x:0;S=S.slice(0,$).concat(S.slice($+1))}N++,v=JSON.stringify(S)}o("bytes",N);try{m(Xn,v)}catch(x){}}return{push:h=>z(null,null,function*(){try{const m=c(er());m.push({envelope:h,timestamp:Date.now()}),o("count",a(m,"push")),u(m)}catch(m){console.warn("[sentry-miniapp] Failed to push to offline store",m)}}),unshift:h=>z(null,null,function*(){try{const m=c(er());m.unshift({envelope:h,timestamp:Date.now()}),o("count",a(m,"unshift")),u(m)}catch(m){console.warn("[sentry-miniapp] Failed to unshift to offline store",m)}}),shift:()=>z(null,null,function*(){try{const h=er(),m=c(h);if(m.length===0){h.length>0&&u(m);return}const S=m.shift();return u(m),S==null?void 0:S.envelope}catch(h){console.warn("[sentry-miniapp] Failed to shift from offline store",h);return}})}}function Zn(t){try{return(t[1]||[]).some(n=>{var s;const r=(s=n[0])==null?void 0:s.type;return r==="event"||r==="error"})}catch(e){return!0}}function vu(t,e){for(let n=t.length-1;n>=0;n--)if(e(t[n]))return n;return-1}function er(){const t=w().getStorageSync;if(!t)return[];let e;try{e=t(Xn)}catch(n){return[]}if(!e)return[];try{const n=typeof e=="string"?JSON.parse(e):e;return Array.isArray(n)&&n.length>0?n[0]&&"envelope"in n[0]&&"timestamp"in n[0]?n:n.map(r=>({envelope:r,timestamp:Date.now()})):[]}catch(n){return Iu(),[]}}function Iu(){try{const t=w().removeStorageSync;t&&t(Xn)}catch(t){}}function wu(t){let e=0;for(let n=0;n<t.length;n++){const r=t.charCodeAt(n);r<128?e+=1:r<2048?e+=2:r>=55296&&r<=56319?(e+=4,n++):e+=3}return e}const ku=Object.freeze(Object.defineProperty({__proto__:null,createMiniappOfflineStore:Qn,createMiniappTransport:ri},Symbol.toStringTag,{value:"Module"}));function Au(t){return t!==null&&(typeof t=="object"||typeof t=="function")}function Ru(t){return t!==null&&typeof t=="object"}function St(t,e,n){!Au(n)||e.has(n)||(e.add(n),t.push(n))}function Cu(){const t=[],e=new Set;return typeof window!="undefined"&&St(t,e,window),typeof global!="undefined"&&St(t,e,global),typeof globalThis!="undefined"&&St(t,e,globalThis),typeof self!="undefined"&&St(t,e,self),t}function si(t,e,n){const r={};let s=!1;for(const i of e){const o=i[n];if(Ru(o))for(const[c,a]of Object.entries(o))r[c]===void 0&&(r[c]=a,s=!0)}s&&(t[n]=r)}function Pu(){if(typeof globalThis=="undefined")return;const t=globalThis,e=Cu();si(t,e,"_sentryDebugIds"),si(t,e,"_debugIds")}class ii extends Ic{constructor(e={}){fu({required:e.requireConsent===!0,cacheLimit:e.consentCacheLimit,cacheMaxBytes:e.consentCacheMaxBytes,cacheMaxAge:e.consentCacheMaxAge,onDrop:e.onConsentCacheDrop}),super(I(l({},e),{transport:n=>{var i;const r=n,s=e.transport?e.transport(r):ri(I(l({},r),{headers:(i=r.headers)!=null?i:{}}));return e.requireConsent===!0?Ws(()=>s)(I(l({},n),{createStore:o=>{var c;return Qn(I(l({},o),{offlineCacheLimit:(c=e.consentCacheLimit)!=null?c:100,offlineCacheMaxAge:e.consentCacheMaxAge,maxBytes:e.consentCacheMaxBytes,evictionMode:"preserve-oldest",onDrop:pu}))},shouldSend:()=>Vs(),flushAtStartup:!0})):!e.transport&&e.enableOfflineCache!==!1?Ws(()=>s)(I(l({},n),{createStore:o=>Qn(I(l({},o),{offlineCacheLimit:e.offlineCacheLimit,offlineCacheMaxAge:e.offlineCacheMaxAge})),flushAtStartup:!0})):s}}))}eventFromException(e){const n={type:e.name||"Error",value:e.message||String(e)};if(e.stack){const r=this.getOptions().stackParser;if(r&&typeof r=="function"){const s=r(e.stack,1);s.length&&(n.stacktrace={frames:s})}}return Promise.resolve({exception:{values:[n]},level:"error"})}eventFromMessage(e,n="info"){return Promise.resolve({message:e,level:n})}_prepareEvent(e,n,r){e.platform=e.platform||this.getOptions().platform||"javascript",e.sdk=I(l({},e.sdk),{name:Js,packages:[...e.sdk&&e.sdk.packages||[],{name:"npm:@sentry/miniapp",version:gt}],version:gt}),e.contexts||(e.contexts={}),e.contexts.miniapp={platform:ei(),sdk_version:gt};try{if(this.getOptions().stackParser)try{Pu()}catch(o){this.getOptions().debug&&console.warn("[sentry-miniapp] Debug ID 全局同步失败:",o)}const s=r||T(),i=F();return Promise.resolve(super._prepareEvent(e,n||{},s,i)).then(o=>this._fillDefaultContexts(o))}catch(s){return this.getOptions().debug&&console.warn("[sentry-miniapp] _prepareEvent 兜底（scope 未就绪）:",s),Promise.resolve(this._fillDefaultContexts(e))}}_fillDefaultContexts(e){if(!e||this.getOptions().enableSystemInfo===!1)return e;const n=yt();e.contexts||(e.contexts={});const r=e.contexts;return r.device=l({brand:(n==null?void 0:n.brand)||"unknown",model:(n==null?void 0:n.model)||"unknown",screen_resolution:`${(n==null?void 0:n.screenWidth)||0}x${(n==null?void 0:n.screenHeight)||0}`,language:(n==null?void 0:n.language)||"unknown",version:(n==null?void 0:n.version)||"unknown",system:(n==null?void 0:n.system)||"unknown",platform:(n==null?void 0:n.platform)||"unknown"},r.device),r.os=l({name:(n==null?void 0:n.system)||"unknown",version:(n==null?void 0:n.version)||"unknown"},r.os),r.app=l({app_version:(n==null?void 0:n.SDKVersion)||"unknown"},r.app),e}close(e){const n=this.getOptions().integrations;if(Array.isArray(n))for(const r of n){const s=r.cleanup;if(typeof s=="function")try{s.call(r)}catch(i){this.getOptions().debug&&console.warn(`[sentry-miniapp] 集成 ${r.name} cleanup 失败:`,i)}if(Array.isArray(We)){const i=We.indexOf(r.name);i!==-1&&We.splice(i,1)}}return super.close(e)}showReportDialog(e={}){console.warn("[sentry-miniapp] showReportDialog is deprecated and does nothing. Please build your own UI and use `Sentry.captureFeedback()` instead.")}captureFeedback(e){const n={contexts:{feedback:{contact_email:e.email,name:e.name,message:e.message,url:e.url,source:e.source,associated_event_id:e.associatedEventId}},type:"feedback",level:"info",tags:e.tags||{}};return T().captureEvent(n)}}const Ot=class Ot{constructor(e){this.name=Ot.id,this._onErrorHandlerInstalled=!1,this._onUnhandledRejectionHandlerInstalled=!1,this._onPageNotFoundHandlerInstalled=!1,this._onMemoryWarningHandlerInstalled=!1,this._errorHandler=null,this._rejectionHandler=null,this._pageNotFoundHandler=null,this._memoryWarningHandler=null,this._options=l({onerror:!0,onunhandledrejection:!0,onpagenotfound:!0,onmemorywarning:!0},e)}setupOnce(){Error.stackTraceLimit=50,this._options.onerror&&this._installGlobalOnErrorHandler(),this._options.onunhandledrejection&&this._installGlobalOnUnhandledRejectionHandler(),this._options.onpagenotfound&&this._installGlobalOnPageNotFoundHandler(),this._options.onmemorywarning&&this._installGlobalOnMemoryWarningHandler()}_installGlobalOnErrorHandler(){var e,n;this._onErrorHandlerInstalled||(w().onError&&(this._errorHandler=r=>{const s=typeof r=="string"?new Error(r):r;we(s,{mechanism:{type:"onerror",handled:!1}})},(n=(e=w()).onError)==null||n.call(e,this._errorHandler)),this._onErrorHandlerInstalled=!0)}_installGlobalOnUnhandledRejectionHandler(){var e,n;this._onUnhandledRejectionHandlerInstalled||(w().onUnhandledRejection&&(this._rejectionHandler=({reason:r,promise:s})=>{const i=typeof r=="string"?new Error(r):r;we(i,{mechanism:{type:"onunhandledrejection",handled:!1},data:{promise:s}})},(n=(e=w()).onUnhandledRejection)==null||n.call(e,this._rejectionHandler)),this._onUnhandledRejectionHandlerInstalled=!0)}_installGlobalOnPageNotFoundHandler(){var e,n;this._onPageNotFoundHandlerInstalled||(w().onPageNotFound&&(this._pageNotFoundHandler=r=>{const s=T(),i=r.path.split("?")[0];s.setTag("pagenotfound",i),s.setContext("page_not_found",{path:r.path,query:r.query,isEntryPage:r.isEntryPage}),we(new Error(`页面无法找到: ${i}`),{mechanism:{type:"onpagenotfound",handled:!0}})},(n=(e=w()).onPageNotFound)==null||n.call(e,this._pageNotFoundHandler)),this._onPageNotFoundHandlerInstalled=!0)}_installGlobalOnMemoryWarningHandler(){var e,n;this._onMemoryWarningHandlerInstalled||(w().onMemoryWarning&&(this._memoryWarningHandler=({level:r=-1})=>{let s="没有获取到告警级别信息";switch(r){case 5:s="TRIM_MEMORY_RUNNING_MODERATE";break;case 10:s="TRIM_MEMORY_RUNNING_LOW";break;case 15:s="TRIM_MEMORY_RUNNING_CRITICAL";break;default:return}const i=T();i.setTag("memory-warning",String(r)),i.setContext("memory_warning",{level:r,message:s}),we(new Error("内存不足告警"),{mechanism:{type:"onmemorywarning",handled:!0}})},(n=(e=w()).onMemoryWarning)==null||n.call(e,this._memoryWarningHandler)),this._onMemoryWarningHandlerInstalled=!0)}cleanup(){try{const e=w();this._errorHandler&&e.offError&&e.offError(this._errorHandler),this._rejectionHandler&&e.offUnhandledRejection&&e.offUnhandledRejection(this._rejectionHandler),this._pageNotFoundHandler&&e.offPageNotFound&&e.offPageNotFound(this._pageNotFoundHandler),this._memoryWarningHandler&&e.offMemoryWarning&&e.offMemoryWarning(this._memoryWarningHandler)}catch(e){}this._errorHandler=null,this._rejectionHandler=null,this._pageNotFoundHandler=null,this._memoryWarningHandler=null,this._onErrorHandlerInstalled=!1,this._onUnhandledRejectionHandlerInstalled=!1,this._onPageNotFoundHandlerInstalled=!1,this._onMemoryWarningHandlerInstalled=!1}};Ot.id="GlobalHandlers";let bt=Ot;function oi(t,e={},n){if(typeof t!="function")return t;try{if(t.__sentry__)return t;if(t.__sentry_wrapped__)return t.__sentry_wrapped__}catch(s){return t}const r=function(...s){try{return t.apply(this,s)}catch(i){throw ie(o=>{o.addEventProcessor(c=>I(l({},c),{extra:I(l({},c.extra),{arguments:s})})),we(i,e.mechanism?{mechanism:e.mechanism}:void 0)}),i}};try{for(const s in t)Object.prototype.hasOwnProperty.call(t,s)&&(r[s]=t[s])}catch(s){}t.prototype=t.prototype||{},r.prototype=t.prototype,Object.defineProperty(t,"__sentry_wrapped__",{enumerable:!1,value:r}),Object.defineProperties(r,{__sentry__:{enumerable:!1,value:!0},__sentry_original__:{enumerable:!1,value:t}});try{Object.getOwnPropertyDescriptor(r,"name").configurable&&Object.defineProperty(r,"name",{get(){return t.name}})}catch(s){}return r}function ai(t){try{return t&&t.name||"<anonymous>"}catch(e){return"<anonymous>"}}function Ye(t,e,n){if(!(e in t))return;const r=t[e],s=n(r);if(typeof s=="function")try{s.prototype=s.prototype||{},s.prototype.constructor=s}catch(i){}t[e]=s}const Mt=class Mt{constructor(){this.name=Mt.id}_wrapTimeFunction(e){return function(...n){const r=n[0];return n[0]=oi(r,{mechanism:{data:{function:ai(e)},handled:!0,type:"instrument"}}),e.apply(this,n)}}_wrapRAF(e){return function(n){return e(oi(n,{mechanism:{data:{function:"requestAnimationFrame",handler:ai(e)},handled:!0,type:"instrument"}}))}}setupOnce(){const e=globalThis;e.setTimeout&&Ye(e,"setTimeout",this._wrapTimeFunction.bind(this)),e.setInterval&&Ye(e,"setInterval",this._wrapTimeFunction.bind(this)),e.requestAnimationFrame&&Ye(e,"requestAnimationFrame",this._wrapRAF.bind(this))}};Mt.id="TryCatch";let Et=Mt;const Nu="cause",Du=5,xt=class xt{constructor(e={}){this.name=xt.id,this._key=e.key||Nu,this._limit=e.limit||Du}setupOnce(){}processEvent(e,n){return T().getClient()?this._handler(e,n):e}_handler(e,n){if(!e.exception||!e.exception.values||!n||!ci(n.originalException,Error))return e;const r=this._walkErrorTree(n.originalException,this._key);return e.exception.values=[...r,...e.exception.values],e}_walkErrorTree(e,n,r=[]){if(!ci(e[n],Error)||r.length+1>=this._limit)return r;const s=Xc(()=>[],e[n]);return this._walkErrorTree(e[n],n,[s,...r])}};xt.id="LinkedErrors";let Tt=xt;function ci(t,e){try{return t instanceof e}catch(n){return!1}}const Ft=class Ft{constructor(){this.name=Ft.id}setupOnce(){}processEvent(e){var s,i;const n=this._getMiniappVersion(),r=yu();return e.contexts=I(l({},e.contexts),{runtime:I(l({},((s=e.contexts)==null?void 0:s.runtime)||{}),{name:"miniapp",version:n}),app:I(l({},((i=e.contexts)==null?void 0:i.app)||{}),{name:r.appId,version:r.version})}),e}_getMiniappVersion(){const e=yt();return(e==null?void 0:e.version)||(e==null?void 0:e.SDKVersion)||"unknown"}};Ft.id="HttpContext";let vt=Ft;const Lt=class Lt{constructor(e={}){this.name=Lt.id,this._fuzzyMatch=!!e.fuzzyMatch}setupOnce(){}processEvent(e,n){if(e.type)return e;try{if(this._shouldDropEvent(e,this._previousEvent))return null}catch(r){return this._previousEvent=this._cloneEventForComparison(e),e}return this._previousEvent=this._cloneEventForComparison(e),e}_shouldDropEvent(e,n){return n?!!(this._isSameMessageEvent(e,n)||this._isSameExceptionEvent(e,n)):!1}_isSameMessageEvent(e,n){const r=e.message,s=n.message;return!(!r&&!s||r&&!s||!r&&s||r!==s||!this._isSameFingerprint(e,n)||!this._isSameStacktrace(e,n))}_isSameExceptionEvent(e,n){const r=this._getExceptionFromEvent(e),s=this._getExceptionFromEvent(n);if(!r||!s||r.type!==s.type)return!1;if(this._fuzzyMatch){if(!Ou(r.value,s.value))return!1}else if(r.value!==s.value)return!1;return!(!this._isSameFingerprint(e,n)||!this._isSameStacktrace(e,n))}_isSameStacktrace(e,n){let r=this._getFramesFromEvent(e),s=this._getFramesFromEvent(n);if(!r&&!s)return!0;if(r&&!s||!r&&s||(r=r,s=s,s.length!==r.length))return!1;for(let i=0;i<s.length;i++){const o=s[i],c=r[i];if(o.filename!==c.filename||o.lineno!==c.lineno||o.colno!==c.colno||o.function!==c.function)return!1}return!0}_isSameFingerprint(e,n){let r=e.fingerprint,s=n.fingerprint;if(!r&&!s)return!0;if(r&&!s||!r&&s)return!1;r=r,s=s;try{return r.join("")===s.join("")}catch(i){return!1}}_getExceptionFromEvent(e){return e.exception&&e.exception.values&&e.exception.values[0]}_cloneEventForComparison(e){var n;return{message:e.message,fingerprint:e.fingerprint?[...e.fingerprint]:void 0,exception:e.exception?{values:(n=e.exception.values)==null?void 0:n.map(r=>{var s;return{type:r.type,value:r.value,stacktrace:r.stacktrace?{frames:(s=r.stacktrace.frames)==null?void 0:s.map(i=>l({},i))}:void 0}})}:void 0}}_getFramesFromEvent(e){const n=e.exception;if(n)try{return n.values[0].stacktrace.frames}catch(r){}}};Lt.id="Dedupe";let It=Lt;const ui=/\d{10,}|[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}|\d{4}-\d{2}-\d{2}T\d{2}:\d{2}/g;function Ou(t,e){if(t===e)return!0;if(!t||!e)return!1;const n=t.replace(ui,"<dynamic>"),r=e.replace(ui,"<dynamic>");return n===r}const $t=class $t{constructor(){this.name=$t.id}setupOnce(){this._addSystemContext(),this._addNetworkContext(),this._addLocationContext(),this._addStorageContext(),this._checkAppUpdate()}_addSystemContext(){var e,n,r,s;try{const i=yt();if(i){const o=T();o.setContext("device",{name:i.model,model:i.model,brand:i.brand,family:i.platform,arch:i.platform}),o.setContext("os",{name:((e=i.system)==null?void 0:e.split(" ")[0])||"unknown",version:((n=i.system)==null?void 0:n.split(" ")[1])||"unknown",kernel_version:i.version}),o.setContext("app",{app_name:i.appName||"unknown",app_version:i.version}),o.setContext("screen",{screen_width:i.screenWidth,screen_height:i.screenHeight,screen_density:i.pixelRatio}),o.setTag("device.model",i.model),o.setTag("device.brand",i.brand),o.setTag("os.name",((r=i.system)==null?void 0:r.split(" ")[0])||"unknown"),o.setTag("os.version",((s=i.system)==null?void 0:s.split(" ")[1])||"unknown"),o.setTag("app.version",i.version),o.setTag("language",i.language)}}catch(i){}}_addNetworkContext(){try{w().getNetworkType&&w().getNetworkType({success:e=>{const n=T();n.setContext("network",{type:e.networkType,connected:e.isConnected!==!1}),n.setTag("network.type",e.networkType)},fail:()=>{}})}catch(e){}}_addLocationContext(){try{w().getLocation&&w().getLocation({type:"gcj02",success:e=>{T().setContext("location",{latitude:e.latitude,longitude:e.longitude,accuracy:e.accuracy})},fail:()=>{}})}catch(e){}}_addStorageContext(){try{const e=w();if(e&&typeof e.getStorageInfoSync=="function"){const n=e.getStorageInfoSync();if(n){const r=n.currentSize||0,s=n.limitSize||0,i=s>0?Math.round(r/s*100):0;T().setContext("storage",{currentSize:r,limitSize:s,usagePercent:i}),i>=80&&M({category:"storage.warning",message:`存储使用率 ${i}%（${r}KB / ${s}KB）`,level:"warning",data:{currentSize:r,limitSize:s,usagePercent:i}})}}}catch(e){}}_checkAppUpdate(){try{const e=w();if(e&&typeof e.getUpdateManager=="function"){const n=e.getUpdateManager();n&&(typeof n.onCheckForUpdate=="function"&&n.onCheckForUpdate(r=>{r&&r.hasUpdate&&(T().setTag("has_update","true"),M({category:"app.update",message:"检测到新版本可用",level:"info",data:{hasUpdate:!0}}))}),typeof n.onUpdateReady=="function"&&n.onUpdateReady(()=>{M({category:"app.update",message:"新版本已下载完成",level:"info",data:{updateReady:!0}})}))}}catch(e){}}};$t.id="System";let tr=$t;const jt=class jt{constructor(){this.name=jt.id,this._lastRoute="",this._monitorTimer=null}setupOnce(){this._instrumentNavigation(),this._startRouteMonitoring()}_instrumentNavigation(){let e;try{e=w()}catch(r){return}const n=["navigateTo","redirectTo","switchTab","reLaunch"];for(const r of n)if(e[r]){const s=e[r];e[r]=i=>(this._recordNavigation(r,i.url,this._getCurrentRoute()),s.call(e,i))}if(e.navigateBack){const r=e.navigateBack;e.navigateBack=(s={})=>(this._recordNavigation("navigateBack","back",this._getCurrentRoute(),s.delta),r.call(e,s))}}_startRouteMonitoring(){this._monitorTimer=setInterval(()=>{const e=this._getCurrentRoute();e&&e!==this._lastRoute&&(this._recordRouteChange(this._lastRoute,e),this._lastRoute=e)},1e3)}cleanup(){this._monitorTimer&&(clearInterval(this._monitorTimer),this._monitorTimer=null)}_getCurrentRoute(){try{const e=globalThis;if(e.getCurrentPages){const n=e.getCurrentPages();if(n&&n.length>0){const r=n[n.length-1];return r.route||r.__route__||""}}}catch(e){}return""}_recordNavigation(e,n,r,s){const i=T();M({category:"navigation",data:{action:e,from:r,to:n,delta:s},message:`Navigation ${e}: ${r} -> ${n}`,type:"navigation"}),i.setTag("route",n==="back"?r:n),i.setContext("navigation",{action:e,from:r,to:n,delta:s,timestamp:Date.now()})}_recordRouteChange(e,n){const r=T();M({category:"navigation",data:{from:e,to:n},message:`Route changed: ${e} -> ${n}`,type:"navigation"}),r.setTag("route",n),r.setContext("route",{current:n,previous:e,timestamp:Date.now()})}};jt.id="Router";let nr=jt;const Bt=class Bt{constructor(e={}){this.name=Bt.id,this._performanceManager=null,this._observers=[],this._entryBuffer=[],this._reportTimer=null,this._options=I(l({enableNavigation:!0,enableRender:!0,enableResource:!0,enableUserTiming:!1,sampleRate:1,bufferSize:100,reportInterval:3e4,enableMemory:!1},e),{thresholds:l({navigation:3e3,render:1e3,resource:2e3,setData:50},e==null?void 0:e.thresholds)})}setupOnce(){this._initializePerformanceManager(),this._setupPerformanceObservers(),this._startAutoReporting(),this._addPerformanceContext()}_initializePerformanceManager(){try{if(this._performanceManager=ti(),!this._performanceManager){console.warn("[sentry-miniapp] Performance API not available");return}const e=T();e.setTag("performance.api.available",!0),e.setContext("performance",{api_version:"miniapp-1.0",sample_rate:this._options.sampleRate,buffer_size:this._options.bufferSize})}catch(e){console.warn("[sentry-miniapp] Failed to initialize performance manager:",e)}}_setupPerformanceObservers(){if(this._performanceManager)try{const e=[];this._options.enableNavigation&&e.push("navigation"),this._options.enableRender&&e.push("render"),this._options.enableResource&&e.push("resource");let n=this._options.enableUserTiming;if(n)try{const i=yt();if(i&&i.platform==="devtools"&&(n=!1),n){const c=(typeof globalThis!="undefined"?globalThis:typeof window!="undefined"?window:{}).PerformanceObserver,a=c&&Array.isArray(c.supportedEntryTypes)?c.supportedEntryTypes:void 0;(!a||!a.includes("measure")||!a.includes("mark"))&&(n=!1)}}catch(i){n=!1}if(n&&e.push("measure","mark"),e.length===0)return;const r=this._performanceManager.createObserver(i=>{this._handlePerformanceEntries(i)});try{r.observe({entryTypes:e})}catch(i){const o=e.filter(c=>c!=="measure"&&c!=="mark");if(o.length<e.length&&o.length>0)r.observe({entryTypes:o}),console.warn("[sentry-miniapp] Failed to observe all types, falling back to:",o),e.length=0,e.push(...o);else throw i}this._observers.push(r);const s=typeof globalThis!="undefined"?globalThis.process:void 0;s&&s.env&&s.env.NODE_ENV!=="production"&&console.log("[sentry-miniapp] Performance observers setup for:",e)}catch(e){console.warn("[sentry-miniapp] Failed to setup performance observers:",e)}}_handlePerformanceEntries(e){if(!e)return;let n;if(Array.isArray(e))n=e;else if(typeof e=="object"&&e.getEntries)n=e.getEntries();else if(typeof e=="object")n=[e];else{console.warn("[sentry-miniapp] Invalid entries format:",typeof e);return}n.length!==0&&(Math.random()>this._options.sampleRate||n.forEach(r=>{try{this._processPerformanceEntry(r),this._addToBuffer(r)}catch(s){console.warn("[sentry-miniapp] Failed to process performance entry:",s)}}))}_processPerformanceEntry(e){switch(e.entryType){case"navigation":this._processNavigationEntry(e);break;case"render":this._processRenderEntry(e);break;case"resource":this._processResourceEntry(e);break;case"measure":case"mark":this._processUserTimingEntry(e);break;default:console.log("[sentry-miniapp] Unknown entry type:",e.entryType)}}_processNavigationEntry(e){T().addBreadcrumb({message:`页面导航: ${e.name}`,category:"performance.navigation",level:"info",data:{duration:e.duration,appLaunchTime:e.appLaunchTime,pageReadyTime:e.pageReadyTime}}),He({name:`Navigation: ${e.name}`,op:"navigation",startTime:e.startTime/1e3},r=>{r.setAttributes({"navigation.name":e.name,"navigation.duration":e.duration,"navigation.app_launch_time":e.appLaunchTime||0,"navigation.page_ready_time":e.pageReadyTime||0,"navigation.first_render_time":e.firstRenderTime||0}),r.end((e.startTime+e.duration)/1e3)})}_processRenderEntry(e){var r,s;He({name:`Render: ${e.name}`,op:"render",startTime:e.startTime/1e3},i=>{i.setAttributes({"render.name":e.name,"render.duration":e.duration,"render.start":e.renderStart||0,"render.end":e.renderEnd||0,"render.script_start":e.scriptStart||0,"render.script_end":e.scriptEnd||0}),i.end((e.startTime+e.duration)/1e3)});const n=(s=(r=this._options.thresholds)==null?void 0:r.setData)!=null?s:50;e.duration>n&&T().addBreadcrumb({message:`慢渲染检测: ${e.name} (${e.duration.toFixed(1)}ms)`,category:"performance.setData.slow",level:"warning",data:{name:e.name,duration:e.duration,threshold:n,renderStart:e.renderStart,renderEnd:e.renderEnd,scriptStart:e.scriptStart,scriptEnd:e.scriptEnd}})}_processResourceEntry(e){He({name:`Resource: ${e.name}`,op:"resource",startTime:e.startTime/1e3},n=>{n.setAttributes({"resource.name":e.name,"resource.duration":e.duration,"resource.type":e.initiatorType||"unknown","resource.transfer_size":e.transferSize||0,"resource.encoded_size":e.encodedBodySize||0,"resource.decoded_size":e.decodedBodySize||0}),e.fetchStart&&e.responseEnd&&n.setAttributes({"resource.fetch_start":e.fetchStart,"resource.response_end":e.responseEnd,"resource.network_time":e.responseEnd-e.fetchStart}),n.end((e.startTime+e.duration)/1e3)})}_processUserTimingEntry(e){e.entryType==="measure"?He({name:`Measure: ${e.name}`,op:"measure",startTime:e.startTime/1e3},n=>{n.setAttributes({"measure.name":e.name,"measure.duration":e.duration,"measure.detail":e.detail?JSON.stringify(e.detail):void 0}),n.end((e.startTime+e.duration)/1e3)}):e.entryType==="mark"&&T().addBreadcrumb({message:`性能标记: ${e.name}`,category:"performance.mark",level:"info",data:{timestamp:e.startTime,detail:e.detail}})}_addToBuffer(e){for(this._entryBuffer.push(e);this._entryBuffer.length>this._options.bufferSize;)this._entryBuffer.shift()}_startAutoReporting(){this._options.reportInterval<=0||(this._reportTimer=setInterval(()=>{this._reportBufferedEntries()},this._options.reportInterval))}_reportBufferedEntries(){if(this._entryBuffer.length!==0)try{const e=T(),n=this._calculatePerformanceStats(),r=this._collectMemoryInfo();r&&(n.memory=r),e.setContext("performance_summary",l({total_entries:this._entryBuffer.length,navigation_count:this._entryBuffer.filter(s=>s.entryType==="navigation").length,render_count:this._entryBuffer.filter(s=>s.entryType==="render").length,resource_count:this._entryBuffer.filter(s=>s.entryType==="resource").length,measure_count:this._entryBuffer.filter(s=>s.entryType==="measure").length,mark_count:this._entryBuffer.filter(s=>s.entryType==="mark").length,report_time:new Date().toISOString()},n)),this._checkPerformanceThresholds(n),this._reportToNativeAPI(),this._entryBuffer=[]}catch(e){console.warn("[sentry-miniapp] Failed to report buffered entries:",e)}}_calculatePerformanceStats(){const e=this._entryBuffer.filter(i=>i.entryType==="navigation"),n=this._entryBuffer.filter(i=>i.entryType==="render"),r=this._entryBuffer.filter(i=>i.entryType==="resource"),s={};if(e.length>0){const i=e.map(o=>o.duration);s.navigation_stats={avg_duration:i.reduce((o,c)=>o+c,0)/i.length,max_duration:Math.max(...i),min_duration:Math.min(...i)}}if(n.length>0){const i=n.map(c=>c.duration);s.render_stats={avg_duration:i.reduce((c,a)=>c+a,0)/i.length,max_duration:Math.max(...i),min_duration:Math.min(...i)};const o=n.filter(c=>{var a,u;return c.duration>((u=(a=this._options.thresholds)==null?void 0:a.setData)!=null?u:50)});s.render_stats.slow_render_count=o.length,s.render_stats.slow_render_ratio=o.length/n.length}if(r.length>0){const i=r.map(c=>c.duration),o=r.map(c=>c.transferSize||0).filter(c=>c>0);s.resource_stats={avg_load_time:i.reduce((c,a)=>c+a,0)/i.length,max_load_time:Math.max(...i),total_transfer_size:o.reduce((c,a)=>c+a,0),avg_transfer_size:o.length>0?o.reduce((c,a)=>c+a,0)/o.length:0}}return s}_checkPerformanceThresholds(e){var o,c,a,u,p,f,_,d,h;const n=T(),r=(c=(o=this._options.thresholds)==null?void 0:o.navigation)!=null?c:3e3;((a=e.navigation_stats)==null?void 0:a.avg_duration)>r&&n.addBreadcrumb({message:"页面导航性能较慢",category:"performance.warning",level:"warning",data:{avg_duration:e.navigation_stats.avg_duration,threshold:r}});const s=(p=(u=this._options.thresholds)==null?void 0:u.render)!=null?p:1e3;((f=e.render_stats)==null?void 0:f.avg_duration)>s&&n.addBreadcrumb({message:"页面渲染性能较慢",category:"performance.warning",level:"warning",data:{avg_duration:e.render_stats.avg_duration,threshold:s}});const i=(d=(_=this._options.thresholds)==null?void 0:_.resource)!=null?d:2e3;((h=e.resource_stats)==null?void 0:h.avg_load_time)>i&&n.addBreadcrumb({message:"资源加载性能较慢",category:"performance.warning",level:"warning",data:{avg_load_time:e.resource_stats.avg_load_time,threshold:i}})}_reportToNativeAPI(){try{const e=w();if(e.reportPerformance&&this._entryBuffer.length>0){const n={entries:this._entryBuffer.map(r=>({name:r.name,entryType:r.entryType,startTime:r.startTime,duration:r.duration})),timestamp:Date.now(),sampleRate:this._options.sampleRate};e.reportPerformance(n)}}catch(e){console.warn("[sentry-miniapp] Failed to report to native API:",e)}}_collectMemoryInfo(){if(!this._options.enableMemory)return null;try{const e=w();if(e.getPerformance){const n=e.getPerformance();if(n&&n.memory)return{jsHeapSizeUsed:n.memory.jsHeapSizeUsed,jsHeapSizeLimit:n.memory.jsHeapSizeLimit}}}catch(e){}return null}_addPerformanceContext(){try{const e=T(),n=w(),r=!!n.getPerformance,s=!!n.reportPerformance;e.setContext("performance_support",{has_performance_api:r,has_report_api:s,integration_enabled:!0,options:this._options}),e.setTag("performance.integration","enabled")}catch(e){console.warn("[sentry-miniapp] Failed to add performance context:",e)}}cleanup(){this._observers.forEach(e=>{try{e.disconnect()}catch(n){console.warn("[sentry-miniapp] Failed to disconnect observer:",n)}}),this._observers=[],this._reportTimer&&(clearInterval(this._reportTimer),this._reportTimer=null),this._reportBufferedEntries()}};Bt.id="PerformanceAPI";let wt=Bt;const rr=t=>()=>new wt(t),Ut=class Ut{constructor(e={}){this.name=Ut.id,this._prefix=e.prefix||"app:///"}setupOnce(){}processEvent(e){return e.exception&&e.exception.values&&e.exception.values.forEach(n=>{n.stacktrace&&n.stacktrace.frames&&n.stacktrace.frames.forEach(r=>{r.filename&&(r.filename=this._normalizeFilename(r.filename))})}),e}_normalizeFilename(e){let n=e;return n.startsWith(this._prefix)?n:(n=n.replace(/^(appservice|app-service|WAService)\//i,"").replace(/^https?:\/\/[^/]+\//i,"").replace(/^chunks:\/\/\/?/i,"chunks/").replace(/^[a-z]+:\/\//i,"").replace(/^\//,""),`${this._prefix}${n}`)}};Ut.id="RewriteFrames";let kt=Ut;const Ht=class Ht{constructor(e={}){this.name=Ht.id,this._originalRequest=null,this._originalHttpRequest=null,this._traceNetworkBody=!!e.traceNetworkBody,this._sensitiveKeys=(e.sensitiveKeys||["password","passwd","secret","token","access_token","refresh_token","authorization","cookie","session","creditcard","credit_card","card_number","cvv","ssn","id_card"]).map(n=>n.toLowerCase()),this._denyUrls=(e.denyBodyUrls||[]).map(n=>typeof n=="string"?new RegExp(n):n),this._enableTracePropagation=e.enableTracePropagation!==!1,this._tracePropagationTargets=e.tracePropagationTargets||[],this._propagateTraceparent=e.propagateTraceparent===!0}setupOnce(){const e=w();e&&typeof e.request=="function"&&(this._originalRequest=e.request,Ye(e,"request",this._createRequestWrapper.bind(this))),e&&typeof e.httpRequest=="function"&&(this._originalHttpRequest=e.httpRequest,Ye(e,"httpRequest",this._createRequestWrapper.bind(this)))}cleanup(){const e=w();e&&(this._originalRequest&&typeof e.request=="function"&&(e.request=this._originalRequest),this._originalHttpRequest&&typeof e.httpRequest=="function"&&(e.httpRequest=this._originalHttpRequest)),this._originalRequest=null,this._originalHttpRequest=null}_createRequestWrapper(e){const n=this._traceNetworkBody,r=this._sanitizeBody.bind(this),s=this._shouldDenyBodyUrl.bind(this),i=this._enableTracePropagation,o=this._shouldPropagateTrace.bind(this),c=this._propagateTraceparent;return function(a){if(!a||typeof a!="object")return e.call(this,a);const u=Bu(a.url),p=P();let f="";if(p){const A=p.getOptions().dsn;if(A)try{const L=A.match(/^(?:https?:\/\/)?(?:[^@\n]+@)?([^:/\n]+)/i);L&&L[1]&&(f=L[1])}catch(L){}}const _=u.match(/^https?:\/\/([^:/\n]+)/i),d=_&&_[1]?_[1]:"";if(f&&d===f||d==="sentry.io"||d.endsWith(".sentry.io"))return e.call(this,a);const m=Uu(a.method),S=a.data,v=Date.now(),N=Mu(m,u);let x=!1;const $=A=>{x||(x=!0,ju(N,A))};i&&o(u)&&$u(a,N,c);const k={url:u,method:m};if(n&&S&&!s(u))try{const A=typeof S=="string"?S:JSON.stringify(S);k.request_body=r(A),k.request_size=A.length}catch(A){k.request_body="[Cannot serialize request body]"}const J=a.success,ce=a.fail,Yt=a.complete;a.success=function(...A){const L=A[0]||{},H=li(L),ue=Date.now()-v;if(k.status_code=H,k.duration=ue,$({statusCode:H,status:sr(H)?"error":"ok"}),n&&L.data&&!s(u))try{const Pe=typeof L.data=="string"?L.data:JSON.stringify(L.data);k.response_body=r(Pe),k.response_size=Pe.length}catch(Pe){k.response_body="[Cannot serialize response body]"}const Kt=sr(H)||ue>3e3?"warning":"info";if(M({type:"http",category:"xhr",data:k,level:Kt}),typeof J=="function")return J.apply(this,A)},a.fail=function(...A){const L=A[0]||{},H=Date.now()-v,ue=L.errMsg||L.errorMessage||"Network request failed";if(k.error=ue,k.duration=H,$({status:"error",errorMessage:ue}),M({type:"http",category:"xhr",data:k,level:"error"}),typeof ce=="function")return ce.apply(this,A)},a.complete=function(...A){const L=A[0]||{},H=li(L);if($({statusCode:H,status:sr(H)?"error":"ok"}),typeof Yt=="function")return Yt.apply(this,A)};try{return e.call(this,a)}catch(A){throw $({status:"error",errorMessage:A instanceof Error?A.message:String(A)}),A}}}_shouldPropagateTrace(e){return this._tracePropagationTargets.length===0?!0:this._tracePropagationTargets.some(n=>{if(typeof n=="string")return e.includes(n);n.lastIndex=0;const r=n.test(e);return n.lastIndex=0,r})}_shouldDenyBodyUrl(e){return this._denyUrls.some(n=>n.test(e))}_sanitizeBody(e){if(this._sensitiveKeys.length===0)return e;try{const n=JSON.parse(e);if(typeof n=="object"&&n!==null)return this._sanitizeObject(n),JSON.stringify(n)}catch(n){for(const r of this._sensitiveKeys){const s=new RegExp(`(${r})=[^&]*`,"gi");e=e.replace(s,"$1=[Filtered]")}}return e}_sanitizeObject(e){for(const n of Object.keys(e))this._sensitiveKeys.includes(n.toLowerCase())?e[n]="[Filtered]":typeof e[n]=="object"&&e[n]!==null&&this._sanitizeObject(e[n])}};Ht.id="NetworkBreadcrumbs";let At=Ht;function Mu(t,e){try{const n=qu(e);return ft({name:`${t} ${xu(e)}`,op:"http.client",kind:2,attributes:{"http.request.method":t,"url.full":e,"server.address":n||void 0}})}catch(n){return null}}function xu(t){return t.startsWith("data:")?Lu(t):Fu(t).replace(/^([a-z][a-z0-9+.-]*:\/\/)([^/?#@]+@)/i,"$1[filtered]:[filtered]@")}function Fu(t){const e=t.split(/[?#]/,1)[0];return e===void 0?t:e}function Lu(t){const e=t.match(/^data:([^;,]+)/);return`data:${e&&e[1]?e[1]:"text/plain"}`}function $u(t,e,n){if(e)try{const r=Mc(n?{span:e,propagateTraceparent:!0}:{span:e}),s=r["sentry-trace"];if(!s)return;const i=l(l({},fi(t.headers)?t.headers:{}),fi(t.header)?t.header:{});i["sentry-trace"]||(i["sentry-trace"]=s),r.baggage&&(i.baggage=zu(i.baggage,r.baggage)),n&&r.traceparent&&!Hu(i,"traceparent")&&(i.traceparent=r.traceparent),t.header=i,t.headers=i}catch(r){}}function ju(t,e){if(t)try{const n=di(e.statusCode);n!==void 0?co(t,n):t.setStatus({code:e.status==="error"?j:it,message:e.status==="error"?e.errorMessage||"error":"ok"}),e.errorMessage&&t.setAttribute("error.message",e.errorMessage),t.end()}catch(n){}}function Bu(t){return typeof t=="string"?t:t==null?"":String(t)}function Uu(t){return typeof t=="string"&&t.trim()!==""?t.toUpperCase():"GET"}function li(t){if(!(!t||typeof t!="object"))return t.statusCode!==void 0&&t.statusCode!==null?t.statusCode:t.status}function sr(t){const e=di(t);return e!==void 0&&e>=400}function fi(t){return typeof t=="object"&&t!==null&&!Array.isArray(t)}function Hu(t,e){const n=e.toLowerCase();return Object.keys(t).some(r=>r.toLowerCase()===n)}function zu(t,e){const n=typeof t=="string"?t:Array.isArray(t)?t.filter(s=>typeof s=="string").join(","):"";return n?n.split(",").some(s=>s.trim().startsWith("sentry-"))?n:`${n},${e}`:e}function di(t){if(typeof t=="number"&&Number.isFinite(t))return t;if(typeof t=="string"&&t.trim()!==""){const e=Number(t);if(Number.isFinite(e))return e}}function qu(t){try{const e=t.match(/^https?:\/\/([^:/\n]+)/i);return e&&e[1]?e[1]:""}catch(e){return""}}const Wu=["onLaunch","onShow","onHide","onError"],Rt=new Set;let Ct=null,Pt=!1;function Gu(t,e){for(const n of[...Rt]){const r=n[t];if(typeof r=="function")try{r(e)}catch(s){}}}function Ju(){const t=globalThis;if(Pt||typeof t.App!="function")return;const e=t.App;Ct=e;const n=function(r={}){if(r&&typeof r=="object")for(const s of Wu){const i=r[s];r[s]=function(...o){if(Gu(s,o[0]),typeof i=="function")return i.apply(this,o)}}return e.call(this,r)};n.__sentryAppWrapper=!0,t.App=n,Pt=!0}function Yu(){if(!Pt||Rt.size>0)return;const t=globalThis;t.App&&t.App.__sentryAppWrapper&&Ct&&(t.App=Ct),Ct=null,Pt=!1}function pi(t){if(typeof globalThis.App!="function")return()=>{};Rt.add(t),Ju();let e=!0;return()=>{e&&(e=!1,Rt.delete(t),Yu())}}const hi=["onLoad","onShow","onHide","onUnload","onReady"];function Ku(t){return hi.includes(t)||t.startsWith("_")?!1:/^(on|handle|bind)[A-Z]/.test(t)||/[Tt]ap$/.test(t)||/[Cc]lick$/.test(t)||/[Cc]hange$/.test(t)||/[Ss]ubmit$/.test(t)||/[Ss]croll$/.test(t)||/[Ii]nput$/.test(t)}const zt=class zt{constructor(e={}){this.name=zt.id,this._originalPage=null,this._wrappedPage=null,this._unsubscribeApp=null,this._launchTime=0,this._firstPageReady=!1,this._options=l({enableLifecycle:!0,enableUserInteraction:!0},e)}setupOnce(){this._wrapPage(),this._subscribeApp()}_wrapPage(){const e=globalThis;if(typeof e.Page!="function"||e.Page.__sentryPageWrapper)return;this._originalPage=e.Page;const n=e.Page,r=this._options,s=()=>this._firstPageReady,i=a=>{this._firstPageReady=a},o=()=>this._launchTime,c=function(a){if(a&&typeof a=="object"){if(r.enableLifecycle){for(const u of hi)if(typeof a[u]=="function"){const p=a[u];a[u]=function(...f){const _=(this==null?void 0:this.route)||(this==null?void 0:this.__route__)||"unknown",d={action:u,page:_};if(u==="onLoad"&&f[0]&&typeof f[0]=="object"&&(d.query=f[0]),u==="onReady"&&!s()&&o()>0){i(!0);const h=Date.now()-o();d.coldStartDuration=h,re("startup",{coldStartDuration:h,firstPage:_})}return M({category:"page.lifecycle",message:`${u}: ${_}`,level:"info",data:d}),p.apply(this,f)}}}if(r.enableUserInteraction){for(const u of Object.keys(a))if(typeof a[u]=="function"&&Ku(u)){const p=a[u];a[u]=function(f,..._){const d=(this==null?void 0:this.route)||(this==null?void 0:this.__route__)||"unknown",h={handler:u,page:d};if(f&&typeof f=="object"&&(f.target&&(f.target.id&&(h.targetId=f.target.id),f.target.dataset&&(h.dataset=f.target.dataset)),f.type&&(h.eventType=f.type),f.detail&&(typeof f.detail.x=="number"&&(h.x=f.detail.x),typeof f.detail.y=="number"&&(h.y=f.detail.y)),f.touches&&f.touches.length>0)){const m=f.touches[0];m&&(h.touchX=m.pageX,h.touchY=m.pageY)}return M({category:"user.interaction",message:`${u} on ${d}`,level:"info",data:h}),p.apply(this,[f,..._])}}}}return n(a)};Object.defineProperty(c,"__sentryPageWrapper",{value:!0,configurable:!0}),this._wrappedPage=c,e.Page=c}_subscribeApp(){this._options.enableLifecycle&&(this._unsubscribeApp=pi({onLaunch:()=>{this._launchTime=Date.now(),this._appBreadcrumb("onLaunch")},onShow:()=>this._appBreadcrumb("onShow"),onHide:()=>this._appBreadcrumb("onHide")}))}_appBreadcrumb(e){M({category:"app.lifecycle",message:`App.${e}`,level:"info",data:{action:e}})}cleanup(){const e=globalThis;this._originalPage&&(e.Page===this._wrappedPage&&(e.Page=this._originalPage),this._originalPage=null,this._wrappedPage=null),this._unsubscribeApp&&(this._unsubscribeApp(),this._unsubscribeApp=null)}};zt.id="PageBreadcrumbs";let Ke=zt;const Vu=t=>new Ke(t),Xu=["debug","info","warn","error","log"],Qu={debug:"debug",info:"info",log:"info",warn:"warning",error:"error"},qt=class qt{constructor(e={}){this.name=qt.id,this._originalMethods={},this._levels=e.levels||[...Xu]}setupOnce(){for(const e of this._levels){if(typeof console[e]!="function")continue;const n=console[e];this._originalMethods[e]=n,console[e]=function(...r){return M({category:"console",level:Qu[e],message:r.map(s=>{if(typeof s=="string")return s;try{return JSON.stringify(s)}catch(i){return String(s)}}).join(" ")}),n.apply(console,r)}}}cleanup(){for(const e of this._levels){const n=this._originalMethods[e];n&&(console[e]=n)}this._originalMethods={}}};qt.id="ConsoleBreadcrumbs";let Ve=qt;const Zu=t=>new Ve(t),Wt=class Wt{constructor(){this.name=Wt.id,this._isSessionActive=!1,this._unsubscribe=null}setupOnce(){this._unsubscribe=pi({onLaunch:()=>this._startSession(),onShow:()=>{this._isSessionActive||this._startSession()},onHide:()=>this._endSession()})}_startSession(){try{vs({ignoreDuration:!0}),Pn(),this._isSessionActive=!0}catch(e){}}_endSession(){try{pt(),Pn(),this._isSessionActive=!1}catch(e){}}cleanup(){this._unsubscribe&&(this._unsubscribe(),this._unsubscribe=null),this._isSessionActive=!1}};Wt.id="Session";let Nt=Wt;const Gt=class Gt{constructor(){this.name=Gt.id,this._statusChangeHandler=null,this._lastConnected=null}setupOnce(){const e=w();if(e){if(typeof e.getNetworkType=="function")try{e.getNetworkType({success:n=>{const r=n.networkType||"unknown";this._lastConnected=r!=="none",re("network",{type:r,isConnected:this._lastConnected})}})}catch(n){}typeof e.onNetworkStatusChange=="function"&&(this._statusChangeHandler=n=>{var i;const r=n.networkType||"unknown",s=n.isConnected!==void 0?n.isConnected:r!=="none";if(re("network",{type:r,isConnected:s}),M({category:"network.change",message:`网络状态变化: ${r}`,level:s?"info":"warning",data:{networkType:r,isConnected:s}}),s&&this._lastConnected===!1)try{(i=P())==null||i.flush()}catch(o){}this._lastConnected=s},e.onNetworkStatusChange(this._statusChangeHandler))}}cleanup(){if(this._statusChangeHandler){const e=w();if(e&&typeof e.offNetworkStatusChange=="function")try{e.offNetworkStatusChange(this._statusChangeHandler)}catch(n){}this._statusChangeHandler=null}}};Gt.id="NetworkStatus";let Dt=Gt;const Jt=class Jt{constructor(){this.name=Jt.id,this._initTs=Je(),this._initEpoch=Vn(),this._showHandler=null,this._hideHandler=null,this._coldStartReported=!1,this._minigameContext={runtime:"minigame"}}setupOnce(){const e=w();if(e){if(typeof e.getLaunchOptionsSync=="function")try{const n=e.getLaunchOptionsSync()||{};this._minigameContext.scene=n.scene,this._minigameContext.path=n.path,this._minigameContext.query=n.query,re("minigame",l({},this._minigameContext)),M({category:"minigame.launch",message:"小游戏冷启动",level:"info",data:{scene:n.scene,path:n.path}})}catch(n){}this._measureColdStart(),typeof e.onShow=="function"&&(this._showHandler=n=>{M({category:"minigame.lifecycle",message:"小游戏 onShow（进入前台）",level:"info",data:{scene:n&&n.scene}})},e.onShow(this._showHandler)),typeof e.onHide=="function"&&(this._hideHandler=()=>{M({category:"minigame.lifecycle",message:"小游戏 onHide（退到后台）",level:"info"})},e.onHide(this._hideHandler))}}_measureColdStart(){const e=globalThis.requestAnimationFrame;typeof e=="function"&&e(()=>{if(this._coldStartReported)return;this._coldStartReported=!0;const n=Je(),r=Math.max(0,Math.round(n-this._initTs));this._minigameContext.coldStartMs=r,re("minigame",l({},this._minigameContext)),M({category:"minigame.performance",message:`SDK 初始化到首帧耗时: ${r}ms`,level:"info",data:{coldStartMs:r}});const s=ft({name:"minigame.coldstart",op:"app.start",forceTransaction:!0,startTime:this._initEpoch/1e3});s.setAttributes({"minigame.scene":this._minigameContext.scene,"minigame.path":this._minigameContext.path,"minigame.cold_start_ms":r}),Ie("cold_start",r,"millisecond",s),s.end((this._initEpoch+r)/1e3)})}cleanup(){const e=w();if(e){try{this._showHandler&&typeof e.offShow=="function"&&e.offShow(this._showHandler),this._hideHandler&&typeof e.offHide=="function"&&e.offHide(this._hideHandler)}catch(n){}this._showHandler=null,this._hideHandler=null}}};Jt.id="Minigame";let Re=Jt;const el=()=>new Re,tl=["minor","major","severe"];function nl(t){if(!t)return[];const e=[];for(const n of tl){const r=t[n];typeof r=="number"&&Number.isFinite(r)&&r>0&&e.push({name:n,threshold:r})}if(e.length===0)return[];for(let n=1;n<e.length;n++){const r=e[n-1],s=e[n];if(r&&s&&s.threshold<=r.threshold)return console.warn("[sentry-miniapp] jankLevels 阈值须按 minor < major < severe 严格递增，当前配置不满足，已忽略分级并回退到单档 longFrameThresholdMs。"),[]}return e}function rl(t,e){for(let n=t.length-1;n>=0;n--){const r=t[n];if(r&&e>r.threshold)return r.name}return null}function sl(t,e){var r;if(t.length===0)return e;const n=[...t].sort((s,i)=>s-i);return(r=n[Math.floor(.95*(n.length-1))])!=null?r:e}const G=class G{constructor(e={}){var r,s,i,o;this.name=G.id,this._running=!1,this._windowElapsed=0,this._lastFrameTs=0,this._frameCount=0,this._jankCount=0,this._jankBreadcrumbs=0,this._jankByTier={minor:0,major:0,severe:0},this._maxFrameDelta=0,this._sessionEpochStart=0,this._sessionElapsed=0,this._sessionFrames=0,this._sessionJank=0,this._sessionJankByTier={minor:0,major:0,severe:0},this._sessionWorstFrame=0,this._fpsSamples=[],this._showHandler=null,this._hideHandler=null,this._options={fpsWarningThreshold:(r=e.fpsWarningThreshold)!=null?r:30,longFrameThresholdMs:(s=e.longFrameThresholdMs)!=null?s:50,reportInterval:(i=e.reportInterval)!=null?i:1e4,maxJankBreadcrumbsPerWindow:(o=e.maxJankBreadcrumbsPerWindow)!=null?o:3},this._tiers=nl(e.jankLevels);const n=this._tiers[0];this._tiered=n!==void 0,this._jankEntryThreshold=n?n.threshold:this._options.longFrameThresholdMs}setupOnce(){const e=globalThis.requestAnimationFrame;if(typeof e!="function"){console.warn("[sentry-miniapp] requestAnimationFrame 不可用，小游戏帧率监控已跳过（小程序逻辑层不支持）");return}this._running=!0;const n=Je();this._lastFrameTs=n,this._sessionEpochStart=Vn();const r=()=>{if(!this._running)return;const i=Je();this._onFrame(i-this._lastFrameTs),this._lastFrameTs=i,e(r)};e(r);const s=w();s&&typeof s.onHide=="function"&&(this._hideHandler=()=>{this._flushSummary()&&this._flushPendingEvents()},s.onHide(this._hideHandler)),s&&typeof s.onShow=="function"&&(this._showHandler=()=>this._restartOnResume(),s.onShow(this._showHandler))}_onFrame(e){const n=this._sanitizeFrameDelta(e);if(n===null){(!Number.isFinite(e)||e>G._MAX_REASONABLE_FRAME_DELTA_MS)&&this._rollupWindow();return}if(this._frameCount+=1,this._windowElapsed+=n,n>this._maxFrameDelta&&(this._maxFrameDelta=n),n>this._jankEntryThreshold){this._jankCount+=1;const r=this._tiered?rl(this._tiers,n):null;r&&(this._jankByTier[r]+=1),this._jankBreadcrumbs<this._options.maxJankBreadcrumbsPerWindow&&(this._jankBreadcrumbs+=1,M({category:"minigame.jank",message:`检测到卡顿帧: ${Math.round(n)}ms`,level:"warning",data:r?{frameDurationMs:Math.round(n),jankLevel:r}:{frameDurationMs:Math.round(n)}}))}this._windowElapsed>=this._options.reportInterval&&this._report()}_sanitizeFrameDelta(e){return!Number.isFinite(e)||e<=0||e>G._MAX_REASONABLE_FRAME_DELTA_MS?null:e}_report(){const e=this._rollupWindow();e&&(re("minigame.framerate",{fps:e.fps,minFps:e.minFps,worstFrameMs:e.worstFrameMs,jankCount:e.jankCount,frames:e.frames}),M({category:"minigame.framerate",message:`FPS: ${e.fps}（最低 ${e.minFps}，卡顿 ${e.jankCount} 次）`,level:e.fps<this._options.fpsWarningThreshold?"warning":"info",data:e}))}_rollupWindow(){const e=this._windowElapsed;if(this._frameCount===0||e<=0)return this._resetWindow(),null;const n=Math.round(this._frameCount/e*1e3),r=this._maxFrameDelta>0?Math.round(1e3/this._maxFrameDelta):n,s={fps:n,minFps:r,worstFrameMs:Math.round(this._maxFrameDelta),jankCount:this._jankCount,frames:this._frameCount};return this._fpsSamples.push(n),this._fpsSamples.length>G._MAX_FPS_SAMPLES&&this._fpsSamples.shift(),this._sessionElapsed+=e,this._sessionFrames+=this._frameCount,this._sessionJank+=this._jankCount,this._sessionJankByTier.minor+=this._jankByTier.minor,this._sessionJankByTier.major+=this._jankByTier.major,this._sessionJankByTier.severe+=this._jankByTier.severe,this._maxFrameDelta>this._sessionWorstFrame&&(this._sessionWorstFrame=this._maxFrameDelta),this._resetWindow(),s}_resetWindow(){this._windowElapsed=0,this._frameCount=0,this._jankCount=0,this._jankBreadcrumbs=0,this._maxFrameDelta=0,this._jankByTier={minor:0,major:0,severe:0}}_resetSession(){this._sessionEpochStart=Vn(),this._sessionElapsed=0,this._sessionFrames=0,this._sessionJank=0,this._sessionJankByTier={minor:0,major:0,severe:0},this._sessionWorstFrame=0,this._fpsSamples=[]}_restartOnResume(){this._resetSession();const e=Je();this._lastFrameTs=e,this._resetWindow()}_flushSummary(){if(this._rollupWindow(),this._sessionFrames===0)return!1;const e=this._sessionElapsed,n=e>0?Math.round(this._sessionFrames/e*1e3):0,r=this._sessionWorstFrame>0?Math.round(1e3/this._sessionWorstFrame):n,s=sl(this._fpsSamples,n),i=Math.round(this._sessionWorstFrame),o=this._sessionJank,c=this._sessionFrames,a=ft({name:"minigame.framerate.summary",op:"ui.framerate",forceTransaction:!0,startTime:this._sessionEpochStart/1e3});if(a.setAttributes({"fps.avg":n,"fps.p95":s,"fps.min":r,"frames.total":c,"jank.count":o,"frame.worst_ms":i}),Ie("fps_avg",n,"none",a),Ie("fps_p95",s,"none",a),Ie("fps_min",r,"none",a),Ie("jank_count",o,"none",a),this._tiered){const u={};for(const p of this._tiers){const f=this._sessionJankByTier[p.name];u[`jank.${p.name}`]=f,Ie(`jank_${p.name}_count`,f,"none",a)}a.setAttributes(u)}return a.end((this._sessionEpochStart+Math.max(0,e))/1e3),this._resetSession(),!0}_flushPendingEvents(){Promise.resolve(Es(G._HIDE_FLUSH_TIMEOUT_MS)).catch(()=>{})}cleanup(){this._running=!1,this._flushSummary()&&this._flushPendingEvents();const e=w();if(e)try{this._hideHandler&&typeof e.offHide=="function"&&e.offHide(this._hideHandler),this._showHandler&&typeof e.offShow=="function"&&e.offShow(this._showHandler)}catch(n){}this._hideHandler=null,this._showHandler=null}};G.id="MinigameFrameRate",G._MAX_FPS_SAMPLES=2e3,G._MAX_REASONABLE_FRAME_DELTA_MS=5e3,G._HIDE_FLUSH_TIMEOUT_MS=2e3;let Ce=G;const il=Object.freeze(Object.defineProperty({__proto__:null,ConsoleBreadcrumbs:Ve,Dedupe:It,GlobalHandlers:bt,HttpContext:vt,LinkedErrors:Tt,MinigameFrameRateIntegration:Ce,MinigameIntegration:Re,NetworkBreadcrumbs:At,NetworkStatusIntegration:Dt,PageBreadcrumbs:Ke,PerformanceIntegration:wt,RewriteFrames:kt,Router:nr,SessionIntegration:Nt,System:tr,TryCatch:Et,consoleBreadcrumbsIntegration:Zu,minigameFrameRateIntegration:t=>()=>new Ce(t),minigameIntegration:el,pageBreadcrumbsIntegration:Vu,performanceIntegration:rr},Symbol.toStringTag,{value:"Module"}));function ir(){return[new vt,new It,new bt,new Et,new Tt,rr({enableNavigation:!0,enableRender:!0,enableResource:!0,enableUserTiming:!0,sampleRate:1,reportInterval:3e4})]}const ol=ir();function al(t){return t===!1?[]:Array.isArray(t)?[...t]:ir()}function cl(t={}){if(!Zs()){console.warn("[sentry-miniapp] Not running in a supported miniapp environment");return}const e=[...t.integrations||al(t.defaultIntegrations)],n=I(l({},t),{defaultIntegrations:[],integrations:e,stackParser:Ks,transport:t.transport});n.enableSourceMap!==!1&&n.integrations.push(new kt);const r={traceNetworkBody:n.traceNetworkBody};n.enableTracePropagation!==void 0&&(r.enableTracePropagation=n.enableTracePropagation),n.tracePropagationTargets!==void 0&&(r.tracePropagationTargets=n.tracePropagationTargets),n.propagateTraceparent!==void 0&&(r.propagateTraceparent=n.propagateTraceparent),n.integrations.push(new At(r)),n.enableAutoSessionTracking!==!1&&n.integrations.push(new Nt);const s=n.enableNavigationBreadcrumbs!==!1,i=n.enableUserInteractionBreadcrumbs!==!1;(s||i)&&n.integrations.push(new Ke({enableLifecycle:s,enableUserInteraction:i})),n.enableNetworkStatusMonitoring!==!1&&n.integrations.push(new Dt),n.enableConsoleBreadcrumbs&&n.integrations.push(new Ve);const o=Su(),c=a=>n.integrations.some(u=>u&&u.name===a);if(!c("EventFilters")&&!c("InboundFilters")){const a={};n.allowUrls&&(a.allowUrls=n.allowUrls),n.denyUrls&&(a.denyUrls=n.denyUrls),n.ignoreErrors&&(a.ignoreErrors=n.ignoreErrors),n.integrations.push(Bc(a))}return!c(Re.id)&&(n.enableMinigameLifecycle===!0||o&&n.enableMinigameLifecycle!==!1)&&n.integrations.push(new Re),!c(Ce.id)&&(n.enableMinigameFrameRate===!0||o&&n.enableMinigameFrameRate!==!1)&&n.integrations.push(new Ce(n.minigameFrameRateOptions)),re("miniapp",{platform:ei(),environment:"miniapp"}),Cc(ii,n),T().getClient()}function ul(t={}){console.warn("[sentry-miniapp] showReportDialog is deprecated and does nothing. Please build your own UI and use `Sentry.captureFeedback()` instead.")}function ll(t){return function(...e){return ie(()=>{try{return t.apply(this,e)}catch(n){throw T().captureException(n),n}})}}function fl(t){var e,n,r;du(t),t&&((r=(n=(e=P())==null?void 0:e.getTransport())==null?void 0:n.flush)==null||r.call(n))}function dl(){return Vs()}function pl(t){const e=T().getClient();return e?e.captureFeedback(t):(console.warn("[sentry-miniapp] No client available for captureFeedback"),"")}U(),g.Integrations=il,g.MiniappClient=ii,g.SDK_NAME=Js,g.SDK_VERSION=gt,g.Transports=ku,g.addBreadcrumb=M,g.addEventProcessor=ja,g.addIntegration=Ga,g.captureEvent=Na,g.captureException=we,g.captureFeedback=pl,g.captureMessage=Pa,g.captureSession=Pn,g.close=$a,g.closeSession=Er,g.defaultIntegrations=ol,g.endSession=pt,g.flush=Es,g.getClient=P,g.getConsent=dl,g.getCurrentScope=T,g.getDefaultIntegrations=ir,g.getIsolationScope=F,g.getPerformanceManager=ti,g.init=cl,g.isEnabled=Ts,g.lastEventId=La,g.logger=su,g.makeSession=br,g.miniappStackParser=Ks,g.performanceIntegration=rr,g.setConsent=fl,g.setContext=re,g.setExtra=Oa,g.setExtras=Da,g.setTag=xa,g.setTags=Ma,g.setUser=Fa,g.showReportDialog=ul,g.startInactiveSpan=ft,g.startSession=vs,g.startSpan=He,g.updateSession=se,g.withScope=ie,g.wrap=ll,Object.defineProperty(g,Symbol.toStringTag,{value:"Module"})}));
+详情参考：https://docs.sentry.io/platforms/javascript/`):console.warn("[sentry-miniapp] 未检测到已支持的小程序平台，SDK 将以降级模式运行"),{request:()=>{},httpRequest:()=>{},getSystemInfoSync:()=>({}),URLSearchParams:()=>{}};const e=t.sdk;t.name==="alipay"&&!e.request&&e.httpRequest&&(e.request=e.httpRequest);const n=e;if((t.name==="alipay"||t.name==="dingtalk")&&!n.__sentryStorageAdapted){if(e.getStorageSync){const r=e.getStorageSync;e.getStorageSync=s=>{const i=r.call(e,{key:s});return i?i.data:null}}if(e.setStorageSync){const r=e.setStorageSync;e.setStorageSync=(s,i)=>{r.call(e,{key:s,data:i})}}if(e.removeStorageSync){const r=e.removeStorageSync;e.removeStorageSync=s=>{r.call(e,{key:s})}}n.__sentryStorageAdapted=!0}return e},Qs=()=>{var t,e;return(e=(t=Xs())==null?void 0:t.name)!=null?e:"unknown"},gu=()=>{try{const t=qn(),e={};let n=!1;if(t.getAppBaseInfo){const s=t.getAppBaseInfo();Object.assign(e,s),n=!0}if(t.getWindowInfo){const s=t.getWindowInfo();Object.assign(e,s),n=!0}if(t.getDeviceInfo){const s=t.getDeviceInfo();Object.assign(e,s),n=!0}if(t.getAppAuthorizeSetting){const s=t.getAppAuthorizeSetting();e.albumAuthorized=s.albumAuthorized==="authorized",e.cameraAuthorized=s.cameraAuthorized==="authorized",e.locationAuthorized=s.locationAuthorized==="authorized",e.microphoneAuthorized=s.microphoneAuthorized==="authorized",e.notificationAuthorized=s.notificationAuthorized==="authorized"}if(t.getSystemSetting){const s=t.getSystemSetting();Object.assign(e,s)}if(n&&!!(e.brand||e.model||e.system))return e;if(t.getSystemInfoSync){const s=t.getSystemInfoSync();return!s.SDKVersion&&s.version&&(s.SDKVersion=s.version),s}return n?e:null}catch(t){return console.warn("[sentry-miniapp] Failed to get system info:",t),null}};let Wn=null;const St=()=>(Wn===null&&(Wn=gu()),Wn),yu=()=>{var t,e;try{const n=qn();if(n.getAccountInfoSync){const r=n.getAccountInfoSync();return{appId:((t=r==null?void 0:r.miniProgram)==null?void 0:t.appId)||"unknown",version:((e=r==null?void 0:r.miniProgram)==null?void 0:e.version)||"unknown"}}}catch(n){}return{appId:"unknown",version:"unknown"}};let Gn=null;const Su=()=>{if(Gn===null){const t=yu();return(t.appId!=="unknown"||t.version!=="unknown")&&(Gn=t),t}return Gn},Zs=()=>Qs()!=="unknown";let Jn=null,Yn=null;const w=()=>(Jn===null&&(Jn=qn()),Jn),ei=()=>(Yn===null&&(Yn=Qs()),Yn);let Kn=null;const bu=()=>{if(Kn===null){const t=globalThis,e=typeof t.GameGlobal!="undefined",n=typeof t.App!="function"&&typeof t.Page!="function"&&typeof t.getCurrentPages!="function";Kn=Zs()&&(e||n)}return Kn},ti=()=>{try{const t=w();if(t.getPerformance&&typeof t.getPerformance=="function")return t.getPerformance()}catch(t){console.warn("Failed to get performance manager:",t)}return null},Je=()=>Date.now(),Vn=()=>Date.now(),ni="application/x-sentry-envelope";function ri(t){const e=t.url,n=t.headers||{};function r(s){return new Promise((i,o)=>{var a,u,p,f;const c={url:e,method:"POST",data:s.body,header:l(l({"Content-Type":ni},n),s.headers),headers:l(l({"Content-Type":ni},n),s.headers),timeout:1e4,success:_=>{const d=_.statusCode||_.status,h=_.header||_.headers||{};i({statusCode:d,headers:{"x-sentry-rate-limits":h["x-sentry-rate-limits"],"retry-after":h["retry-after"]}})},fail:_=>{o(new Error(`Network request failed: ${_.errMsg||_.errorMessage||_.message||"Unknown error"}`))}};w().request?(u=(a=w()).request)==null||u.call(a,c):w().httpRequest?(f=(p=w()).httpRequest)==null||f.call(p,c):o(new Error("No request method available in current miniapp environment"))})}return mc(t,r)}const Eu=30,Tu=864e5,Xn="sentry_offline_store",vu=900*1024;function Qn(t){var p,f,_,d;const e=(p=t.offlineCacheLimit)!=null?p:Eu,n=(f=t.offlineCacheMaxAge)!=null?f:Tu,r=(_=t.maxBytes)!=null?_:vu,s=(d=t.evictionMode)!=null?d:"error-priority",i=t.onDrop;function o(h,m){if(m>0&&typeof i=="function")try{i(h,m)}catch(S){}}function c(h){const m=Date.now(),S=h.filter(v=>m-v.timestamp<n);return o("age",h.length-S.length),S}function a(h,m){let S=0;for(;h.length>e;){if(s==="preserve-oldest")h.pop();else if(m==="push"){const v=h.findIndex(N=>!Zn(N.envelope));v>=0?h.splice(v,1):h.shift()}else{const v=Iu(h,N=>!Zn(N.envelope));v>=0?h.splice(v,1):h.pop()}S++}return S}function u(h){const m=w().setStorageSync;if(!m)return;let S=h,v=JSON.stringify(S),N=0;for(;S.length>0&&ku(v)>r;){if(s==="preserve-oldest")S=S.slice(0,S.length-1);else{const x=S.findIndex(k=>!Zn(k.envelope)),$=x>=0?x:0;S=S.slice(0,$).concat(S.slice($+1))}N++,v=JSON.stringify(S)}o("bytes",N);try{m(Xn,v)}catch(x){}}return{push:h=>z(null,null,function*(){try{const m=c(er());m.push({envelope:h,timestamp:Date.now()}),o("count",a(m,"push")),u(m)}catch(m){console.warn("[sentry-miniapp] Failed to push to offline store",m)}}),unshift:h=>z(null,null,function*(){try{const m=c(er());m.unshift({envelope:h,timestamp:Date.now()}),o("count",a(m,"unshift")),u(m)}catch(m){console.warn("[sentry-miniapp] Failed to unshift to offline store",m)}}),shift:()=>z(null,null,function*(){try{const h=er(),m=c(h);if(m.length===0){h.length>0&&u(m);return}const S=m.shift();return u(m),S==null?void 0:S.envelope}catch(h){console.warn("[sentry-miniapp] Failed to shift from offline store",h);return}})}}function Zn(t){try{return(t[1]||[]).some(n=>{var s;const r=(s=n[0])==null?void 0:s.type;return r==="event"||r==="error"})}catch(e){return!0}}function Iu(t,e){for(let n=t.length-1;n>=0;n--)if(e(t[n]))return n;return-1}function er(){const t=w().getStorageSync;if(!t)return[];let e;try{e=t(Xn)}catch(n){return[]}if(!e)return[];try{const n=typeof e=="string"?JSON.parse(e):e;return Array.isArray(n)&&n.length>0?n[0]&&"envelope"in n[0]&&"timestamp"in n[0]?n:n.map(r=>({envelope:r,timestamp:Date.now()})):[]}catch(n){return wu(),[]}}function wu(){try{const t=w().removeStorageSync;t&&t(Xn)}catch(t){}}function ku(t){let e=0;for(let n=0;n<t.length;n++){const r=t.charCodeAt(n);r<128?e+=1:r<2048?e+=2:r>=55296&&r<=56319?(e+=4,n++):e+=3}return e}const Au=Object.freeze(Object.defineProperty({__proto__:null,createMiniappOfflineStore:Qn,createMiniappTransport:ri},Symbol.toStringTag,{value:"Module"}));function Ru(t){return t!==null&&(typeof t=="object"||typeof t=="function")}function si(t,e,n){const r={};let s=!1;for(const i of e){const o=i[n];if(o)for(const[c,a]of Object.entries(o))r[c]===void 0&&typeof a=="string"&&a&&(r[c]=a,s=!0)}s&&(t[n]=r)}function Cu(){if(typeof globalThis=="undefined")return;const t=globalThis,e=hu().filter(Ru);si(t,e,"_sentryDebugIds"),si(t,e,"_debugIds")}class ii extends Ic{constructor(e={}){fu({required:e.requireConsent===!0,cacheLimit:e.consentCacheLimit,cacheMaxBytes:e.consentCacheMaxBytes,cacheMaxAge:e.consentCacheMaxAge,onDrop:e.onConsentCacheDrop}),super(I(l({},e),{transport:n=>{var i;const r=n,s=e.transport?e.transport(r):ri(I(l({},r),{headers:(i=r.headers)!=null?i:{}}));return e.requireConsent===!0?Ws(()=>s)(I(l({},n),{createStore:o=>{var c;return Qn(I(l({},o),{offlineCacheLimit:(c=e.consentCacheLimit)!=null?c:100,offlineCacheMaxAge:e.consentCacheMaxAge,maxBytes:e.consentCacheMaxBytes,evictionMode:"preserve-oldest",onDrop:pu}))},shouldSend:()=>Vs(),flushAtStartup:!0})):!e.transport&&e.enableOfflineCache!==!1?Ws(()=>s)(I(l({},n),{createStore:o=>Qn(I(l({},o),{offlineCacheLimit:e.offlineCacheLimit,offlineCacheMaxAge:e.offlineCacheMaxAge})),flushAtStartup:!0})):s}}))}eventFromException(e){const n={type:e.name||"Error",value:e.message||String(e)};if(e.stack){const r=this.getOptions().stackParser;if(r&&typeof r=="function"){const s=r(e.stack,1);s.length&&(n.stacktrace={frames:s})}}return Promise.resolve({exception:{values:[n]},level:"error"})}eventFromMessage(e,n="info"){return Promise.resolve({message:e,level:n})}_prepareEvent(e,n,r){e.platform=e.platform||this.getOptions().platform||"javascript",e.sdk=I(l({},e.sdk),{name:Js,packages:[...e.sdk&&e.sdk.packages||[],{name:"npm:@sentry/miniapp",version:gt}],version:gt}),e.contexts||(e.contexts={}),e.contexts.miniapp={platform:ei(),sdk_version:gt};try{if(this.getOptions().stackParser)try{Cu()}catch(o){this.getOptions().debug&&console.warn("[sentry-miniapp] Debug ID 全局同步失败:",o)}const s=r||T(),i=F();return Promise.resolve(super._prepareEvent(e,n||{},s,i)).then(o=>this._fillDefaultContexts(o))}catch(s){return this.getOptions().debug&&console.warn("[sentry-miniapp] _prepareEvent 兜底（scope 未就绪）:",s),Promise.resolve(this._fillDefaultContexts(e))}}_fillDefaultContexts(e){if(!e||this.getOptions().enableSystemInfo===!1)return e;const n=St();e.contexts||(e.contexts={});const r=e.contexts;return r.device=l({brand:(n==null?void 0:n.brand)||"unknown",model:(n==null?void 0:n.model)||"unknown",screen_resolution:`${(n==null?void 0:n.screenWidth)||0}x${(n==null?void 0:n.screenHeight)||0}`,language:(n==null?void 0:n.language)||"unknown",version:(n==null?void 0:n.version)||"unknown",system:(n==null?void 0:n.system)||"unknown",platform:(n==null?void 0:n.platform)||"unknown"},r.device),r.os=l({name:(n==null?void 0:n.system)||"unknown",version:(n==null?void 0:n.version)||"unknown"},r.os),r.app=l({app_version:(n==null?void 0:n.SDKVersion)||"unknown"},r.app),e}close(e){const n=this.getOptions().integrations;if(Array.isArray(n))for(const r of n){const s=r.cleanup;if(typeof s=="function")try{s.call(r)}catch(i){this.getOptions().debug&&console.warn(`[sentry-miniapp] 集成 ${r.name} cleanup 失败:`,i)}if(Array.isArray(We)){const i=We.indexOf(r.name);i!==-1&&We.splice(i,1)}}return super.close(e)}showReportDialog(e={}){console.warn("[sentry-miniapp] showReportDialog is deprecated and does nothing. Please build your own UI and use `Sentry.captureFeedback()` instead.")}captureFeedback(e){const n={contexts:{feedback:{contact_email:e.email,name:e.name,message:e.message,url:e.url,source:e.source,associated_event_id:e.associatedEventId}},type:"feedback",level:"info",tags:e.tags||{}};return T().captureEvent(n)}}const Dt=class Dt{constructor(e){this.name=Dt.id,this._onErrorHandlerInstalled=!1,this._onUnhandledRejectionHandlerInstalled=!1,this._onPageNotFoundHandlerInstalled=!1,this._onMemoryWarningHandlerInstalled=!1,this._errorHandler=null,this._rejectionHandler=null,this._pageNotFoundHandler=null,this._memoryWarningHandler=null,this._options=l({onerror:!0,onunhandledrejection:!0,onpagenotfound:!0,onmemorywarning:!0},e)}setupOnce(){Error.stackTraceLimit=50,this._options.onerror&&this._installGlobalOnErrorHandler(),this._options.onunhandledrejection&&this._installGlobalOnUnhandledRejectionHandler(),this._options.onpagenotfound&&this._installGlobalOnPageNotFoundHandler(),this._options.onmemorywarning&&this._installGlobalOnMemoryWarningHandler()}_installGlobalOnErrorHandler(){var e,n;this._onErrorHandlerInstalled||(w().onError&&(this._errorHandler=r=>{const s=typeof r=="string"?new Error(r):r;we(s,{mechanism:{type:"onerror",handled:!1}})},(n=(e=w()).onError)==null||n.call(e,this._errorHandler)),this._onErrorHandlerInstalled=!0)}_installGlobalOnUnhandledRejectionHandler(){var e,n;this._onUnhandledRejectionHandlerInstalled||(w().onUnhandledRejection&&(this._rejectionHandler=({reason:r,promise:s})=>{const i=typeof r=="string"?new Error(r):r;we(i,{mechanism:{type:"onunhandledrejection",handled:!1},data:{promise:s}})},(n=(e=w()).onUnhandledRejection)==null||n.call(e,this._rejectionHandler)),this._onUnhandledRejectionHandlerInstalled=!0)}_installGlobalOnPageNotFoundHandler(){var e,n;this._onPageNotFoundHandlerInstalled||(w().onPageNotFound&&(this._pageNotFoundHandler=r=>{const s=T(),i=r.path.split("?")[0];s.setTag("pagenotfound",i),s.setContext("page_not_found",{path:r.path,query:r.query,isEntryPage:r.isEntryPage}),we(new Error(`页面无法找到: ${i}`),{mechanism:{type:"onpagenotfound",handled:!0}})},(n=(e=w()).onPageNotFound)==null||n.call(e,this._pageNotFoundHandler)),this._onPageNotFoundHandlerInstalled=!0)}_installGlobalOnMemoryWarningHandler(){var e,n;this._onMemoryWarningHandlerInstalled||(w().onMemoryWarning&&(this._memoryWarningHandler=({level:r=-1})=>{let s="没有获取到告警级别信息";switch(r){case 5:s="TRIM_MEMORY_RUNNING_MODERATE";break;case 10:s="TRIM_MEMORY_RUNNING_LOW";break;case 15:s="TRIM_MEMORY_RUNNING_CRITICAL";break;default:return}const i=T();i.setTag("memory-warning",String(r)),i.setContext("memory_warning",{level:r,message:s}),we(new Error("内存不足告警"),{mechanism:{type:"onmemorywarning",handled:!0}})},(n=(e=w()).onMemoryWarning)==null||n.call(e,this._memoryWarningHandler)),this._onMemoryWarningHandlerInstalled=!0)}cleanup(){try{const e=w();this._errorHandler&&e.offError&&e.offError(this._errorHandler),this._rejectionHandler&&e.offUnhandledRejection&&e.offUnhandledRejection(this._rejectionHandler),this._pageNotFoundHandler&&e.offPageNotFound&&e.offPageNotFound(this._pageNotFoundHandler),this._memoryWarningHandler&&e.offMemoryWarning&&e.offMemoryWarning(this._memoryWarningHandler)}catch(e){}this._errorHandler=null,this._rejectionHandler=null,this._pageNotFoundHandler=null,this._memoryWarningHandler=null,this._onErrorHandlerInstalled=!1,this._onUnhandledRejectionHandlerInstalled=!1,this._onPageNotFoundHandlerInstalled=!1,this._onMemoryWarningHandlerInstalled=!1}};Dt.id="GlobalHandlers";let bt=Dt;function oi(t,e={},n){if(typeof t!="function")return t;try{if(t.__sentry__)return t;if(t.__sentry_wrapped__)return t.__sentry_wrapped__}catch(s){return t}const r=function(...s){try{return t.apply(this,s)}catch(i){throw ie(o=>{o.addEventProcessor(c=>I(l({},c),{extra:I(l({},c.extra),{arguments:s})})),we(i,e.mechanism?{mechanism:e.mechanism}:void 0)}),i}};try{for(const s in t)Object.prototype.hasOwnProperty.call(t,s)&&(r[s]=t[s])}catch(s){}t.prototype=t.prototype||{},r.prototype=t.prototype,Object.defineProperty(t,"__sentry_wrapped__",{enumerable:!1,value:r}),Object.defineProperties(r,{__sentry__:{enumerable:!1,value:!0},__sentry_original__:{enumerable:!1,value:t}});try{Object.getOwnPropertyDescriptor(r,"name").configurable&&Object.defineProperty(r,"name",{get(){return t.name}})}catch(s){}return r}function ai(t){try{return t&&t.name||"<anonymous>"}catch(e){return"<anonymous>"}}function Ye(t,e,n){if(!(e in t))return;const r=t[e],s=n(r);if(typeof s=="function")try{s.prototype=s.prototype||{},s.prototype.constructor=s}catch(i){}t[e]=s}const Mt=class Mt{constructor(){this.name=Mt.id}_wrapTimeFunction(e){return function(...n){const r=n[0];return n[0]=oi(r,{mechanism:{data:{function:ai(e)},handled:!0,type:"instrument"}}),e.apply(this,n)}}_wrapRAF(e){return function(n){return e(oi(n,{mechanism:{data:{function:"requestAnimationFrame",handler:ai(e)},handled:!0,type:"instrument"}}))}}setupOnce(){const e=globalThis;e.setTimeout&&Ye(e,"setTimeout",this._wrapTimeFunction.bind(this)),e.setInterval&&Ye(e,"setInterval",this._wrapTimeFunction.bind(this)),e.requestAnimationFrame&&Ye(e,"requestAnimationFrame",this._wrapRAF.bind(this))}};Mt.id="TryCatch";let Et=Mt;const Pu="cause",Nu=5,xt=class xt{constructor(e={}){this.name=xt.id,this._key=e.key||Pu,this._limit=e.limit||Nu}setupOnce(){}processEvent(e,n){return T().getClient()?this._handler(e,n):e}_handler(e,n){if(!e.exception||!e.exception.values||!n||!ci(n.originalException,Error))return e;const r=this._walkErrorTree(n.originalException,this._key);return e.exception.values=[...r,...e.exception.values],e}_walkErrorTree(e,n,r=[]){if(!ci(e[n],Error)||r.length+1>=this._limit)return r;const s=Xc(()=>[],e[n]);return this._walkErrorTree(e[n],n,[s,...r])}};xt.id="LinkedErrors";let Tt=xt;function ci(t,e){try{return t instanceof e}catch(n){return!1}}const Ft=class Ft{constructor(){this.name=Ft.id}setupOnce(){}processEvent(e){var s,i;const n=this._getMiniappVersion(),r=Su();return e.contexts=I(l({},e.contexts),{runtime:I(l({},((s=e.contexts)==null?void 0:s.runtime)||{}),{name:"miniapp",version:n}),app:I(l({},((i=e.contexts)==null?void 0:i.app)||{}),{name:r.appId,version:r.version})}),e}_getMiniappVersion(){const e=St();return(e==null?void 0:e.version)||(e==null?void 0:e.SDKVersion)||"unknown"}};Ft.id="HttpContext";let vt=Ft;const Lt=class Lt{constructor(e={}){this.name=Lt.id,this._fuzzyMatch=!!e.fuzzyMatch}setupOnce(){}processEvent(e,n){if(e.type)return e;try{if(this._shouldDropEvent(e,this._previousEvent))return null}catch(r){return this._previousEvent=this._cloneEventForComparison(e),e}return this._previousEvent=this._cloneEventForComparison(e),e}_shouldDropEvent(e,n){return n?!!(this._isSameMessageEvent(e,n)||this._isSameExceptionEvent(e,n)):!1}_isSameMessageEvent(e,n){const r=e.message,s=n.message;return!(!r&&!s||r&&!s||!r&&s||r!==s||!this._isSameFingerprint(e,n)||!this._isSameStacktrace(e,n))}_isSameExceptionEvent(e,n){const r=this._getExceptionFromEvent(e),s=this._getExceptionFromEvent(n);if(!r||!s||r.type!==s.type)return!1;if(this._fuzzyMatch){if(!Ou(r.value,s.value))return!1}else if(r.value!==s.value)return!1;return!(!this._isSameFingerprint(e,n)||!this._isSameStacktrace(e,n))}_isSameStacktrace(e,n){let r=this._getFramesFromEvent(e),s=this._getFramesFromEvent(n);if(!r&&!s)return!0;if(r&&!s||!r&&s||(r=r,s=s,s.length!==r.length))return!1;for(let i=0;i<s.length;i++){const o=s[i],c=r[i];if(o.filename!==c.filename||o.lineno!==c.lineno||o.colno!==c.colno||o.function!==c.function)return!1}return!0}_isSameFingerprint(e,n){let r=e.fingerprint,s=n.fingerprint;if(!r&&!s)return!0;if(r&&!s||!r&&s)return!1;r=r,s=s;try{return r.join("")===s.join("")}catch(i){return!1}}_getExceptionFromEvent(e){return e.exception&&e.exception.values&&e.exception.values[0]}_cloneEventForComparison(e){var n;return{message:e.message,fingerprint:e.fingerprint?[...e.fingerprint]:void 0,exception:e.exception?{values:(n=e.exception.values)==null?void 0:n.map(r=>{var s;return{type:r.type,value:r.value,stacktrace:r.stacktrace?{frames:(s=r.stacktrace.frames)==null?void 0:s.map(i=>l({},i))}:void 0}})}:void 0}}_getFramesFromEvent(e){const n=e.exception;if(n)try{return n.values[0].stacktrace.frames}catch(r){}}};Lt.id="Dedupe";let It=Lt;const ui=/\d{10,}|[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}|\d{4}-\d{2}-\d{2}T\d{2}:\d{2}/g;function Ou(t,e){if(t===e)return!0;if(!t||!e)return!1;const n=t.replace(ui,"<dynamic>"),r=e.replace(ui,"<dynamic>");return n===r}const $t=class $t{constructor(){this.name=$t.id}setupOnce(){this._addSystemContext(),this._addNetworkContext(),this._addLocationContext(),this._addStorageContext(),this._checkAppUpdate()}_addSystemContext(){var e,n,r,s;try{const i=St();if(i){const o=T();o.setContext("device",{name:i.model,model:i.model,brand:i.brand,family:i.platform,arch:i.platform}),o.setContext("os",{name:((e=i.system)==null?void 0:e.split(" ")[0])||"unknown",version:((n=i.system)==null?void 0:n.split(" ")[1])||"unknown",kernel_version:i.version}),o.setContext("app",{app_name:i.appName||"unknown",app_version:i.version}),o.setContext("screen",{screen_width:i.screenWidth,screen_height:i.screenHeight,screen_density:i.pixelRatio}),o.setTag("device.model",i.model),o.setTag("device.brand",i.brand),o.setTag("os.name",((r=i.system)==null?void 0:r.split(" ")[0])||"unknown"),o.setTag("os.version",((s=i.system)==null?void 0:s.split(" ")[1])||"unknown"),o.setTag("app.version",i.version),o.setTag("language",i.language)}}catch(i){}}_addNetworkContext(){try{w().getNetworkType&&w().getNetworkType({success:e=>{const n=T();n.setContext("network",{type:e.networkType,connected:e.isConnected!==!1}),n.setTag("network.type",e.networkType)},fail:()=>{}})}catch(e){}}_addLocationContext(){try{w().getLocation&&w().getLocation({type:"gcj02",success:e=>{T().setContext("location",{latitude:e.latitude,longitude:e.longitude,accuracy:e.accuracy})},fail:()=>{}})}catch(e){}}_addStorageContext(){try{const e=w();if(e&&typeof e.getStorageInfoSync=="function"){const n=e.getStorageInfoSync();if(n){const r=n.currentSize||0,s=n.limitSize||0,i=s>0?Math.round(r/s*100):0;T().setContext("storage",{currentSize:r,limitSize:s,usagePercent:i}),i>=80&&M({category:"storage.warning",message:`存储使用率 ${i}%（${r}KB / ${s}KB）`,level:"warning",data:{currentSize:r,limitSize:s,usagePercent:i}})}}}catch(e){}}_checkAppUpdate(){try{const e=w();if(e&&typeof e.getUpdateManager=="function"){const n=e.getUpdateManager();n&&(typeof n.onCheckForUpdate=="function"&&n.onCheckForUpdate(r=>{r&&r.hasUpdate&&(T().setTag("has_update","true"),M({category:"app.update",message:"检测到新版本可用",level:"info",data:{hasUpdate:!0}}))}),typeof n.onUpdateReady=="function"&&n.onUpdateReady(()=>{M({category:"app.update",message:"新版本已下载完成",level:"info",data:{updateReady:!0}})}))}}catch(e){}}};$t.id="System";let tr=$t;const jt=class jt{constructor(){this.name=jt.id,this._lastRoute="",this._monitorTimer=null}setupOnce(){this._instrumentNavigation(),this._startRouteMonitoring()}_instrumentNavigation(){let e;try{e=w()}catch(r){return}const n=["navigateTo","redirectTo","switchTab","reLaunch"];for(const r of n)if(e[r]){const s=e[r];e[r]=i=>(this._recordNavigation(r,i.url,this._getCurrentRoute()),s.call(e,i))}if(e.navigateBack){const r=e.navigateBack;e.navigateBack=(s={})=>(this._recordNavigation("navigateBack","back",this._getCurrentRoute(),s.delta),r.call(e,s))}}_startRouteMonitoring(){this._monitorTimer=setInterval(()=>{const e=this._getCurrentRoute();e&&e!==this._lastRoute&&(this._recordRouteChange(this._lastRoute,e),this._lastRoute=e)},1e3)}cleanup(){this._monitorTimer&&(clearInterval(this._monitorTimer),this._monitorTimer=null)}_getCurrentRoute(){try{const e=globalThis;if(e.getCurrentPages){const n=e.getCurrentPages();if(n&&n.length>0){const r=n[n.length-1];return r.route||r.__route__||""}}}catch(e){}return""}_recordNavigation(e,n,r,s){const i=T();M({category:"navigation",data:{action:e,from:r,to:n,delta:s},message:`Navigation ${e}: ${r} -> ${n}`,type:"navigation"}),i.setTag("route",n==="back"?r:n),i.setContext("navigation",{action:e,from:r,to:n,delta:s,timestamp:Date.now()})}_recordRouteChange(e,n){const r=T();M({category:"navigation",data:{from:e,to:n},message:`Route changed: ${e} -> ${n}`,type:"navigation"}),r.setTag("route",n),r.setContext("route",{current:n,previous:e,timestamp:Date.now()})}};jt.id="Router";let nr=jt;const Bt=class Bt{constructor(e={}){this.name=Bt.id,this._performanceManager=null,this._observers=[],this._entryBuffer=[],this._reportTimer=null,this._options=I(l({enableNavigation:!0,enableRender:!0,enableResource:!0,enableUserTiming:!1,sampleRate:1,bufferSize:100,reportInterval:3e4,enableMemory:!1},e),{thresholds:l({navigation:3e3,render:1e3,resource:2e3,setData:50},e==null?void 0:e.thresholds)})}setupOnce(){this._initializePerformanceManager(),this._setupPerformanceObservers(),this._startAutoReporting(),this._addPerformanceContext()}_initializePerformanceManager(){try{if(this._performanceManager=ti(),!this._performanceManager){console.warn("[sentry-miniapp] Performance API not available");return}const e=T();e.setTag("performance.api.available",!0),e.setContext("performance",{api_version:"miniapp-1.0",sample_rate:this._options.sampleRate,buffer_size:this._options.bufferSize})}catch(e){console.warn("[sentry-miniapp] Failed to initialize performance manager:",e)}}_setupPerformanceObservers(){if(this._performanceManager)try{const e=[];this._options.enableNavigation&&e.push("navigation"),this._options.enableRender&&e.push("render"),this._options.enableResource&&e.push("resource");let n=this._options.enableUserTiming;if(n)try{const i=St();if(i&&i.platform==="devtools"&&(n=!1),n){const c=(typeof globalThis!="undefined"?globalThis:typeof window!="undefined"?window:{}).PerformanceObserver,a=c&&Array.isArray(c.supportedEntryTypes)?c.supportedEntryTypes:void 0;(!a||!a.includes("measure")||!a.includes("mark"))&&(n=!1)}}catch(i){n=!1}if(n&&e.push("measure","mark"),e.length===0)return;const r=this._performanceManager.createObserver(i=>{this._handlePerformanceEntries(i)});try{r.observe({entryTypes:e})}catch(i){const o=e.filter(c=>c!=="measure"&&c!=="mark");if(o.length<e.length&&o.length>0)r.observe({entryTypes:o}),console.warn("[sentry-miniapp] Failed to observe all types, falling back to:",o),e.length=0,e.push(...o);else throw i}this._observers.push(r);const s=typeof globalThis!="undefined"?globalThis.process:void 0;s&&s.env&&s.env.NODE_ENV!=="production"&&console.log("[sentry-miniapp] Performance observers setup for:",e)}catch(e){console.warn("[sentry-miniapp] Failed to setup performance observers:",e)}}_handlePerformanceEntries(e){if(!e)return;let n;if(Array.isArray(e))n=e;else if(typeof e=="object"&&e.getEntries)n=e.getEntries();else if(typeof e=="object")n=[e];else{console.warn("[sentry-miniapp] Invalid entries format:",typeof e);return}n.length!==0&&(Math.random()>this._options.sampleRate||n.forEach(r=>{try{this._processPerformanceEntry(r),this._addToBuffer(r)}catch(s){console.warn("[sentry-miniapp] Failed to process performance entry:",s)}}))}_processPerformanceEntry(e){switch(e.entryType){case"navigation":this._processNavigationEntry(e);break;case"render":this._processRenderEntry(e);break;case"resource":this._processResourceEntry(e);break;case"measure":case"mark":this._processUserTimingEntry(e);break;default:console.log("[sentry-miniapp] Unknown entry type:",e.entryType)}}_processNavigationEntry(e){T().addBreadcrumb({message:`页面导航: ${e.name}`,category:"performance.navigation",level:"info",data:{duration:e.duration,appLaunchTime:e.appLaunchTime,pageReadyTime:e.pageReadyTime}}),He({name:`Navigation: ${e.name}`,op:"navigation",startTime:e.startTime/1e3},r=>{r.setAttributes({"navigation.name":e.name,"navigation.duration":e.duration,"navigation.app_launch_time":e.appLaunchTime||0,"navigation.page_ready_time":e.pageReadyTime||0,"navigation.first_render_time":e.firstRenderTime||0}),r.end((e.startTime+e.duration)/1e3)})}_processRenderEntry(e){var r,s;He({name:`Render: ${e.name}`,op:"render",startTime:e.startTime/1e3},i=>{i.setAttributes({"render.name":e.name,"render.duration":e.duration,"render.start":e.renderStart||0,"render.end":e.renderEnd||0,"render.script_start":e.scriptStart||0,"render.script_end":e.scriptEnd||0}),i.end((e.startTime+e.duration)/1e3)});const n=(s=(r=this._options.thresholds)==null?void 0:r.setData)!=null?s:50;e.duration>n&&T().addBreadcrumb({message:`慢渲染检测: ${e.name} (${e.duration.toFixed(1)}ms)`,category:"performance.setData.slow",level:"warning",data:{name:e.name,duration:e.duration,threshold:n,renderStart:e.renderStart,renderEnd:e.renderEnd,scriptStart:e.scriptStart,scriptEnd:e.scriptEnd}})}_processResourceEntry(e){He({name:`Resource: ${e.name}`,op:"resource",startTime:e.startTime/1e3},n=>{n.setAttributes({"resource.name":e.name,"resource.duration":e.duration,"resource.type":e.initiatorType||"unknown","resource.transfer_size":e.transferSize||0,"resource.encoded_size":e.encodedBodySize||0,"resource.decoded_size":e.decodedBodySize||0}),e.fetchStart&&e.responseEnd&&n.setAttributes({"resource.fetch_start":e.fetchStart,"resource.response_end":e.responseEnd,"resource.network_time":e.responseEnd-e.fetchStart}),n.end((e.startTime+e.duration)/1e3)})}_processUserTimingEntry(e){e.entryType==="measure"?He({name:`Measure: ${e.name}`,op:"measure",startTime:e.startTime/1e3},n=>{n.setAttributes({"measure.name":e.name,"measure.duration":e.duration,"measure.detail":e.detail?JSON.stringify(e.detail):void 0}),n.end((e.startTime+e.duration)/1e3)}):e.entryType==="mark"&&T().addBreadcrumb({message:`性能标记: ${e.name}`,category:"performance.mark",level:"info",data:{timestamp:e.startTime,detail:e.detail}})}_addToBuffer(e){for(this._entryBuffer.push(e);this._entryBuffer.length>this._options.bufferSize;)this._entryBuffer.shift()}_startAutoReporting(){this._options.reportInterval<=0||(this._reportTimer=setInterval(()=>{this._reportBufferedEntries()},this._options.reportInterval))}_reportBufferedEntries(){if(this._entryBuffer.length!==0)try{const e=T(),n=this._calculatePerformanceStats(),r=this._collectMemoryInfo();r&&(n.memory=r),e.setContext("performance_summary",l({total_entries:this._entryBuffer.length,navigation_count:this._entryBuffer.filter(s=>s.entryType==="navigation").length,render_count:this._entryBuffer.filter(s=>s.entryType==="render").length,resource_count:this._entryBuffer.filter(s=>s.entryType==="resource").length,measure_count:this._entryBuffer.filter(s=>s.entryType==="measure").length,mark_count:this._entryBuffer.filter(s=>s.entryType==="mark").length,report_time:new Date().toISOString()},n)),this._checkPerformanceThresholds(n),this._reportToNativeAPI(),this._entryBuffer=[]}catch(e){console.warn("[sentry-miniapp] Failed to report buffered entries:",e)}}_calculatePerformanceStats(){const e=this._entryBuffer.filter(i=>i.entryType==="navigation"),n=this._entryBuffer.filter(i=>i.entryType==="render"),r=this._entryBuffer.filter(i=>i.entryType==="resource"),s={};if(e.length>0){const i=e.map(o=>o.duration);s.navigation_stats={avg_duration:i.reduce((o,c)=>o+c,0)/i.length,max_duration:Math.max(...i),min_duration:Math.min(...i)}}if(n.length>0){const i=n.map(c=>c.duration);s.render_stats={avg_duration:i.reduce((c,a)=>c+a,0)/i.length,max_duration:Math.max(...i),min_duration:Math.min(...i)};const o=n.filter(c=>{var a,u;return c.duration>((u=(a=this._options.thresholds)==null?void 0:a.setData)!=null?u:50)});s.render_stats.slow_render_count=o.length,s.render_stats.slow_render_ratio=o.length/n.length}if(r.length>0){const i=r.map(c=>c.duration),o=r.map(c=>c.transferSize||0).filter(c=>c>0);s.resource_stats={avg_load_time:i.reduce((c,a)=>c+a,0)/i.length,max_load_time:Math.max(...i),total_transfer_size:o.reduce((c,a)=>c+a,0),avg_transfer_size:o.length>0?o.reduce((c,a)=>c+a,0)/o.length:0}}return s}_checkPerformanceThresholds(e){var o,c,a,u,p,f,_,d,h;const n=T(),r=(c=(o=this._options.thresholds)==null?void 0:o.navigation)!=null?c:3e3;((a=e.navigation_stats)==null?void 0:a.avg_duration)>r&&n.addBreadcrumb({message:"页面导航性能较慢",category:"performance.warning",level:"warning",data:{avg_duration:e.navigation_stats.avg_duration,threshold:r}});const s=(p=(u=this._options.thresholds)==null?void 0:u.render)!=null?p:1e3;((f=e.render_stats)==null?void 0:f.avg_duration)>s&&n.addBreadcrumb({message:"页面渲染性能较慢",category:"performance.warning",level:"warning",data:{avg_duration:e.render_stats.avg_duration,threshold:s}});const i=(d=(_=this._options.thresholds)==null?void 0:_.resource)!=null?d:2e3;((h=e.resource_stats)==null?void 0:h.avg_load_time)>i&&n.addBreadcrumb({message:"资源加载性能较慢",category:"performance.warning",level:"warning",data:{avg_load_time:e.resource_stats.avg_load_time,threshold:i}})}_reportToNativeAPI(){try{const e=w();if(e.reportPerformance&&this._entryBuffer.length>0){const n={entries:this._entryBuffer.map(r=>({name:r.name,entryType:r.entryType,startTime:r.startTime,duration:r.duration})),timestamp:Date.now(),sampleRate:this._options.sampleRate};e.reportPerformance(n)}}catch(e){console.warn("[sentry-miniapp] Failed to report to native API:",e)}}_collectMemoryInfo(){if(!this._options.enableMemory)return null;try{const e=w();if(e.getPerformance){const n=e.getPerformance();if(n&&n.memory)return{jsHeapSizeUsed:n.memory.jsHeapSizeUsed,jsHeapSizeLimit:n.memory.jsHeapSizeLimit}}}catch(e){}return null}_addPerformanceContext(){try{const e=T(),n=w(),r=!!n.getPerformance,s=!!n.reportPerformance;e.setContext("performance_support",{has_performance_api:r,has_report_api:s,integration_enabled:!0,options:this._options}),e.setTag("performance.integration","enabled")}catch(e){console.warn("[sentry-miniapp] Failed to add performance context:",e)}}cleanup(){this._observers.forEach(e=>{try{e.disconnect()}catch(n){console.warn("[sentry-miniapp] Failed to disconnect observer:",n)}}),this._observers=[],this._reportTimer&&(clearInterval(this._reportTimer),this._reportTimer=null),this._reportBufferedEntries()}};Bt.id="PerformanceAPI";let wt=Bt;const rr=t=>()=>new wt(t),Ut=class Ut{constructor(e={}){this.name=Ut.id,this._prefix=e.prefix||"app:///"}setupOnce(){}processEvent(e){return e.exception&&e.exception.values&&e.exception.values.forEach(n=>{n.stacktrace&&n.stacktrace.frames&&n.stacktrace.frames.forEach(r=>{r.filename&&(r.filename=this._normalizeFilename(r.filename))})}),e}_normalizeFilename(e){let n=e;return n.startsWith(this._prefix)?n:(n=n.replace(/^(appservice|app-service|WAService)\//i,"").replace(/^https?:\/\/[^/]+\//i,"").replace(/^chunks:\/\/\/?/i,"chunks/").replace(/^[a-z]+:\/\//i,"").replace(/^\//,""),`${this._prefix}${n}`)}};Ut.id="RewriteFrames";let kt=Ut;const Ht=class Ht{constructor(e={}){this.name=Ht.id,this._originalRequest=null,this._originalHttpRequest=null,this._traceNetworkBody=!!e.traceNetworkBody,this._sensitiveKeys=(e.sensitiveKeys||["password","passwd","secret","token","access_token","refresh_token","authorization","cookie","session","creditcard","credit_card","card_number","cvv","ssn","id_card"]).map(n=>n.toLowerCase()),this._denyUrls=(e.denyBodyUrls||[]).map(n=>typeof n=="string"?new RegExp(n):n),this._enableTracePropagation=e.enableTracePropagation!==!1,this._tracePropagationTargets=e.tracePropagationTargets||[],this._propagateTraceparent=e.propagateTraceparent===!0}setupOnce(){const e=w();e&&typeof e.request=="function"&&(this._originalRequest=e.request,Ye(e,"request",this._createRequestWrapper.bind(this))),e&&typeof e.httpRequest=="function"&&(this._originalHttpRequest=e.httpRequest,Ye(e,"httpRequest",this._createRequestWrapper.bind(this)))}cleanup(){const e=w();e&&(this._originalRequest&&typeof e.request=="function"&&(e.request=this._originalRequest),this._originalHttpRequest&&typeof e.httpRequest=="function"&&(e.httpRequest=this._originalHttpRequest)),this._originalRequest=null,this._originalHttpRequest=null}_createRequestWrapper(e){const n=this._traceNetworkBody,r=this._sanitizeBody.bind(this),s=this._shouldDenyBodyUrl.bind(this),i=this._enableTracePropagation,o=this._shouldPropagateTrace.bind(this),c=this._propagateTraceparent;return function(a){if(!a||typeof a!="object")return e.call(this,a);const u=ju(a.url),p=P();let f="";if(p){const A=p.getOptions().dsn;if(A)try{const L=A.match(/^(?:https?:\/\/)?(?:[^@\n]+@)?([^:/\n]+)/i);L&&L[1]&&(f=L[1])}catch(L){}}const _=u.match(/^https?:\/\/([^:/\n]+)/i),d=_&&_[1]?_[1]:"";if(f&&d===f||d==="sentry.io"||d.endsWith(".sentry.io"))return e.call(this,a);const m=Bu(a.method),S=a.data,v=Date.now(),N=Du(m,u);let x=!1;const $=A=>{x||(x=!0,$u(N,A))};i&&o(u)&&Lu(a,N,c);const k={url:u,method:m};if(n&&S&&!s(u))try{const A=typeof S=="string"?S:JSON.stringify(S);k.request_body=r(A),k.request_size=A.length}catch(A){k.request_body="[Cannot serialize request body]"}const J=a.success,ce=a.fail,Yt=a.complete;a.success=function(...A){const L=A[0]||{},H=li(L),ue=Date.now()-v;if(k.status_code=H,k.duration=ue,$({statusCode:H,status:sr(H)?"error":"ok"}),n&&L.data&&!s(u))try{const Pe=typeof L.data=="string"?L.data:JSON.stringify(L.data);k.response_body=r(Pe),k.response_size=Pe.length}catch(Pe){k.response_body="[Cannot serialize response body]"}const Kt=sr(H)||ue>3e3?"warning":"info";if(M({type:"http",category:"xhr",data:k,level:Kt}),typeof J=="function")return J.apply(this,A)},a.fail=function(...A){const L=A[0]||{},H=Date.now()-v,ue=L.errMsg||L.errorMessage||"Network request failed";if(k.error=ue,k.duration=H,$({status:"error",errorMessage:ue}),M({type:"http",category:"xhr",data:k,level:"error"}),typeof ce=="function")return ce.apply(this,A)},a.complete=function(...A){const L=A[0]||{},H=li(L);if($({statusCode:H,status:sr(H)?"error":"ok"}),typeof Yt=="function")return Yt.apply(this,A)};try{return e.call(this,a)}catch(A){throw $({status:"error",errorMessage:A instanceof Error?A.message:String(A)}),A}}}_shouldPropagateTrace(e){return this._tracePropagationTargets.length===0?!0:this._tracePropagationTargets.some(n=>{if(typeof n=="string")return e.includes(n);n.lastIndex=0;const r=n.test(e);return n.lastIndex=0,r})}_shouldDenyBodyUrl(e){return this._denyUrls.some(n=>n.test(e))}_sanitizeBody(e){if(this._sensitiveKeys.length===0)return e;try{const n=JSON.parse(e);if(typeof n=="object"&&n!==null)return this._sanitizeObject(n),JSON.stringify(n)}catch(n){for(const r of this._sensitiveKeys){const s=new RegExp(`(${r})=[^&]*`,"gi");e=e.replace(s,"$1=[Filtered]")}}return e}_sanitizeObject(e){for(const n of Object.keys(e))this._sensitiveKeys.includes(n.toLowerCase())?e[n]="[Filtered]":typeof e[n]=="object"&&e[n]!==null&&this._sanitizeObject(e[n])}};Ht.id="NetworkBreadcrumbs";let At=Ht;function Du(t,e){try{const n=zu(e);return ft({name:`${t} ${Mu(e)}`,op:"http.client",kind:2,attributes:{"http.request.method":t,"url.full":e,"server.address":n||void 0}})}catch(n){return null}}function Mu(t){return t.startsWith("data:")?Fu(t):xu(t).replace(/^([a-z][a-z0-9+.-]*:\/\/)([^/?#@]+@)/i,"$1[filtered]:[filtered]@")}function xu(t){const e=t.split(/[?#]/,1)[0];return e===void 0?t:e}function Fu(t){const e=t.match(/^data:([^;,]+)/);return`data:${e&&e[1]?e[1]:"text/plain"}`}function Lu(t,e,n){if(e)try{const r=Mc(n?{span:e,propagateTraceparent:!0}:{span:e}),s=r["sentry-trace"];if(!s)return;const i=l(l({},fi(t.headers)?t.headers:{}),fi(t.header)?t.header:{});i["sentry-trace"]||(i["sentry-trace"]=s),r.baggage&&(i.baggage=Hu(i.baggage,r.baggage)),n&&r.traceparent&&!Uu(i,"traceparent")&&(i.traceparent=r.traceparent),t.header=i,t.headers=i}catch(r){}}function $u(t,e){if(t)try{const n=di(e.statusCode);n!==void 0?co(t,n):t.setStatus({code:e.status==="error"?j:it,message:e.status==="error"?e.errorMessage||"error":"ok"}),e.errorMessage&&t.setAttribute("error.message",e.errorMessage),t.end()}catch(n){}}function ju(t){return typeof t=="string"?t:t==null?"":String(t)}function Bu(t){return typeof t=="string"&&t.trim()!==""?t.toUpperCase():"GET"}function li(t){if(!(!t||typeof t!="object"))return t.statusCode!==void 0&&t.statusCode!==null?t.statusCode:t.status}function sr(t){const e=di(t);return e!==void 0&&e>=400}function fi(t){return typeof t=="object"&&t!==null&&!Array.isArray(t)}function Uu(t,e){const n=e.toLowerCase();return Object.keys(t).some(r=>r.toLowerCase()===n)}function Hu(t,e){const n=typeof t=="string"?t:Array.isArray(t)?t.filter(s=>typeof s=="string").join(","):"";return n?n.split(",").some(s=>s.trim().startsWith("sentry-"))?n:`${n},${e}`:e}function di(t){if(typeof t=="number"&&Number.isFinite(t))return t;if(typeof t=="string"&&t.trim()!==""){const e=Number(t);if(Number.isFinite(e))return e}}function zu(t){try{const e=t.match(/^https?:\/\/([^:/\n]+)/i);return e&&e[1]?e[1]:""}catch(e){return""}}const qu=["onLaunch","onShow","onHide","onError"],Rt=new Set;let Ct=null,Pt=!1;function Wu(t,e){for(const n of[...Rt]){const r=n[t];if(typeof r=="function")try{r(e)}catch(s){}}}function Gu(){const t=globalThis;if(Pt||typeof t.App!="function")return;const e=t.App;Ct=e;const n=function(r={}){if(r&&typeof r=="object")for(const s of qu){const i=r[s];r[s]=function(...o){if(Wu(s,o[0]),typeof i=="function")return i.apply(this,o)}}return e.call(this,r)};n.__sentryAppWrapper=!0,t.App=n,Pt=!0}function Ju(){if(!Pt||Rt.size>0)return;const t=globalThis;t.App&&t.App.__sentryAppWrapper&&Ct&&(t.App=Ct),Ct=null,Pt=!1}function pi(t){if(typeof globalThis.App!="function")return()=>{};Rt.add(t),Gu();let e=!0;return()=>{e&&(e=!1,Rt.delete(t),Ju())}}const hi=["onLoad","onShow","onHide","onUnload","onReady"];function Yu(t){return hi.includes(t)||t.startsWith("_")?!1:/^(on|handle|bind)[A-Z]/.test(t)||/[Tt]ap$/.test(t)||/[Cc]lick$/.test(t)||/[Cc]hange$/.test(t)||/[Ss]ubmit$/.test(t)||/[Ss]croll$/.test(t)||/[Ii]nput$/.test(t)}const zt=class zt{constructor(e={}){this.name=zt.id,this._originalPage=null,this._wrappedPage=null,this._unsubscribeApp=null,this._launchTime=0,this._firstPageReady=!1,this._options=l({enableLifecycle:!0,enableUserInteraction:!0},e)}setupOnce(){this._wrapPage(),this._subscribeApp()}_wrapPage(){const e=globalThis;if(typeof e.Page!="function"||e.Page.__sentryPageWrapper)return;this._originalPage=e.Page;const n=e.Page,r=this._options,s=()=>this._firstPageReady,i=a=>{this._firstPageReady=a},o=()=>this._launchTime,c=function(a){if(a&&typeof a=="object"){if(r.enableLifecycle){for(const u of hi)if(typeof a[u]=="function"){const p=a[u];a[u]=function(...f){const _=(this==null?void 0:this.route)||(this==null?void 0:this.__route__)||"unknown",d={action:u,page:_};if(u==="onLoad"&&f[0]&&typeof f[0]=="object"&&(d.query=f[0]),u==="onReady"&&!s()&&o()>0){i(!0);const h=Date.now()-o();d.coldStartDuration=h,re("startup",{coldStartDuration:h,firstPage:_})}return M({category:"page.lifecycle",message:`${u}: ${_}`,level:"info",data:d}),p.apply(this,f)}}}if(r.enableUserInteraction){for(const u of Object.keys(a))if(typeof a[u]=="function"&&Yu(u)){const p=a[u];a[u]=function(f,..._){const d=(this==null?void 0:this.route)||(this==null?void 0:this.__route__)||"unknown",h={handler:u,page:d};if(f&&typeof f=="object"&&(f.target&&(f.target.id&&(h.targetId=f.target.id),f.target.dataset&&(h.dataset=f.target.dataset)),f.type&&(h.eventType=f.type),f.detail&&(typeof f.detail.x=="number"&&(h.x=f.detail.x),typeof f.detail.y=="number"&&(h.y=f.detail.y)),f.touches&&f.touches.length>0)){const m=f.touches[0];m&&(h.touchX=m.pageX,h.touchY=m.pageY)}return M({category:"user.interaction",message:`${u} on ${d}`,level:"info",data:h}),p.apply(this,[f,..._])}}}}return n(a)};Object.defineProperty(c,"__sentryPageWrapper",{value:!0,configurable:!0}),this._wrappedPage=c,e.Page=c}_subscribeApp(){this._options.enableLifecycle&&(this._unsubscribeApp=pi({onLaunch:()=>{this._launchTime=Date.now(),this._appBreadcrumb("onLaunch")},onShow:()=>this._appBreadcrumb("onShow"),onHide:()=>this._appBreadcrumb("onHide")}))}_appBreadcrumb(e){M({category:"app.lifecycle",message:`App.${e}`,level:"info",data:{action:e}})}cleanup(){const e=globalThis;this._originalPage&&(e.Page===this._wrappedPage&&(e.Page=this._originalPage),this._originalPage=null,this._wrappedPage=null),this._unsubscribeApp&&(this._unsubscribeApp(),this._unsubscribeApp=null)}};zt.id="PageBreadcrumbs";let Ke=zt;const Ku=t=>new Ke(t),Vu=["debug","info","warn","error","log"],Xu={debug:"debug",info:"info",log:"info",warn:"warning",error:"error"},qt=class qt{constructor(e={}){this.name=qt.id,this._originalMethods={},this._levels=e.levels||[...Vu]}setupOnce(){for(const e of this._levels){if(typeof console[e]!="function")continue;const n=console[e];this._originalMethods[e]=n,console[e]=function(...r){return M({category:"console",level:Xu[e],message:r.map(s=>{if(typeof s=="string")return s;try{return JSON.stringify(s)}catch(i){return String(s)}}).join(" ")}),n.apply(console,r)}}}cleanup(){for(const e of this._levels){const n=this._originalMethods[e];n&&(console[e]=n)}this._originalMethods={}}};qt.id="ConsoleBreadcrumbs";let Ve=qt;const Qu=t=>new Ve(t),Wt=class Wt{constructor(){this.name=Wt.id,this._isSessionActive=!1,this._unsubscribe=null}setupOnce(){this._unsubscribe=pi({onLaunch:()=>this._startSession(),onShow:()=>{this._isSessionActive||this._startSession()},onHide:()=>this._endSession()})}_startSession(){try{vs({ignoreDuration:!0}),Pn(),this._isSessionActive=!0}catch(e){}}_endSession(){try{pt(),Pn(),this._isSessionActive=!1}catch(e){}}cleanup(){this._unsubscribe&&(this._unsubscribe(),this._unsubscribe=null),this._isSessionActive=!1}};Wt.id="Session";let Nt=Wt;const Gt=class Gt{constructor(){this.name=Gt.id,this._statusChangeHandler=null,this._lastConnected=null}setupOnce(){const e=w();if(e){if(typeof e.getNetworkType=="function")try{e.getNetworkType({success:n=>{const r=n.networkType||"unknown";this._lastConnected=r!=="none",re("network",{type:r,isConnected:this._lastConnected})}})}catch(n){}typeof e.onNetworkStatusChange=="function"&&(this._statusChangeHandler=n=>{var i;const r=n.networkType||"unknown",s=n.isConnected!==void 0?n.isConnected:r!=="none";if(re("network",{type:r,isConnected:s}),M({category:"network.change",message:`网络状态变化: ${r}`,level:s?"info":"warning",data:{networkType:r,isConnected:s}}),s&&this._lastConnected===!1)try{(i=P())==null||i.flush()}catch(o){}this._lastConnected=s},e.onNetworkStatusChange(this._statusChangeHandler))}}cleanup(){if(this._statusChangeHandler){const e=w();if(e&&typeof e.offNetworkStatusChange=="function")try{e.offNetworkStatusChange(this._statusChangeHandler)}catch(n){}this._statusChangeHandler=null}}};Gt.id="NetworkStatus";let Ot=Gt;const Jt=class Jt{constructor(){this.name=Jt.id,this._initTs=Je(),this._initEpoch=Vn(),this._showHandler=null,this._hideHandler=null,this._coldStartReported=!1,this._minigameContext={runtime:"minigame"}}setupOnce(){const e=w();if(e){if(typeof e.getLaunchOptionsSync=="function")try{const n=e.getLaunchOptionsSync()||{};this._minigameContext.scene=n.scene,this._minigameContext.path=n.path,this._minigameContext.query=n.query,re("minigame",l({},this._minigameContext)),M({category:"minigame.launch",message:"小游戏冷启动",level:"info",data:{scene:n.scene,path:n.path}})}catch(n){}this._measureColdStart(),typeof e.onShow=="function"&&(this._showHandler=n=>{M({category:"minigame.lifecycle",message:"小游戏 onShow（进入前台）",level:"info",data:{scene:n&&n.scene}})},e.onShow(this._showHandler)),typeof e.onHide=="function"&&(this._hideHandler=()=>{M({category:"minigame.lifecycle",message:"小游戏 onHide（退到后台）",level:"info"})},e.onHide(this._hideHandler))}}_measureColdStart(){const e=globalThis.requestAnimationFrame;typeof e=="function"&&e(()=>{if(this._coldStartReported)return;this._coldStartReported=!0;const n=Je(),r=Math.max(0,Math.round(n-this._initTs));this._minigameContext.coldStartMs=r,re("minigame",l({},this._minigameContext)),M({category:"minigame.performance",message:`SDK 初始化到首帧耗时: ${r}ms`,level:"info",data:{coldStartMs:r}});const s=ft({name:"minigame.coldstart",op:"app.start",forceTransaction:!0,startTime:this._initEpoch/1e3});s.setAttributes({"minigame.scene":this._minigameContext.scene,"minigame.path":this._minigameContext.path,"minigame.cold_start_ms":r}),Ie("cold_start",r,"millisecond",s),s.end((this._initEpoch+r)/1e3)})}cleanup(){const e=w();if(e){try{this._showHandler&&typeof e.offShow=="function"&&e.offShow(this._showHandler),this._hideHandler&&typeof e.offHide=="function"&&e.offHide(this._hideHandler)}catch(n){}this._showHandler=null,this._hideHandler=null}}};Jt.id="Minigame";let Re=Jt;const Zu=()=>new Re,el=["minor","major","severe"];function tl(t){if(!t)return[];const e=[];for(const n of el){const r=t[n];typeof r=="number"&&Number.isFinite(r)&&r>0&&e.push({name:n,threshold:r})}if(e.length===0)return[];for(let n=1;n<e.length;n++){const r=e[n-1],s=e[n];if(r&&s&&s.threshold<=r.threshold)return console.warn("[sentry-miniapp] jankLevels 阈值须按 minor < major < severe 严格递增，当前配置不满足，已忽略分级并回退到单档 longFrameThresholdMs。"),[]}return e}function nl(t,e){for(let n=t.length-1;n>=0;n--){const r=t[n];if(r&&e>r.threshold)return r.name}return null}function rl(t,e){var r;if(t.length===0)return e;const n=[...t].sort((s,i)=>s-i);return(r=n[Math.floor(.95*(n.length-1))])!=null?r:e}const G=class G{constructor(e={}){var r,s,i,o;this.name=G.id,this._running=!1,this._windowElapsed=0,this._lastFrameTs=0,this._frameCount=0,this._jankCount=0,this._jankBreadcrumbs=0,this._jankByTier={minor:0,major:0,severe:0},this._maxFrameDelta=0,this._sessionEpochStart=0,this._sessionElapsed=0,this._sessionFrames=0,this._sessionJank=0,this._sessionJankByTier={minor:0,major:0,severe:0},this._sessionWorstFrame=0,this._fpsSamples=[],this._showHandler=null,this._hideHandler=null,this._options={fpsWarningThreshold:(r=e.fpsWarningThreshold)!=null?r:30,longFrameThresholdMs:(s=e.longFrameThresholdMs)!=null?s:50,reportInterval:(i=e.reportInterval)!=null?i:1e4,maxJankBreadcrumbsPerWindow:(o=e.maxJankBreadcrumbsPerWindow)!=null?o:3},this._tiers=tl(e.jankLevels);const n=this._tiers[0];this._tiered=n!==void 0,this._jankEntryThreshold=n?n.threshold:this._options.longFrameThresholdMs}setupOnce(){const e=globalThis.requestAnimationFrame;if(typeof e!="function"){console.warn("[sentry-miniapp] requestAnimationFrame 不可用，小游戏帧率监控已跳过（小程序逻辑层不支持）");return}this._running=!0;const n=Je();this._lastFrameTs=n,this._sessionEpochStart=Vn();const r=()=>{if(!this._running)return;const i=Je();this._onFrame(i-this._lastFrameTs),this._lastFrameTs=i,e(r)};e(r);const s=w();s&&typeof s.onHide=="function"&&(this._hideHandler=()=>{this._flushSummary()&&this._flushPendingEvents()},s.onHide(this._hideHandler)),s&&typeof s.onShow=="function"&&(this._showHandler=()=>this._restartOnResume(),s.onShow(this._showHandler))}_onFrame(e){const n=this._sanitizeFrameDelta(e);if(n===null){(!Number.isFinite(e)||e>G._MAX_REASONABLE_FRAME_DELTA_MS)&&this._rollupWindow();return}if(this._frameCount+=1,this._windowElapsed+=n,n>this._maxFrameDelta&&(this._maxFrameDelta=n),n>this._jankEntryThreshold){this._jankCount+=1;const r=this._tiered?nl(this._tiers,n):null;r&&(this._jankByTier[r]+=1),this._jankBreadcrumbs<this._options.maxJankBreadcrumbsPerWindow&&(this._jankBreadcrumbs+=1,M({category:"minigame.jank",message:`检测到卡顿帧: ${Math.round(n)}ms`,level:"warning",data:r?{frameDurationMs:Math.round(n),jankLevel:r}:{frameDurationMs:Math.round(n)}}))}this._windowElapsed>=this._options.reportInterval&&this._report()}_sanitizeFrameDelta(e){return!Number.isFinite(e)||e<=0||e>G._MAX_REASONABLE_FRAME_DELTA_MS?null:e}_report(){const e=this._rollupWindow();e&&(re("minigame.framerate",{fps:e.fps,minFps:e.minFps,worstFrameMs:e.worstFrameMs,jankCount:e.jankCount,frames:e.frames}),M({category:"minigame.framerate",message:`FPS: ${e.fps}（最低 ${e.minFps}，卡顿 ${e.jankCount} 次）`,level:e.fps<this._options.fpsWarningThreshold?"warning":"info",data:e}))}_rollupWindow(){const e=this._windowElapsed;if(this._frameCount===0||e<=0)return this._resetWindow(),null;const n=Math.round(this._frameCount/e*1e3),r=this._maxFrameDelta>0?Math.round(1e3/this._maxFrameDelta):n,s={fps:n,minFps:r,worstFrameMs:Math.round(this._maxFrameDelta),jankCount:this._jankCount,frames:this._frameCount};return this._fpsSamples.push(n),this._fpsSamples.length>G._MAX_FPS_SAMPLES&&this._fpsSamples.shift(),this._sessionElapsed+=e,this._sessionFrames+=this._frameCount,this._sessionJank+=this._jankCount,this._sessionJankByTier.minor+=this._jankByTier.minor,this._sessionJankByTier.major+=this._jankByTier.major,this._sessionJankByTier.severe+=this._jankByTier.severe,this._maxFrameDelta>this._sessionWorstFrame&&(this._sessionWorstFrame=this._maxFrameDelta),this._resetWindow(),s}_resetWindow(){this._windowElapsed=0,this._frameCount=0,this._jankCount=0,this._jankBreadcrumbs=0,this._maxFrameDelta=0,this._jankByTier={minor:0,major:0,severe:0}}_resetSession(){this._sessionEpochStart=Vn(),this._sessionElapsed=0,this._sessionFrames=0,this._sessionJank=0,this._sessionJankByTier={minor:0,major:0,severe:0},this._sessionWorstFrame=0,this._fpsSamples=[]}_restartOnResume(){this._resetSession();const e=Je();this._lastFrameTs=e,this._resetWindow()}_flushSummary(){if(this._rollupWindow(),this._sessionFrames===0)return!1;const e=this._sessionElapsed,n=e>0?Math.round(this._sessionFrames/e*1e3):0,r=this._sessionWorstFrame>0?Math.round(1e3/this._sessionWorstFrame):n,s=rl(this._fpsSamples,n),i=Math.round(this._sessionWorstFrame),o=this._sessionJank,c=this._sessionFrames,a=ft({name:"minigame.framerate.summary",op:"ui.framerate",forceTransaction:!0,startTime:this._sessionEpochStart/1e3});if(a.setAttributes({"fps.avg":n,"fps.p95":s,"fps.min":r,"frames.total":c,"jank.count":o,"frame.worst_ms":i}),Ie("fps_avg",n,"none",a),Ie("fps_p95",s,"none",a),Ie("fps_min",r,"none",a),Ie("jank_count",o,"none",a),this._tiered){const u={};for(const p of this._tiers){const f=this._sessionJankByTier[p.name];u[`jank.${p.name}`]=f,Ie(`jank_${p.name}_count`,f,"none",a)}a.setAttributes(u)}return a.end((this._sessionEpochStart+Math.max(0,e))/1e3),this._resetSession(),!0}_flushPendingEvents(){Promise.resolve(Es(G._HIDE_FLUSH_TIMEOUT_MS)).catch(()=>{})}cleanup(){this._running=!1,this._flushSummary()&&this._flushPendingEvents();const e=w();if(e)try{this._hideHandler&&typeof e.offHide=="function"&&e.offHide(this._hideHandler),this._showHandler&&typeof e.offShow=="function"&&e.offShow(this._showHandler)}catch(n){}this._hideHandler=null,this._showHandler=null}};G.id="MinigameFrameRate",G._MAX_FPS_SAMPLES=2e3,G._MAX_REASONABLE_FRAME_DELTA_MS=5e3,G._HIDE_FLUSH_TIMEOUT_MS=2e3;let Ce=G;const sl=Object.freeze(Object.defineProperty({__proto__:null,ConsoleBreadcrumbs:Ve,Dedupe:It,GlobalHandlers:bt,HttpContext:vt,LinkedErrors:Tt,MinigameFrameRateIntegration:Ce,MinigameIntegration:Re,NetworkBreadcrumbs:At,NetworkStatusIntegration:Ot,PageBreadcrumbs:Ke,PerformanceIntegration:wt,RewriteFrames:kt,Router:nr,SessionIntegration:Nt,System:tr,TryCatch:Et,consoleBreadcrumbsIntegration:Qu,minigameFrameRateIntegration:t=>()=>new Ce(t),minigameIntegration:Zu,pageBreadcrumbsIntegration:Ku,performanceIntegration:rr},Symbol.toStringTag,{value:"Module"}));function ir(){return[new vt,new It,new bt,new Et,new Tt,rr({enableNavigation:!0,enableRender:!0,enableResource:!0,enableUserTiming:!0,sampleRate:1,reportInterval:3e4})]}const il=ir();function ol(t){return t===!1?[]:Array.isArray(t)?[...t]:ir()}function al(t={}){if(!Zs()){console.warn("[sentry-miniapp] Not running in a supported miniapp environment");return}const e=[...t.integrations||ol(t.defaultIntegrations)],n=I(l({},t),{defaultIntegrations:[],integrations:e,stackParser:Ks,transport:t.transport});n.enableSourceMap!==!1&&n.integrations.push(new kt);const r={traceNetworkBody:n.traceNetworkBody};n.enableTracePropagation!==void 0&&(r.enableTracePropagation=n.enableTracePropagation),n.tracePropagationTargets!==void 0&&(r.tracePropagationTargets=n.tracePropagationTargets),n.propagateTraceparent!==void 0&&(r.propagateTraceparent=n.propagateTraceparent),n.integrations.push(new At(r)),n.enableAutoSessionTracking!==!1&&n.integrations.push(new Nt);const s=n.enableNavigationBreadcrumbs!==!1,i=n.enableUserInteractionBreadcrumbs!==!1;(s||i)&&n.integrations.push(new Ke({enableLifecycle:s,enableUserInteraction:i})),n.enableNetworkStatusMonitoring!==!1&&n.integrations.push(new Ot),n.enableConsoleBreadcrumbs&&n.integrations.push(new Ve);const o=bu(),c=a=>n.integrations.some(u=>u&&u.name===a);if(!c("EventFilters")&&!c("InboundFilters")){const a={};n.allowUrls&&(a.allowUrls=n.allowUrls),n.denyUrls&&(a.denyUrls=n.denyUrls),n.ignoreErrors&&(a.ignoreErrors=n.ignoreErrors),n.integrations.push(Bc(a))}return!c(Re.id)&&(n.enableMinigameLifecycle===!0||o&&n.enableMinigameLifecycle!==!1)&&n.integrations.push(new Re),!c(Ce.id)&&(n.enableMinigameFrameRate===!0||o&&n.enableMinigameFrameRate!==!1)&&n.integrations.push(new Ce(n.minigameFrameRateOptions)),re("miniapp",{platform:ei(),environment:"miniapp"}),Cc(ii,n),T().getClient()}function cl(t={}){console.warn("[sentry-miniapp] showReportDialog is deprecated and does nothing. Please build your own UI and use `Sentry.captureFeedback()` instead.")}function ul(t){return function(...e){return ie(()=>{try{return t.apply(this,e)}catch(n){throw T().captureException(n),n}})}}function ll(t){var e,n,r;du(t),t&&((r=(n=(e=P())==null?void 0:e.getTransport())==null?void 0:n.flush)==null||r.call(n))}function fl(){return Vs()}function dl(t){const e=T().getClient();return e?e.captureFeedback(t):(console.warn("[sentry-miniapp] No client available for captureFeedback"),"")}U(),g.Integrations=sl,g.MiniappClient=ii,g.SDK_NAME=Js,g.SDK_VERSION=gt,g.Transports=Au,g.addBreadcrumb=M,g.addEventProcessor=ja,g.addIntegration=Ga,g.captureEvent=Na,g.captureException=we,g.captureFeedback=dl,g.captureMessage=Pa,g.captureSession=Pn,g.close=$a,g.closeSession=Er,g.defaultIntegrations=il,g.endSession=pt,g.flush=Es,g.getClient=P,g.getConsent=fl,g.getCurrentScope=T,g.getDefaultIntegrations=ir,g.getIsolationScope=F,g.getPerformanceManager=ti,g.init=al,g.isEnabled=Ts,g.lastEventId=La,g.logger=su,g.makeSession=br,g.miniappStackParser=Ks,g.performanceIntegration=rr,g.setConsent=ll,g.setContext=re,g.setExtra=Da,g.setExtras=Oa,g.setTag=xa,g.setTags=Ma,g.setUser=Fa,g.showReportDialog=cl,g.startInactiveSpan=ft,g.startSession=vs,g.startSpan=He,g.updateSession=se,g.withScope=ie,g.wrap=ul,Object.defineProperty(g,Symbol.toStringTag,{value:"Module"})}));
 //# sourceMappingURL=sentry-miniapp.umd.js.map
